### 2020-01-16 #2018 @whedon
[REVIEW]: Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs
**Submitting author:** @ali-ramadhan (<a href="http://orcid.org/0000-0003-1102-1520">Ali Ramadhan</a>)
**Repository:** <a href="https://github.com/CliMA/Oceananigans.jl" target ="_blank">https://github.com/CliMA/Oceananigans.jl</a>
**Version:** v0.36.0
