### 2020-01-16 #2018 @ali-ramadhan
Thanks so much @funsim and @ysimillides for agreeing to review our submission!

Please let me know if you guys need access to GPUs and we can figure something out.
