### 2020-01-17 d1f64f4
Adds functionality for SeawaterBuoyancy with constant temperature *or* salinity; eliminates buoyancy_frequency_squared in favor of dz_b function, modifications to depth function for nonlinear equations of state
