+++ src/Buoyancy/nonlinear_equation_of_state.jl
@@ -1,21 +1,27 @@
 """ Return the geopotential depth at `i, j, k` at cell centers. """
-@inline Dᵃᵃᶜ(i, j, k, grid) = @inbounds -grid.zC[k]
+@inline Dᵃᵃᶜ(i, j, k, grid::AbstractGrid{FT}) where FT = @inbounds k < 1       ? - grid.zC[1] + (1 - k) * grid.Δz :
+                                                                   k > grid.Nz ? - grid.zC[end] - (k - grid.Nz) * grid.Δz : 
+                                                                                 - grid.zC[k]
 
 """ Return the geopotential depth at `i, j, k` at cell z-interfaces. """
-@inline Dᵃᵃᶠ(i, j, k, grid) = @inbounds -grid.zF[k]
+@inline Dᵃᵃᶠ(i, j, k, grid::AbstractGrid{FT}) where FT = @inbounds k < 1           ? - grid.zF[1] + (1 - k) * grid.Δz :
+                                                                   k > grid.Nz + 1 ? - grid.zF[end] - (k - grid.Nz + 1) * grid.Δz : 
+                                                                                     - grid.zF[k]
+# Dispatch shenanigans
+@inline θ_and_sᴬ(i, j, k, θ::AbstractArray, sᴬ::AbstractArray) = @inbounds θ[i, j, k], sᴬ[i, j, k]
+@inline θ_and_sᴬ(i, j, k, θ::Number,        sᴬ::AbstractArray) = @inbounds θ, sᴬ[i, j, k]
+@inline θ_and_sᴬ(i, j, k, θ::AbstractArray, sᴬ::Number)        = @inbounds θ[i, j, k], sᴬ
+@inline θ_and_sᴬ(i, j, k, θ::Number,        sᴬ::Number)        = @inbounds θ, sᴬ
 
 # Basic functionality
-@inline ρ′(i, j, k, grid, eos, C) = @inbounds ρ′(C.T[i, j, k], C.S[i, j, k], Dᵃᵃᶜ(i, j, k, grid), eos)
+@inline ρ′(i, j, k, grid, eos, θ, sᴬ) = @inbounds ρ′(θ_and_sᴬ(i, j, k, θ, sᴬ)..., Dᵃᵃᶜ(i, j, k, grid), eos)
 
-@inline thermal_expansionᶜᶜᶜ(i, j, k, grid, eos, C) = @inbounds thermal_expansion(C.T[i, j, k], C.S[i, j, k], Dᵃᵃᶜ(i, j, k, grid), eos)
-@inline thermal_expansionᶠᶜᶜ(i, j, k, grid, eos, C) = @inbounds thermal_expansion(ℑxᶠᵃᵃ(i, j, k, grid, C.T), ℑxᶠᵃᵃ(i, j, k, grid, C.S), Dᵃᵃᶜ(i, j, k, grid), eos)
-@inline thermal_expansionᶜᶠᶜ(i, j, k, grid, eos, C) = @inbounds thermal_expansion(ℑyᵃᶠᵃ(i, j, k, grid, C.T), ℑyᵃᶠᵃ(i, j, k, grid, C.S), Dᵃᵃᶜ(i, j, k, grid), eos)
[…]
