+++ test/test_buoyancy.jl
@@ -1,7 +1,7 @@
 using Oceananigans.Models: TracerFields
 
 using Oceananigans.Buoyancy:
-    required_tracers, ρ′, ∂x_b, ∂y_b, buoyancy_frequency_squared,
+    required_tracers, ρ′, ∂x_b, ∂y_b,
     thermal_expansionᶜᶜᶜ, thermal_expansionᶠᶜᶜ, thermal_expansionᶜᶠᶜ, thermal_expansionᶜᶜᶠ,
     haline_contractionᶜᶜᶜ, haline_contractionᶠᶜᶜ, haline_contractionᶜᶠᶜ, haline_contractionᶜᶜᶠ
 
@@ -24,7 +24,7 @@ end
 function density_perturbation_works(arch, FT, eos)
     grid = RegularCartesianGrid(FT; size=(3, 3, 3), length=(1, 1, 1))
     C = datatuple(TracerFields(arch, grid, (:T, :S)))
-    density_anomaly = ρ′(2, 2, 2, grid, eos, C)
+    density_anomaly = ρ′(2, 2, 2, grid, eos, C.T, C.S)
     return true
 end
 
@@ -42,30 +42,30 @@ function ∂y_b_works(arch, FT, buoyancy)
     return true
 end
 
-function buoyancy_frequency_squared_works(arch, FT, buoyancy)
+function ∂z_b_works(arch, FT, buoyancy)
     grid = RegularCartesianGrid(FT; size=(3, 3, 3), length=(1, 1, 1))
     C = datatuple(TracerFields(arch, grid, required_tracers(buoyancy)))
-    N² = buoyancy_frequency_squared(2, 2, 2, grid, buoyancy, C)
+    dbdz = ∂z_b(2, 2, 2, grid, buoyancy, C)
     return true
 end
 
 function thermal_expansion_works(arch, FT, eos)
     grid = RegularCartesianGrid(FT; size=(3, 3, 3), length=(1, 1, 1))
     C = datatuple(TracerFields(arch, grid, (:T, :S)))
-    α = thermal_expansionᶜᶜᶜ(2, 2, 2, grid, eos, C)
-    α = thermal_expansionᶠᶜᶜ(2, 2, 2, grid, eos, C)
-    α = thermal_expansionᶜᶠᶜ(2, 2, 2, grid, eos, C)
-    α = thermal_expansionᶜᶜᶠ(2, 2, 2, grid, eos, C)
+    α = thermal_expansionᶜᶜᶜ(2, 2, 2, grid, eos, C.T, C.S)
+    α = thermal_expansionᶠᶜᶜ(2, 2, 2, grid, eos, C.T, C.S)
+    α = thermal_expansionᶜᶠᶜ(2, 2, 2, grid, eos, C.T, C.S)
+    α = thermal_expansionᶜᶜᶠ(2, 2, 2, grid, eos, C.T, C.S)
     return true
 end
 
 function haline_contraction_works(arch, FT, eos)
     grid = RegularCartesianGrid(FT; size=(3, 3, 3), length=(1, 1, 1))
     C = datatuple(TracerFields(arch, grid, (:T, :S)))
-    β = haline_contractionᶜᶜᶜ(2, 2, 2, grid, eos, C)
-    β = haline_contractionᶠᶜᶜ(2, 2, 2, grid, eos, C)
-    β = haline_contractionᶜᶠᶜ(2, 2, 2, grid, eos, C)
-    β = haline_contractionᶜᶜᶠ(2, 2, 2, grid, eos, C)
+    β = haline_contractionᶜᶜᶜ(2, 2, 2, grid, eos, C.T, C.S)
[…]
