### 2020-01-18 4e878cd
Need halos of size 2 for compressible viscous operators to work.
