+++ test/test_time_stepping.jl
@@ -6,19 +6,20 @@ using Oceananigans, JULES
     @testset "Modified potential temperature" begin
         @info "  Testing time stepping with Θᵐ prognostic temperature..."
 
-        model = CompressibleModel(grid=RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1)))
+        grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1), halo=(2, 2, 2))
+        model = CompressibleModel(grid=grid)
         time_step!(model; Δt=1)
 
-        @test isa(model, CompressibleModel)
+        @test model isa CompressibleModel
     end
 
     @testset "Entropy" begin
         @info "  Testing model construction with S prognostic temperature..."
 
-        model = CompressibleModel(grid=RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1)),
-                                  prognostic_temperature=Entropy(), tracers=(:S,))
+        grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1), halo=(2, 2, 2))
+        model = CompressibleModel(grid=grid, prognostic_temperature=Entropy(), tracers=(:S,))
         time_step!(model; Δt=1)
 
-        @test isa(model, CompressibleModel)
+        @test model isa CompressibleModel
     end
 end
