+++ examples/solve_for_hydrostatic_profiles.jl
@@ -7,7 +7,8 @@ Nx = Ny = 1
 Nz = 32
 L = 10e3
 
-grid = RegularCartesianGrid(size=(Nx, Ny, Nz), x=(0, L), y=(0, L), z=(0, L))
+grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
+                            x=(0, L), y=(0, L), z=(0, L))
 Δz = grid.Δz
 
 pₛ = 100000
