### 2020-01-20 1861cdd
Update author list order
