### 2020-01-21 d79d223
Tweaks determination of depth for nonlinear equations of state. This may need to be revisited.
