### 2020-01-22 7f4b980
Make Value/Gradient BCs and Rayleigh-Benard regression grid agnostic.
