### 2020-01-22 91545b6
Update some turbulence closure and VAMD functions.
