+++ src/TurbulenceClosures/turbulence_closure_implementations/rozema_anisotropic_minimum_dissipation.jl
@@ -1,3 +1,5 @@
+import Oceananigans.Operators: Δx, Δy
+
 struct RozemaAnisotropicMinimumDissipation{FT, K} <: AbstractAnisotropicMinimumDissipation{FT}
      C :: FT
     Cb :: FT
