+++ src/TurbulenceClosures/turbulence_closure_implementations/verstappen_anisotropic_minimum_dissipation.jl
@@ -159,9 +159,9 @@ end
 #####
 
 # Recall that filter widths are 2x the grid spacing in VAMD
-@inline Δᶠxᶜᶜᶜ(i, j, k, grid::RegularCartesianGrid) = 2 * grid.Δx
-@inline Δᶠyᶜᶜᶜ(i, j, k, grid::RegularCartesianGrid) = 2 * grid.Δy
-@inline Δᶠzᶜᶜᶜ(i, j, k, grid::RegularCartesianGrid) = 2 * grid.Δz
+@inline Δᶠxᶜᶜᶜ(i, j, k, grid) = 2 * Δx(i, j, k, grid)
+@inline Δᶠyᶜᶜᶜ(i, j, k, grid) = 2 * Δy(i, j, k, grid)
+@inline Δᶠzᶜᶜᶜ(i, j, k, grid) = 2 * ΔzC(i, j, k, grid)
 
 for loc in (:ccf, :fcc, :cfc, :ffc, :cff, :fcf), ξ in (:x, :y, :z)
     Δ_loc = Symbol(:Δᶠ, ξ, :_, loc)
