+++ src/Buoyancy/Buoyancy.jl
@@ -5,6 +5,12 @@ export
     LinearEquationOfState, RoquetIdealizedNonlinearEquationOfState,
     ∂x_b, ∂y_b, ∂z_b, buoyancy_perturbation, buoyancy_frequency_squared
 
+using 
+    Printf,
+    Oceananigans.Operators
+
+using Oceananigans: AbstractGrid
+
 # Physical constants
 # https://en.wikipedia.org/wiki/Gravitational_acceleration#Gravity_model_for_Earth (30 Oct 2019)
 const g_Earth = 9.80665
@@ -30,13 +36,50 @@ Abstract supertype for nonlinar equations of state.
 """
 abstract type AbstractNonlinearEquationOfState <: AbstractEquationOfState end
 
-include("buoyancy_utils.jl")
-include("no_buoyancy.jl")
-include("buoyancy_tracer.jl")
-include("seawater_buoyancy.jl")
+function validate_buoyancy(buoyancy, tracers)
+    req_tracers = required_tracers(buoyancy)
+
+    all(tracer ∈ tracers for tracer in req_tracers) ||
+        error("$(req_tracers) must be among the list of tracers to use $(typeof(buoyancy).name.wrapper)")
+
+    return nothing
+end
+
+validate_buoyancy(::Nothing, tracers) = nothing
+
+required_tracers(::Nothing) = ()
+
+@inline buoyancy_perturbation(i, j, k, grid::AbstractGrid{FT}, ::Nothing, C) where FT = zero(FT)
+
+@inline ∂x_b(i, j, k, grid::AbstractGrid{FT}, ::Nothing, C) where FT = zero(FT)
+@inline ∂y_b(i, j, k, grid::AbstractGrid{FT}, ::Nothing, C) where FT = zero(FT)
+@inline ∂z_b(i, j, k, grid::AbstractGrid{FT}, ::Nothing, C) where FT = zero(FT)
+
+"""
+    BuoyancyTracer <: AbstractBuoyancy{Nothing}
+
+Type indicating that the tracer `b` represents buoyancy.
+"""
+struct BuoyancyTracer <: AbstractBuoyancy{Nothing} end
+
+required_tracers(::BuoyancyTracer) = (:b,)
+
+@inline buoyancy_perturbation(i, j, k, grid, ::BuoyancyTracer, C) = @inbounds C.b[i, j, k]
+
+@inline ∂x_b(i, j, k, grid, ::BuoyancyTracer, C) = ∂xᶠᵃᵃ(i, j, k, grid, C.b)
+@inline ∂y_b(i, j, k, grid, ::BuoyancyTracer, C) = ∂yᵃᶠᵃ(i, j, k, grid, C.b)
+@inline ∂z_b(i, j, k, grid, ::BuoyancyTracer, C) = ∂zᵃᵃᶠ(i, j, k, grid, C.b)
+
 include("linear_equation_of_state.jl")
 include("nonlinear_equation_of_state.jl")
+include("seawater_buoyancy.jl")
 include("roquet_idealized_nonlinear_eos.jl")
-include("show_buoyancy.jl")
+
+Base.show(io::IO, b::SeawaterBuoyancy{FT}) where FT =
+    println(io, "SeawaterBuoyancy{$FT}: g = $(b.gravitational_acceleration)", '\n',
+                "└── equation of state: $(b.equation_of_state)")
+
+Base.show(io::IO, eos::LinearEquationOfState{FT}) where FT =
+    println(io, "LinearEquationOfState{$FT}: ", @sprintf("α = %.2e, β = %.2e", eos.α, eos.β))
 
 end
