+++ src/Buoyancy/buoyancy_utils.jl
@@ -1,10 +1,0 @@
-function validate_buoyancy(buoyancy, tracers)
-    req_tracers = required_tracers(buoyancy)
-
-    all(tracer ∈ tracers for tracer in req_tracers) ||
-        error("$(req_tracers) must be among the list of tracers to use $(typeof(buoyancy).name.wrapper)")
-
-    return nothing
-end
-
-validate_buoyancy(::Nothing, tracers) = nothing
