+++ src/Buoyancy/Buoyancy.jl
@@ -70,9 +70,9 @@ required_tracers(::BuoyancyTracer) = (:b,)
 @inline ∂y_b(i, j, k, grid, ::BuoyancyTracer, C) = ∂yᵃᶠᵃ(i, j, k, grid, C.b)
 @inline ∂z_b(i, j, k, grid, ::BuoyancyTracer, C) = ∂zᵃᵃᶠ(i, j, k, grid, C.b)
 
+include("seawater_buoyancy.jl")
 include("linear_equation_of_state.jl")
 include("nonlinear_equation_of_state.jl")
-include("seawater_buoyancy.jl")
 include("roquet_idealized_nonlinear_eos.jl")
 
 Base.show(io::IO, b::SeawaterBuoyancy{FT}) where FT =
