### 2020-01-25 9c7eb64
Refactor boundary conditions module by coalescing small files related to the model boundary conditions into one medium size file
