+++ src/BoundaryConditions/BoundaryConditions.jl
@@ -20,11 +20,7 @@ include("boundary_condition_types.jl")
 include("boundary_condition.jl")
 include("coordinate_boundary_conditions.jl")
 include("field_boundary_conditions.jl")
-include("solution_boundary_conditions.jl")
-include("tracer_boundary_conditions.jl")
-include("tendency_boundary_conditions.jl")
-include("pressure_boundary_conditions.jl")
-include("model_boundary_conditions.jl")
+include("solution_and_model_boundary_conditions.jl")
 include("boundary_function.jl")
 include("show_boundary_conditions.jl")
 
