### 2020-01-25 10bb86f
Refactors solution and model boundary conditions file
