### 2020-01-26 2c920ca
Adds diffusivities to ModelBoundaryConditions constructor, plus adds new constructor that permits diffusivity boundary conditions to be changed
