### 2020-01-26 9a5f74c
Fixes bug in ModelBoundaryConditions constructor and initiates use of diffusivities boundary conditions in time stepping
