### 2020-01-26 78d0a22
Update Smagorinsky-Lilly to work with vertically stretched grids.
