+++ src/TurbulenceClosures/turbulence_closure_utils.jl
@@ -20,15 +20,12 @@ end
 convert_diffusivity(T, κ::Number) = convert(T, κ)
 convert_diffusivity(T, κ::NamedTuple) = convert(NamedTuple{propertynames(κ), NTuple{length(κ), T}}, κ)
 
-@inline geo_mean_Δᶠ(i, j, k, grid::RegularCartesianGrid{T}) where T = (grid.Δx * grid.Δy * grid.Δz)^T(1/3)
+@inline geo_mean_Δᶠ(i, j, k, grid::AbstractGrid{T}) where T =
+    (Δx(i, j, k, grid) * Δy(i, j, k, grid) * ΔzC(i, j, k, grid))^T(1/3)
 
 function calculate_nonlinear_viscosity!(νₑ, grid, closure, buoyancy, U, C)
-    @loop for k in (1:grid.Nz; (blockIdx().z - 1) * blockDim().z + threadIdx().z)
-        @loop for j in (1:grid.Ny; (blockIdx().y - 1) * blockDim().y + threadIdx().y)
-            @loop for i in (1:grid.Nx; (blockIdx().x - 1) * blockDim().x + threadIdx().x)
-                @inbounds νₑ[i, j, k] = νᶜᶜᶜ(i, j, k, grid, closure, buoyancy, U, C)
-            end
-        end
+    @loop_xyz i j k grid begin
+        @inbounds νₑ[i, j, k] = νᶜᶜᶜ(i, j, k, grid, closure, buoyancy, U, C)
     end
     return nothing
 end
