+++ src/TurbulenceClosures/turbulence_closure_implementations/smagorinsky_lilly.jl
@@ -190,7 +190,7 @@ end
 #####
 
 "Return the filter width for Constant Smagorinsky on a Regular Cartesian grid."
-@inline Δᶠ(i, j, k, grid::RegularCartesianGrid, ::AbstractSmagorinsky) = geo_mean_Δᶠ(i, j, k, grid)
+@inline Δᶠ(i, j, k, grid, ::AbstractSmagorinsky) = geo_mean_Δᶠ(i, j, k, grid)
 
 # Temporarily set filter widths to cell-size (rather than distance between cell centers, etc.)
 const Δᶠ_ccc = Δᶠ
