+++ src/Solvers/horizontally_periodic_pressure_solver.jl
@@ -2,7 +2,7 @@
 ##### CPU horizontally periodic pressure solver for regular grids
 #####
 
-function HorizontallyPeriodicPressureSolver(::CPU, grid, pressure_bcs, planner_flag=FFTW.PATIENT)
+function HorizontallyPeriodicPressureSolver(::CPU, grid::RegularCartesianGrid, pressure_bcs, planner_flag=FFTW.PATIENT)
     kx², ky², kz² = generate_discrete_eigenvalues(grid, pressure_bcs)
     wavenumbers = (kx² = kx², ky² = ky², kz² = kz²)
 
@@ -35,7 +35,7 @@ mutates to produce the solution, so it will also be stored in solver.storage.
 
 We should describe the algorithm in detail in the documentation.
 """
-function solve_poisson_equation!(solver::PressureSolver{HorizontallyPeriodic, CPU}, grid)
+function solve_poisson_equation!(solver::PressureSolver{HorizontallyPeriodic, CPU}, grid::RegularCartesianGrid)
     Nx, Ny, Nz, _ = unpack_grid(grid)
     kx², ky², kz² = solver.wavenumbers.kx², solver.wavenumbers.ky², solver.wavenumbers.kz²
 
@@ -68,7 +68,7 @@ end
 ##### GPU horizontally periodic pressure solver for regular grids
 #####
 
-function HorizontallyPeriodicPressureSolver(::GPU, grid, pressure_bcs, no_args...)
+function HorizontallyPeriodicPressureSolver(::GPU, grid::RegularCartesianGrid, pressure_bcs, no_args...)
     Nx, Ny, Nz, _ = unpack_grid(grid)
 
     kx², ky², kz² = generate_discrete_eigenvalues(grid, pressure_bcs) .|> CuArray
@@ -124,7 +124,7 @@ The output will be permuted in this way and so the permutation must be undone.
 
 We should describe the algorithm in detail in the documentation.
 """
-function solve_poisson_equation!(solver::PressureSolver{HorizontallyPeriodic, GPU}, grid)
+function solve_poisson_equation!(solver::PressureSolver{HorizontallyPeriodic, GPU}, grid::RegularCartesianGrid)
     kx², ky², kz² = solver.wavenumbers.kx², solver.wavenumbers.ky², solver.wavenumbers.kz²
     ω_4Nz⁺, ω_4Nz⁻ = solver.constants.ω_4Nz⁺, solver.constants.ω_4Nz⁻
 
@@ -160,39 +160,41 @@ end
 ##### Horizontally periodic pressure solver for vertically stretched grids
 #####
 
-# function HorizontallyPeriodicPressureSolver(arch, grid::VerticallyStretchedCartesianGrid, pressure_bcs, planner_flag=nothing)
-#     kx², ky², _ = generate_discrete_eigenvalues(grid, pressure_bcs)
-#     wavenumbers = (kx² = kx², ky² = ky²)
-#
-#     if isnothing(planner_flag) && arch isa CPU
-#         planner_flag = FFTW.PATIENT
-#     end
-#
-#     storage = zeros(Complex{Float64}, grid.Nx, grid.Ny, grid.Nz)
-#
-#     @debug "Planning transforms for PressureSolver{HorizontallyPeriodic, $(typeof(arch)), VerticallyStretchedCartesianGrid}..."
[…]
