### 2020-01-26 b34d4b8
Add vertically stretched horizontally periodic Poisson solver.
