### 2020-01-26 f413445
Dry rising thermal bubble
