### 2020-01-27 80a001d
Switch thermal bubble to dry adiabatic atmosphere
