### 2020-01-27 994d563
Clean up operators a bit.
