+++ examples/thermal_bubble.jl
@@ -95,13 +95,13 @@ for k in 1:Nz, i in 1:Nx
     Θ[i, 1, k] = ρ[i, 1, k] * θ
 end
 
-ρ_plot = contour(model.grid.xC, model.grid.zC, rotr90(ρ.data[1:Nx, 1, 1:Nz] .- ρʰᵈ),
-                 fill=true, levels=10, clims=(-0.008, 0.008), color=:balance, dpi=200)
+ρ_plot = contour(model.grid.xC ./ km, model.grid.zC ./ km, rotr90(ρ.data[1:Nx, 1, 1:Nz] .- ρʰᵈ),
+                 fill=true, levels=10, xlims=(-5, 5), clims=(-0.008, 0.008), color=:balance, dpi=200)
 savefig(ρ_plot, "rho_prime_initial_condition.png")
 
 θ_slice = rotr90(Θ.data[1:Nx, 1, 1:Nz] ./ ρ.data[1:Nx, 1, 1:Nz])
-Θ_plot = contour(model.grid.xC, model.grid.zC, θ_slice,
-                 fill=true, levels=10, color=:thermal, dpi=200)
+Θ_plot = contour(model.grid.xC ./ km, model.grid.zC ./ km, θ_slice,
+                 fill=true, levels=10, xlims=(-5, 5), color=:thermal, dpi=200)
 savefig(Θ_plot, "theta_initial_condition.png")
 
 #####
@@ -122,10 +122,10 @@ for n in 1:200
     θ_slice = rotr90(model.tracers.Θᵐ.data[1:Nx, j, 1:Nz] ./ model.density.data[1:Nx, j, 1:Nz])
 
     u_title = @sprintf("u, t = %d s", round(Int, model.clock.time))
-    pu = contour(xC, zC, u_slice, title=u_title, fill=true, levels=10, color=:balance, clims=(-4, 4))
-    pw = contour(xC, zC, w_slice, title="w", fill=true, levels=10, color=:balance, clims=(-4, 4))
-    pρ = contour(xC, zC, ρ_slice, title="rho_prime", fill=true, levels=10, color=:balance, clims=(-0.01, 0.01))
-    pθ = contour(xC, zC, θ_slice, title="theta", fill=true, levels=10, color=:thermal, clims=(299.9, 302))
+    pu = contour(xC, zC, u_slice, title=u_title, fill=true, levels=10, xlims=(-5, 5), color=:balance, clims=(-4, 4))
+    pw = contour(xC, zC, w_slice, title="w", fill=true, levels=10, xlims=(-5, 5), color=:balance, clims=(-4, 4))
+    pρ = contour(xC, zC, ρ_slice, title="rho_prime", fill=true, levels=10, xlims=(-5, 5), color=:balance, clims=(-0.01, 0.01))
+    pθ = contour(xC, zC, θ_slice, title="theta", fill=true, levels=10, xlims=(-5, 5), color=:thermal, clims=(299.9, 302))
 
     p = plot(pu, pw, pρ, pθ, layout=(2, 2), dpi=200, show=true)
     savefig(p, @sprintf("thermal_bubble_%03d.png", n))
