### 2020-01-27 e89679a
Remove uneccesary halo region filling.
