### 2020-01-27 4af3d43
Refactors PressureBoundaryConditions and DiffusivitiesBoundaryConditions to take either proposal/solution bcs or model bcs
