+++ src/BoundaryConditions/solution_and_model_boundary_conditions.jl
@@ -1,6 +1,8 @@
 using Oceananigans.Utils: with_tracers
 using Oceananigans.TurbulenceClosures: with_tracers
 
+const ModelBoundaryConditions = NamedTuple{(:solution, :tendency, :pressure, :diffusivities)}
+
 #####
 ##### Default boundary conditions on tracers are periodic or no flux and
 ##### can be derived from boundary conditions on any field
@@ -38,44 +40,49 @@ TendenciesBoundaryConditions(solution_bcs) =
 
 #####
 ##### Boundary conditions on pressure are derived from boundary conditions
-##### on the north-south horizontal velocity, v.
+##### on the east-west horizontal velocity, u.
 #####
 
 # Pressure boundary conditions are either zero flux (Neumann) or Periodic.
 PressureBC(::BC)  = BoundaryCondition(Flux, nothing)
 PressureBC(::PBC) = PeriodicBC()
 
-function PressureBoundaryConditions(vbcs)
-    x = CoordinateBoundaryConditions(PressureBC(vbcs.x.left), PressureBC(vbcs.x.right))
-    y = CoordinateBoundaryConditions(PressureBC(vbcs.y.left), PressureBC(vbcs.y.right))
-    z = CoordinateBoundaryConditions(PressureBC(vbcs.z.left), PressureBC(vbcs.z.right))
+function PressureBoundaryConditions(solution_boundary_conditions)
+    ubcs = solution_boundary_conditions.u
+    x = CoordinateBoundaryConditions(PressureBC(ubcs.x.left), PressureBC(ubcs.x.right))
+    y = CoordinateBoundaryConditions(PressureBC(ubcs.y.left), PressureBC(ubcs.y.right))
+    z = CoordinateBoundaryConditions(PressureBC(ubcs.z.left), PressureBC(ubcs.z.right))
     return (x=x, y=y, z=z)
 end
 
+PressureBoundaryConditions(model_boundary_conditions::ModelBoundaryConditions) =
+    PressureBoundaryConditions(model_boundary_conditions.solution)
+
 #####
 ##### Boundary conditions on diffusivities are derived from boundary conditions
-##### on the north-south horizontal velocity, v.
+##### on the east-west horizontal velocity, u.
 #####
 
 # Diffusivity boundary conditions are either zero flux (Neumann) or Periodic.
 DiffusivityBC(::BC)  = BoundaryCondition(Flux, nothing)
 DiffusivityBC(::PBC) = PeriodicBC()
 
-function DiffusivityBoundaryConditions(vbcs)
-    x = CoordinateBoundaryConditions(DiffusivityBC(vbcs.x.left), DiffusivityBC(vbcs.x.right))
-    y = CoordinateBoundaryConditions(DiffusivityBC(vbcs.y.left), DiffusivityBC(vbcs.y.right))
-    z = CoordinateBoundaryConditions(DiffusivityBC(vbcs.z.left), DiffusivityBC(vbcs.z.right))
+function DiffusivityBoundaryConditions(solution_boundary_conditions)
+    ubcs = solution_boundary_conditions.u
+    x = CoordinateBoundaryConditions(DiffusivityBC(ubcs.x.left), DiffusivityBC(ubcs.x.right))
+    y = CoordinateBoundaryConditions(DiffusivityBC(ubcs.y.left), DiffusivityBC(ubcs.y.right))
+    z = CoordinateBoundaryConditions(DiffusivityBC(ubcs.z.left), DiffusivityBC(ubcs.z.right))
     return (x=x, y=y, z=z)
 end
 
 DiffusivitiesBoundaryConditions(::Nothing, args...) = nothing
-DiffusivitiesBoundaryConditions(::AbstractField, vbcs) = DiffusivityBoundaryConditions(vbcs)
[…]
