### 2020-01-27 373bc15
Adds test for setting boundary conditions on diffusivity fields
