### 2020-01-28 705a14a
Run all tests
