### 2020-01-28 ca16487
Moves DiffusivityBoundaryConditions setting test to test_boundary_conditions, turns into a test of fluxes
