### 2020-01-28 f1d341a
Adds functions for obtaining diffusivity boundary conditions and filling halos for tuples of closures
