+++ src/BoundaryConditions/fill_halo_regions.jl
@@ -11,22 +11,9 @@ Fill halo regions for each field in the tuple `fields` according
 to the single instance of `FieldBoundaryConditions` in `bcs`, possibly recursing into
 `fields` if it is a nested tuple-of-tuples.
 """
-function fill_halo_regions!(fields::Union{Tuple, NamedTuple}, bcs::FieldBoundaryConditions, arch, grid, args...)
+function fill_halo_regions!(fields::NamedTuple, bcs::FieldBoundaryConditions, arch, grid, args...)
     for field in fields
         fill_halo_regions!(field, bcs, arch, grid, args...)
-    end
-    return nothing
-end
-
-"""
-    fill_halo_regions!(fields, bcs, arch, grid)
-
-Fill halo regions for all fields in the `NamedTuple` `fields` according
-to the corresponding `NamedTuple` of `bcs`.
-"""
-function fill_halo_regions!(fields::NamedTuple{S}, bcs::NamedTuple{S}, arch, grid, args...) where S
-    for (field, fieldbcs) in zip(fields, bcs)
-        fill_halo_regions!(field, fieldbcs, arch, grid, args...)
     end
     return nothing
 end
@@ -43,3 +30,32 @@ function fill_halo_regions!(c::AbstractArray, fieldbcs, arch, grid, args...)
         fill_top_halo!(c, fieldbcs.z.top,    arch, grid, args...)
     return nothing
 end
+
+#####
+##### Convenience methods for filling halo regions of tupled fields
+#####
+
+function tupled_fill_halo_regions!(fields, bcs, arch, grid, args...)
+    for (field, fieldbcs) in zip(fields, bcs)
+        fill_halo_regions!(field, fieldbcs, arch, grid, args...)
+    end
+    return nothing
+end
+
+"""
+    fill_halo_regions!(fields, bcs, arch, grid)
+
+Fill halo regions for all fields in the `Tuple` `fields` according
+to the corresponding `Tuple` of `bcs`.
+"""
+fill_halo_regions!(fields::Tuple, bcs::Tuple, arch, grid, args...) =
+    tupled_fill_halo_regions!(fields, bcs, arch, grid, args...)
+
+"""
+    fill_halo_regions!(fields, bcs, arch, grid)
+
+Fill halo regions for all fields in the `NamedTuple` `fields` according
+to the corresponding `NamedTuple` of `bcs`.
+"""
+fill_halo_regions!(fields::NamedTuple{S}, bcs::NamedTuple{S}, arch, grid, args...) where S =
+    tupled_fill_halo_regions!(fields, bcs, arch, grid, args...)
