### 2020-01-28 4ecba07
Adds a restarter keyword to Model constructor
