### 2020-01-28 cafd236
Restarter does not have to be a part of the Model struct
