### 2020-01-28 9288c58
Adds documentation and tests
