+++ src/Coriolis/f_plane.jl
@@ -61,9 +61,35 @@ struct NonTraditionalFPlane{FT} <: AbstractRotation
     f′ :: FT
 end
 
-function NonTraditionalFPlane(FT=Float64; rotation_rate=nothing, latitude=nothing)
-    f  = 2rotation_rate*sind(latitude)
-    f′ = 2rotation_rate*cosd(latitude)
+"""
+    NonTraditionalFPlane([FT=Float64;] f=nothing, f′=nothing,
+                                       rotation_rate=Ω_Earth, latitude=nothing)
+
+Returns a parameter object for constant rotation about an arbitrary axis. The two
+perpendicular components of rotation are given by angular frequencies `f/2` and `f′/2`
+and therefore with background vorticities `f` and `f′`, respectively, which can be
+directly specified.
+
+In oceanography `f` and `f′` represent the components of planetary voriticity which
+are perpendicular and parallel to the surface, respectively.
+
+If `f` and `f′` are not specified, they are calculated from `rotation_rate` and `latitude`
+according to the relations `f = 2*rotation_rate*sind(latitude)` and
+`f′ = 2*rotation_rate*cosd(latitude)`, respectively. By default, `rotation_rate`
+is assumed to be Earth's.
+"""
+function NonTraditionalFPlane(FT=Float64; f=nothing, f′=nothing, rotation_rate=Ω_Earth, latitude=nothing)
+
+    use_f = !isnothing(f) && !isnothing(f′) && isnothing(latitude)
+    use_planet_parameters = !isnothing(latitude) && isnothing(f) && isnothing(f′)
+
+    if (use_planet_parameters && use_f) || (!use_planet_parameters && !use_f)
+        throw(ArgumentError("Either both rotation_rate and latitude must be " *
+                            "specified, *or* both f and f′ must be specified."))
+    elseif use_planet_parameters
+        f  = 2rotation_rate*sind(latitude)
+        f′ = 2rotation_rate*cosd(latitude)
+    end
     return NonTraditionalFPlane{FT}(f, f′)
 end
 
@@ -74,4 +100,4 @@ end
 @inline z_f_cross_U(i, j, k, grid, coriolis::NonTraditionalFPlane, U) = - coriolis.f′ * ℑxzᶜᵃᶠ(i, j, k, grid, U.u)
 
 Base.show(io::IO, f_plane::NonTraditionalFPlane{FT}) where FT =
-    println(io, @sprintf("Non-Traditional FPlane{$FT}: f = %.2e, f′ = %.2e", f_plane.f, f_plane.f′))
+    println(io, "Non-Traditional FPlane{$FT}:", @sprintf("f = %.2e, f′ = %.2e", f_plane.f, f_plane.f′))
