### 2020-01-29 03f32f7
Adds ability to set Poincare constant to a variable function in AMD
