### 2020-01-29 08e872d
Better plots to compare with.
