### 2020-01-29 585bcc0
Print min and max θ′ and w.
