### 2020-01-29 165bd0c
Switch nonlinear density current to dry adiabatic atmosphere.
