+++ verification/nonlinear_density_current/nonlinear_density_current.jl
@@ -9,10 +9,13 @@ Straka et al. (1993). "Numerical Solutions of a Nonlinear Density-Current -
 """
 
 using Printf
-using JULES, Oceananigans
 using Plots
+using Oceananigans
+using JULES
+using JULES: Π
 
 const km = 1000
+const hPa = 100
 
 Lx = 51.2km
 Lz = 6.4km
@@ -24,17 +27,17 @@ Ny = 1
 Nz = Int(Lz/Δ)
 
 grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
-                            x=(-Lx/2, Lx/2), y=(-1, 1), z=(0, Lz))
+                            x=(-Lx/2, Lx/2), y=(-Lx/2, Lx/2), z=(0, Lz))
 
 #####
 ##### Initial perturbation
 #####
 
-xᶜ, xʳ = 0, 4km
+xᶜ, xʳ = 0km, 4km
 zᶜ, zʳ = 3km, 2km
 
 function ΔT(x, y, z)
-    L = √( ((x-xᶜ)/xʳ)^2 + ((z-zᶜ)/zʳ)^2 )
+    L = √(((x - xᶜ) / xʳ)^2 + ((z - zᶜ) / zʳ)^2)
     L > 1 && return 0
     L ≤ 1 && return -15 * (1 + cos(π*L)) / 2
 end
@@ -43,30 +46,28 @@ end
 ##### Set up model
 #####
 
-pₛ = 100000
-Tₐ = 300
+gas = IdealGas()
+Rᵈ, cₚ, cᵥ = gas.Rᵈ, gas.cₚ, gas.cᵥ
+
 g  = 9.80665
+pₛ = 1000hPa
+θₛ = 300
+πₛ = 1
 
-buoyancy = IdealGas()
-Rᵈ, cₚ = buoyancy.Rᵈ, buoyancy.cₚ
-
-H = Rᵈ * Tₐ / g    # Scale height [m]
-ρₛ = pₛ / (Rᵈ*Tₐ)  # Surface density [kg/m³]
-
-p₀(x, y, z) = pₛ * exp(-z/H)
-ρ₀(x, y, z) = ρₛ * exp(-z/H)
-
-θ₀(x, y, z) = Tₐ * exp(z/H * Rᵈ/cₚ)
+θ₀(x, y, z) = θₛ
+π₀(x, y, z) = πₛ - g*z / (cₚ*θₛ)
+p₀(x, y, z) = pₛ * π₀(x, y, z)^(cₚ/Rᵈ)
+ρ₀(x, y, z) = pₛ / (Rᵈ * θ₀(x, y, z)) * π₀(x, y, z)^(cᵥ/Rᵈ)
 Θ₀(x, y, z) = ρ₀(x, y, z) * θ₀(x, y, z)
 
 const τ⁻¹ = 1     # Damping/relaxation time scale [s⁻¹]. This is very strong damping.
 const Δμ = 0.1Lz  # Sponge layer width [m] set to 10% of the domain height.
 @inline μ(z, Lz) = τ⁻¹ * exp(-(Lz-z) / Δμ)
 
-@inline Fw(i, j, k, grid, t, Ũ, C̃, p) = @inbounds -μ(grid.zF[k], grid.Lz) * Ũ.ρw[i, j, k]
+@inline Fw(i, j, k, grid, t, Ũ, C̃, p) = @inbounds (t <= 500) * -μ(grid.zF[k], grid.Lz) * Ũ.ρw[i, j, k]
 forcing = ModelForcing(w=Fw)
 
-model = CompressibleModel(grid=grid, buoyancy=buoyancy, reference_pressure=pₛ,
[…]
