### 2020-01-29 ddbe7dd
Properly initialize cold bubble perturbation.
