+++ verification/nonlinear_density_current/nonlinear_density_current.jl
@@ -20,7 +20,7 @@ const hPa = 100
 Lx = 51.2km
 Lz = 6.4km
 
-Δ = 200  # grid spacing [m]
+Δ = 100  # grid spacing [m]
 
 Nx = Int(Lx/Δ)
 Ny = 1
@@ -78,10 +78,10 @@ set!(model.tracers.Θᵐ, Θ₀)
 ##### Run a dry adiabatic atmosphere to hydrostatic balance
 #####
 
-# while model.clock.time < 500
-#     @printf("t = %.2f s\n", model.clock.time)
-#     time_step!(model, Δt=0.5, Nt=100)
-# end
+while model.clock.time < 500
+    @printf("t = %.2f s\n", model.clock.time)
+    time_step!(model, Δt=0.25, Nt=100)
+end
 
 #####
 ##### Now add the cold bubble perturbation.
@@ -107,7 +107,7 @@ for k in 1:Nz, i in 1:Nx
 end
 
 ρ_plot = contour(model.grid.xC ./ km, model.grid.zC ./ km, rotr90(ρ.data[1:Nx, 1, 1:Nz] .- ρʰᵈ),
-                 fill=true, levels=10, ylims=(0, 6.4), color=:balance, aspect_ratio=:equal, dpi=200)
+                 fill=true, levels=10, ylims=(0, 6.4), clims=(-0.05, 0.05), color=:balance, aspect_ratio=:equal, dpi=200)
 savefig(ρ_plot, "rho_prime_initial_condition.png")
 
 θ_slice = rotr90(Θ.data[1:Nx, 1, 1:Nz] ./ ρ.data[1:Nx, 1, 1:Nz])
@@ -119,23 +119,25 @@ savefig(Θ_plot, "theta_initial_condition.png")
 ##### Watch the density current evolve!
 #####
 
-# for i = 1:1000
-#     @printf("t = %.2f s\n", model.clock.time)
-#     time_step!(model, Δt=0.1, Nt=100)
-#
-#     xC, yC, zC = model.grid.xC ./ km, model.grid.yC ./ km, model.grid.zC ./ km
-#     xF, yF, zF = model.grid.xF ./ km, model.grid.yF ./ km, model.grid.zF ./ km
-#
-#     j = 1
-#     U_slice = rotr90(model.momenta.ρu.data[1:Nx, j, 1:Nz])
-#     W_slice = rotr90(model.momenta.ρw.data[1:Nx, j, 1:Nz])
-#     ρ_slice = rotr90(model.density.data[1:Nx, j, 1:Nz] .- ρ_hd)
-#     Θ_slice = rotr90(model.tracers.Θᵐ.data[1:Nx, j, 1:Nz] .- Θ_hd)
-#
-#     pU = contour(xC, zC, U_slice; fill=true, levels=10, color=:balance, clims=(-10, 10))
-#     pW = contour(xC, zC, W_slice; fill=true, levels=10, color=:balance, clims=(-10, 10))
-#     pρ = contour(xC, zC, ρ_slice; fill=true, levels=10, color=:balance, clims=(-0.05, 0.05))
-#     pΘ = contour(xC, zC, Θ_slice; fill=true, levels=10, color=:thermal)
-#
-#     display(plot(pU, pW, pρ, pΘ, layout=(4, 1), show=true))
-# end
+for n = 1:180
+    @printf("t = %.2f s\n", model.clock.time)
+    time_step!(model, Δt=0.1, Nt=50)
+
+    xC, yC, zC = model.grid.xC ./ km, model.grid.yC ./ km, model.grid.zC ./ km
+    xF, yF, zF = model.grid.xF ./ km, model.grid.yF ./ km, model.grid.zF ./ km
+
+    j = 1
[…]
