### 2020-01-30 5e682ac
Implements distinct Poincare constants for momentum and tracers
