+++ src/TurbulenceClosures/turbulence_closure_implementations/verstappen_anisotropic_minimum_dissipation.jl
@@ -4,38 +4,44 @@
 Parameters for the anisotropic minimum dissipation large eddy simulation model proposed by
 Verstappen (2018) and described by Vreugdenhil & Taylor (2018).
 """
-struct VerstappenAnisotropicMinimumDissipation{FT, P, K} <: AbstractAnisotropicMinimumDissipation{FT}
-     C :: P
+struct VerstappenAnisotropicMinimumDissipation{FT, PK, PN, K} <: AbstractAnisotropicMinimumDissipation{FT}
+    Cν :: PN
+    Cκ :: PK
     Cb :: FT
      ν :: FT
      κ :: K
 
-    function VerstappenAnisotropicMinimumDissipation{FT}(C, Cb, ν, κ) where FT
-        return new{FT, typeof(C), typeof(κ)}(C, Cb, ν, convert_diffusivity(FT, κ))
+    function VerstappenAnisotropicMinimumDissipation{FT}(Cν, Cκ, Cb, ν, κ) where FT
+        return new{FT, typeof(Cν), typeof(Cκ), typeof(κ)}(Cν, Cκ, Cb, ν, convert_diffusivity(FT, κ))
     end
 end
 
 const VAMD = VerstappenAnisotropicMinimumDissipation
 
 """
-    VerstappenAnisotropicMinimumDissipation(FT=Float64; C=1/12, Cb=0.0, ν=ν₀, κ=κ₀)
+    VerstappenAnisotropicMinimumDissipation(FT=Float64; C=1/12, Cν=nothing, Cκ=nothing,
+                                            Cb=0.0, ν=ν₀, κ=κ₀)
 
 Returns parameters of type `FT` for the `VerstappenAnisotropicMinimumDissipation`
 turbulence closure.
 
 Keyword arguments
 =================
-    - `C`  : Poincaré constant
+    - `C`  : Poincaré constant for both eddy viscosity and eddy diffusivities
+    - `Cν` : Poincaré constant for eddy viscosity
+    - `Cκ` : Poincaré constant for eddy diffusivities
     - `Cb` : Buoyancy modification multiplier (`Cb = 0` turns it off, `Cb = 1` turns it on)
     - `ν`  : Constant background viscosity for momentum
     - `κ`  : Constant background diffusivity for tracer. Can either be a single number
              applied to all tracers, or `NamedTuple` of diffusivities corresponding to each
              tracer.
 
-By default, `C` = 1/12, which is appropriate for a finite-volume method employing a
+By default: `C = Cν = Cκ` = 1/12, which is appropriate for a finite-volume method employing a
 second-order advection scheme, `Cb` = 0, which terms off the buoyancy modification term,
-the molecular viscosity of seawater at 20 deg C and 35 psu is used for `ν`, and the molecular
-diffusivity of heat in seawater at 20 deg C and 35 psu is used for `κ`.
+the molecular viscosity of seawater at 20 deg C and 35 psu is used for `ν`, and 
+the molecular diffusivity of heat in seawater at 20 deg C and 35 psu is used for `κ`.
+
[…]
