### 2020-01-30 #2018 @funsim
Overall this software package is very well set up and I recommend the submission to be accepted by JOSS after addressing the following comments:

**Software paper**
I cannot find the software paper - could you please send me a link?

**Performance**
* The performance plots on the README.md file show that the single precision speed up for CPUs is below 1. This is surprising and should be commented on.
* The link to benchmarks.jl in the README.md file is broken.

**Functionality documentation**
* Citation commands are not working in documentation page:
https://climate-machine.github.io/Oceananigans.jl/stable/numerical_implementation/large_eddy_simulation/
* I suggest to link the API documentation with the mathematical description. For instance, you could mention the Julia commands when describing the numerical handling of boundary conditions in  https://climate-machine.github.io/Oceananigans.jl/stable/numerical_implementation/boundary_conditions/

**Functionality:**
* It seems that you have not performed a convergence test of your numerical model (based on analytical known solutions or manufactured solutions). I suggest that you add this (for instance based on the Taylor Green example)

I would like to thank @jakobes for helping me in this review.
