### 2020-01-30 #2018 @kthyng
@funsim You can find the paper above by clicking "check article proof"

Please note, for both @funsim and @ali-ramadhan: if there will be more discussion on the points noted above, it's preferable to have that discussion in individual issues that are opened and linked to this review issue. That way this issue remains more of an overview and details can be hashed out elsewhere. Thanks!
