### 2020-01-31 35a4c23
Stommel gyre solution
