### 2020-01-31 6c31dde
Fix typos in advective flux operators
