### 2020-01-31 1b9639b
Creating 2D Stommel gyre animations
