+++ verification/tracer_advection_2d/tracer_advection_2d.jl
@@ -1,6 +1,8 @@
 using Printf
 using OffsetArrays
 using DifferentialEquations
+
+ENV["GKSwstype"] = "nul"
 using Plots
 
 include("wind_driven_gyres.jl")
@@ -50,8 +52,8 @@ end
 ##### Initial (= final) conditions
 #####
 
-@inline ϕ_Gaussian(x, y; A, σˣ, σʸ) = A * exp(-x^2/(2σˣ^2) -y^2/(2σʸ^2))
-@inline ϕ_Square(x, y; A, σˣ, σʸ)   = A * (-σˣ <= x <= σˣ) * (-σʸ <= y <= σʸ)
+@inline ϕ_Gaussian(x, y; L, A, σˣ, σʸ) = A * exp(-(x-L/2)^2/(2σˣ^2) -(y-L/2)^2/(2σʸ^2))
+@inline ϕ_Square(x, y; L, A, σˣ, σʸ)   = A * (-σˣ <= x-L/2 <= σˣ) * (-σʸ <= y-L/2 <= σʸ)
 
 ic_name(::typeof(ϕ_Gaussian)) = "Gaussian"
 ic_name(::typeof(ϕ_Square))   = "Square"
@@ -60,8 +62,9 @@ ic_name(::typeof(ϕ_Square))   = "Square"
 ##### Experiment functions
 #####
 
-function setup_problem(Nx, Ny, T, CFL, ϕₐ, time_stepper, scheme; u, v,
-                       L=4000km, τ₀=1, β=1e-11, r=0.04*β*L, A=1, σˣ=10km, σʸ=10km)
+function setup_problem(Nx, Ny, T, CFL, ϕₐ, time_stepper, scheme;
+                       u, v, L=4000km, τ₀=1, β=1e-11, r=0.04*β*L,
+                       A=1, σˣ=10km, σʸ=10km)
     Δx = L/Nx
     Δy = L/Ny
     Hx = Hy = 3
@@ -79,29 +82,55 @@ function setup_problem(Nx, Ny, T, CFL, ϕₐ, time_stepper, scheme; u, v,
     for j in 1:Ny, i in 1:Nx
         u[i, j] = u_Stommel(xF[i], yF[j], L=L, τ₀=τ₀, β=β, r=r)
         v[i, j] = v_Stommel(xF[i], yF[j], L=L, τ₀=τ₀, β=β, r=r)
-        ϕ[i, j] = ϕₐ(xC[i], yC[j], A=A, σˣ=σˣ, σʸ=σʸ)
+        ϕ[i, j] = ϕₐ(xC[i], yC[j], L=L, A=A, σˣ=σˣ, σʸ=σʸ)
     end
 
-    @show Δx, Δy
-    @show maximum(abs, u)
-    @show maximum(abs, v)
     u = u ./ maximum(abs, u)
     v = v ./ maximum(abs, v)
     Δt = CFL * min(Δx, Δy) / max(maximum(abs, u), maximum(abs, v))
-    @show Δt
+    
+    @info "Nx=$Nx, Ny=$Ny, Δx=$Δx, Δy=$Δy, Δt=$Δt"
 
     tspan = (0.0, T)
     params = (Nx=Nx, Ny=Ny, Hx=Hx, Hy=Hy, Δx=Δx, Δy=Δy, u=u, v=v, scheme=scheme)
     return xC, yC, Δt, ODEProblem(advection!, ϕ, tspan, params)
 end
 
+function create_animation(Nx, Ny, T, CFL, ϕₐ, time_stepper, scheme;
+                          u, v, L=1000km, τ₀=1, β=1e-11, r=0.04*β*L,
[…]
