### 2020-01-31 8e8e04d
Properly normalize divergence-free velocity field
