### 2020-01-31 9e9cfc1
Fix plot orientation
