### 2020-01-31 dcf67e1
Flesh out and document WENO interpolation schemes
