+++ verification/tracer_advection_2d/weno_interpolation.jl
@@ -1,5 +1,6 @@
 """
-WENO interpolation schemes following Shu (§2, 2009).
+WENO interpolation schemes following Shu (§2, 2009). Combining k interpolants
+each of order k will produce a scheme of order 2k+1.
 
 Shu (2009) "High Order Weighted Essentially Nonoscillatory Schemes for
     Convection Dominated Problems", SIAM Review 51(1), pp. 82–126.
@@ -8,14 +9,19 @@ Shu (2009) "High Order Weighted Essentially Nonoscillatory Schemes for
 
 const Δx = 1
 x(i) = i*Δx
-ℓ(i, k, x′) = prod([(x′ - x(j)) // (x(i) - x(j)) for j in 0:k if j != i])
-csw(k, s) = [ℓ(i, k, s + 1//2) for i in 0:k]
 
-ssign(n) = n >= 0 ? "+" : "-"
-ssubscript(n) = n == 0 ? "i" : "i" * ssign(n) * string(abs(n))
+# Lagrange interpolants or Lagrange basis polynomials
+ℓ(i, k, x′) = prod([(x′ - x(j)) // (x(i) - x(j)) for j in 0:k if j != i])
+
+# Return array of coefficients for interpolant of order k centered around x = s + 1/2.
+c_ks(k, s) = [ℓ(i, k, s + 1//2) for i in 0:k]
 
 function print_interpolant(k, s)
     cs = csw(k, s)
+
+    ssign(n) = n >= 0 ? "+" : "-"
+    ssubscript(n) = n == 0 ? "i" : "i" * ssign(n) * string(abs(n))
+
     print("u(i+1//2) = ")
     for (j, c) in enumerate(cs)
         c_s = ssign(c) * " " * string(abs(c))
@@ -31,12 +37,12 @@ function print_interpolants(k)
     end
 end
 
-println("WENO-3 [Compare with Shu (2009) equations (2.1)-(3.3)]:")
+println("WENO-5 [Compare with Shu (2009) equations (2.1)-(3.3)]:")
 print_interpolants(2)
 
-println("\nWENO-5:")
+println("\nWENO-7:")
 print_interpolants(4)
 
-println("\nWENO-7:")
+println("\nWENO-9:")
 print_interpolants(6)
 
