### 2020-01-31 474baff
Clean up 2D advection script a bit
