### 2020-01-31 CliMA/Oceananigans.jl#607 @ali-ramadhan
Update performance benchmarks plot and discuss results
> The performance plots on the README.md file show that the single precision speed up for CPUs is below 1. This is surprising and should be commented on.
>
> _Originally posted by @funsim in https://github.com/openjournals/joss-reviews/issues/2018#issuecomment-580286479_
