### 2020-01-31 CliMA/Oceananigans.jl#608 @ali-ramadhan
Update README and fix broken links
> The link to benchmarks.jl in the README.md file is broken.
>
> _Originally posted by @funsim in https://github.com/openjournals/joss-reviews/issues/2018#issuecomment-580286479_
