### 2020-01-31 CliMA/Oceananigans.jl#609 @ali-ramadhan
Fix citations and references in documentation
> Citation commands are not working in documentation page:
https://climate-machine.github.io/Oceananigans.jl/stable/numerical_implementation/large_eddy_simulation/
>
> _Originally posted by @funsim in https://github.com/openjournals/joss-reviews/issues/2018#issuecomment-580286479_
