### 2020-01-31 CliMA/Oceananigans.jl#610 @ali-ramadhan
Link documentation of numerical methods with code and API
> I suggest to link the API documentation with the mathematical description. For instance, you could mention the Julia commands when describing the numerical handling of boundary conditions in https://climate-machine.github.io/Oceananigans.jl/stable/numerical_implementation/boundary_conditions/
>
> _Originally posted by @funsim in https://github.com/openjournals/joss-reviews/issues/2018#issuecomment-580286479_
