### 2020-01-31 CliMA/Oceananigans.jl#611 @ali-ramadhan
Taylor-Green vortex verification experiment and convergence test
> It seems that you have not performed a convergence test of your numerical model (based on analytical known solutions or manufactured solutions). I suggest that you add this (for instance based on the Taylor Green example)
>
> _Originally posted by @funsim in https://github.com/openjournals/joss-reviews/issues/2018#issuecomment-580286479_
