### 2020-01-31 #2018 @ali-ramadhan
Thank you @funsim for the quick review! You bring up good points that will improve the package. Certainly the documentation needs more work as you've pointed out, and we've been meaning to add a verification experiment with some convergence results so thank you for bringing it up.

I have opened an issue for each of your comments where we will work towards addressing them:
1. https://github.com/climate-machine/Oceananigans.jl/issues/607
2. https://github.com/climate-machine/Oceananigans.jl/issues/608
3. https://github.com/climate-machine/Oceananigans.jl/issues/609
4. https://github.com/climate-machine/Oceananigans.jl/issues/610
5. https://github.com/climate-machine/Oceananigans.jl/issues/611
