### 2020-02-01 90cc192
Add closure and diffusivities to CompressibleModel
