+++ src/Operators/Operators.jl
@@ -6,9 +6,9 @@ export
     ρᵐ, ρᵈ_over_ρᵐ,
     x_f_cross_U, y_f_cross_U, z_f_cross_U,
     hdivᶜᶜᵃ, divᶜᶜᶜ, ∇²,
-    div_flux, div_κ∇c,
+    div_flux, ∇_κ_∇c,
     div_ρuũ, div_ρvũ, div_ρwũ,
-    div_μ∇u, div_μ∇v, div_μ∇w
+    ∂ⱼ_2ν_Σ₁ⱼ, ∂ⱼ_2ν_Σ₂ⱼ, ∂ⱼ_2ν_Σ₃ⱼ
 
 include("compressible_operators.jl")
 
