### 2020-02-01 e876148
Clean up and update slow forcing terms
