+++ verification/tracer_advection_2d/weno_reconstruction.jl
@@ -69,4 +69,3 @@ print_interpolants(4)
 
 println("\nWENO-9 [Compare with Table 2.1 of Shu (1998) and equation (2.14) of Shu (2009)]:")
 print_interpolants(5)
-
