### 2020-02-01 51d2d12
Update verification experiments to use a closure
