### 2020-02-01 5c695b6
Add CFL utility.
