+++ src/JULES.jl
@@ -21,5 +21,6 @@ include("models.jl")
 include("right_hand_sides.jl")
 include("time_stepping_kernels.jl")
 include("time_stepping.jl")
+include("utils.jl")
 
 end # module
