### 2020-02-01 302534c
Return CFL number instead of cell advection timescale
