### 2020-02-01 c2d063b
Add acoustic CFL utility.
