+++ .appveyor.yml
@@ -1,7 +1,7 @@
 # Documentation: https://github.com/JuliaCI/Appveyor.jl
 environment:
   matrix:
-  - julia_version: 1.1
+  - julia_version: 1.3
   - julia_version: nightly
 
 platform:
