+++ src/models.jl
@@ -3,11 +3,11 @@ using Oceananigans
 using Oceananigans.Models: AbstractModel, Clock
 using Oceananigans.Forcing: zeroforcing
 
-####
-#### Definition of a compressible model
-####
+#####
+##### Definition of a compressible model
+#####
 
-mutable struct CompressibleModel{A, FT, G, M, D, T, TS, B, C, TC, DF, F, P, SF, RHS, IV, ATS} <: AbstractModel
+mutable struct CompressibleModel{A, FT, G, M, D, T, TS, MX, B, C, TC, DF, F, P, SF, RHS, IV, ATS} <: AbstractModel
              architecture :: A
                      grid :: G
                     clock :: Clock{FT}
@@ -15,6 +15,7 @@ mutable struct CompressibleModel{A, FT, G, M, D, T, TS, B, C, TC, DF, F, P, SF, 
                   density :: D
    prognostic_temperature :: T
                   tracers :: TS
+            mixing_ratios :: MX
                  buoyancy :: B
                  coriolis :: C
                   closure :: TC
@@ -28,9 +29,9 @@ mutable struct CompressibleModel{A, FT, G, M, D, T, TS, B, C, TC, DF, F, P, SF, 
     acoustic_time_stepper :: ATS
 end
 
-####
-#### Constructor for compressible models
-####
+#####
+##### Constructor for compressible models
+#####
 
 function CompressibleModel(;
                      grid,
@@ -40,7 +41,8 @@ function CompressibleModel(;
                   momenta = MomentumFields(architecture, grid),
                   density = CellField(architecture, grid),
    prognostic_temperature = ModifiedPotentialTemperature(),
-                  tracers = (:Θᵐ, :Qv, :Ql, :Qi),
+                  tracers = (:Θᵐ,),
+            mixing_ratios = (:Qv, :Ql, :Qi),
                  buoyancy = IdealGas(float_type),
                  coriolis = nothing,
                   closure = ConstantIsotropicDiffusivity(float_type, ν=0.5, κ=0.5),
@@ -58,19 +60,20 @@ function CompressibleModel(;
 
     reference_pressure = float_type(reference_pressure)
     tracers = TracerFields(architecture, grid, tracers)
+    mixing_ratios = MixingRatioFields(arch, grid, mixing_ratios)
 
     forcing = ModelForcing(tracernames(tracers), forcing)
     closure = with_tracers(tracernames(tracers), closure)
 
     return CompressibleModel(architecture, grid, clock, momenta, density, prognostic_temperature,
-                             tracers, buoyancy, coriolis, closure, diffusivities, forcing, parameters,
-                             reference_pressure, slow_forcings, right_hand_sides, intermediate_vars,
-                             acoustic_time_stepper)
+                             tracers, mixing_ratios, buoyancy, coriolis, closure, diffusivities,
+                             forcing, parameters, reference_pressure, slow_forcings, right_hand_sides,
+                             intermediate_vars, acoustic_time_stepper)
 end
 
-####
-#### Utilities for constructing compressible models
-####
+#####
+##### Utilities for constructing compressible models
+#####
 
 function MomentumFields(arch, grid)
     ρu = FaceFieldX(arch, grid)
@@ -87,6 +90,11 @@ tracernames(::NamedTuple{names}) where names = tracernames(names)
 function TracerFields(arch, grid, tracernames)
     tracerfields = Tuple(CellField(arch, grid) for c in tracernames)
     return NamedTuple{tracernames}(tracerfields)
+end
+
+function MixingRatioFields(arch, grid, mixing_ratio_names)
[…]
