### 2020-02-01 22ca111
Simplify and generalize prognostic temperature validation.
