### 2020-02-01 77f6664
Test that CompressibleModel constructor throws exceptions when it should
