+++ test/test_models.jl
@@ -5,14 +5,27 @@ using Oceananigans, JULES
 
     @testset "Modified potential temperature" begin
         @info "  Testing model construction with Θᵐ prognostic temperature..."
-        model = CompressibleModel(grid=RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1)))
-        @test isa(model, CompressibleModel)
+
+        grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
+        pt   = ModifiedPotentialTemperature()
+
+        model = CompressibleModel(grid=grid, prognostic_temperature=pt, tracers=(:Θᵐ,))
+        @test model isa CompressibleModel
+
+        @test_throws(ArgumentError,
+            CompressibleModel(grid=grid, prognostic_temperature=pt, tracers=(:T,)))
     end
 
     @testset "Entropy" begin
         @info "  Testing model construction with S prognostic temperature..."
-        model = CompressibleModel(grid=RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1)),
-                                  prognostic_temperature=Entropy(), tracers=(:S,))
-        @test isa(model, CompressibleModel)
+
+        grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
+        pt   = Entropy()
+
+        model = CompressibleModel(grid=grid, prognostic_temperature=pt, tracers=(:S,))
+        @test model isa CompressibleModel
+
+        @test_throws(ArgumentError,
+            CompressibleModel(grid=grid, prognostic_temperature=pt, tracers=(:T,)))
     end
 end
