### 2020-02-01 b31411c
Cleanup and test model with placeholder microphysics schemes.
