+++ src/microphysics.jl
@@ -5,7 +5,7 @@ struct VaporLiquidIcePlaceholder <: AbstractMicrophysics end
 validate_microphysics(::Nothing, tracers) = true
 
 missing_tracers_error(names, mp) =
-    "Must specify $names as tracers to use $mp microphysics."
+    "Must specify $names as tracers to use $(typeof(mp)) microphysics."
 
 required_tracers(::VaporPlaceholder) = (:Qv,)
 required_tracers(::VaporLiquidIcePlaceholder) = (:Qv, :Ql, :Qi)
@@ -13,10 +13,7 @@ required_tracers(::VaporLiquidIcePlaceholder) = (:Qv, :Ql, :Qi)
 function validate_microphysics(microphysics, tracers)
     C̃ = required_tracers(microphysics)
     for c in C̃
-        if c ∉ tracers
-            throw(ArgumentError(missing_tracers_error(C̃, microphysics)))
-        else
-            return true
-        end
+        c ∉ tracers && throw(ArgumentError(missing_tracers_error(C̃, microphysics)))
     end
+    return true
 end
