+++ src/time_stepping.jl
@@ -52,6 +52,7 @@ function time_step!(model::CompressibleModel; Δt, Nt=1)
     IV = model.intermediate_vars
 
     pₛ = model.reference_pressure
+    g  = model.gravity
 
     # On third RK3 step, we update Φ⁺ instead of model.intermediate_vars
     Φ⁺ = merge(Ũ, C̃, (ρ=ρᵈ,))
@@ -79,12 +80,12 @@ function time_step!(model::CompressibleModel; Δt, Nt=1)
 
             @debug "  Computing right hand sides..."
             if rk3_iter == 1
-                compute_rhs_args = (R, grid, prog_temp, buoyancy, pₛ, microphysics, ρᵈ, Ũ, C̃, F)
+                compute_rhs_args = (R, grid, prog_temp, buoyancy, microphysics, pₛ, g, ρᵈ, Ũ, C̃, F)
                 fill_halo_regions!(ρᵈ.data, hpbcs, arch, grid)
                 fill_halo_regions!(datatuple(merge(Ũ, C̃)), hpbcs, arch, grid)
                 fill_halo_regions!(Ũ.ρw.data, hpbcs_np, arch, grid)
             else
-                compute_rhs_args = (R, grid, prog_temp, buoyancy, pₛ, microphysics, IV.ρ, IV_Ũ, IV_C̃, F)
+                compute_rhs_args = (R, grid, prog_temp, buoyancy, microphysics, pₛ, g, IV.ρ, IV_Ũ, IV_C̃, F)
                 fill_halo_regions!(IV.ρ.data, hpbcs, arch, grid)
                 fill_halo_regions!(datatuple(merge(IV_Ũ, IV_C̃)), hpbcs, arch, grid)
                 fill_halo_regions!(IV_Ũ.ρw.data, hpbcs_np, arch, grid)
