### 2020-02-01 7b14e99
Fixes bug in type signature of AMD paramter struct constructor
