+++ src/Grids/Grids.jl
@@ -4,6 +4,35 @@ export RegularCartesianGrid, VerticallyStretchedCartesianGrid
 
 using Oceananigans
 
+"""
+    AbstractTopology
+
+Abstract supertype for grid topologies.
+"""
+abstract type AbstractTopology end
+
+"""
+    Periodic
+
+Grid topology for periodic dimensions.
+"""
+struct Periodic <: AbstractTopology end
+
+"""
+    Bounded
+
+Grid topology for bounded dimensions. These could be wall-bounded dimensions
+or dimensions
+"""
+struct Bounded <: AbstractTopology end
+
+"""
+    Singleton
+
+Grid topology for singleton dimensions with one grid point.
+"""
+struct Singleton <: AbstractTopology end
+
 include("grid_utils.jl")
 include("regular_cartesian_grid.jl")
 include("vertically_stretched_cartesian_grid.jl")
