### 2020-02-03 71973bc
Get rid of grid.Tx, Ty, Tz
