+++ test/test_operators.jl
@@ -77,7 +77,7 @@ end
         grid = RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz))
 
         Hx, Hy, Hz = grid.Hx, grid.Hy, grid.Hz
-        Tx, Ty, Tz = grid.Tx, grid.Ty, grid.Tz
+        Tx, Ty, Tz = Nx+2Hx, Ny+2Hy, Nz+2Hz
 
         A3 = OffsetArray(zeros(Tx, Ty, Tz), 1-Hx:Nx+Hx, 1-Hy:Ny+Hy, 1-Hz:Nz+Hz)
         @. @views A3[1:Nx, 1:Ny, 1:Nz] = rand()
