### 2020-02-03 504b8d2
Move `AbstractGrid` to Grid submodule and export it
