+++ src/TimeSteppers/adams_bashforth.jl
@@ -1,7 +1,5 @@
 import Oceananigans.OutputWriters: saveproperty!
 
-using Oceananigans.Utils: datatuples
-
 """
     AdamsBashforthTimeStepper(float_type, arch, grid, tracers, χ)
 
