### 2020-02-03 cdf2b46
Kill off regular temperature as a thermodynamic variable
