### 2020-02-03 cafd654
Change prognostic temperature to thermodynamic variable
