### 2020-02-03 4a28cc5
Add topology kwarg and validate topology input
