### 2020-02-03 ad9166c
Generalize grid property functions
