### 2020-02-03 dffa04f
Do not export BoundaryConditions.Periodic for now
