+++ test/test_forcings.jl
@@ -14,7 +14,7 @@ function time_step_with_forcing_functions(arch)
 
     forcing = ModelForcing(u=Fu, v=Fv, w=Fw)
 
-    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
+    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1), topology=DT)
     model = Model(grid=grid, architecture=arch, forcing=forcing)
     time_step!(model, 1, 1)
     return true
@@ -27,7 +27,7 @@ function time_step_with_forcing_functions_params(arch)
 
     forcing = ModelForcing(u=Fu, v=Fv, w=Fw)
 
-    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
+    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1), topology=DT)
     model = Model(grid=grid, architecture=arch, forcing=forcing, parameters=(τ=60,))
     time_step!(model, 1, 1)
     return true
@@ -39,7 +39,7 @@ function time_step_with_forcing_functions_sin_exp(arch)
 
     forcing = ModelForcing(u=Fu, T=FT)
 
-    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
+    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1), topology=DT)
     model = Model(grid=grid, architecture=arch, forcing=forcing)
     time_step!(model, 1, 1)
     return true
@@ -47,7 +47,7 @@ end
 
 function time_step_with_simple_forcing(arch)
     u_forcing = SimpleForcing((x, y, z, t) -> sin(x))
-    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
+    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1), topology=DT)
     model = Model(grid=grid, architecture=arch, forcing=ModelForcing(u=u_forcing))
     time_step!(model, 1, 1)
     return true
@@ -55,7 +55,7 @@ end
 
 function time_step_with_simple_forcing_parameters(arch)
     u_forcing = SimpleForcing((x, y, z, t, p) -> sin(p.ω * x), parameters=(ω=π,))
-    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
+    grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1), topology=DT)
     model = Model(grid=grid, architecture=arch, forcing=ModelForcing(u=u_forcing))
     time_step!(model, 1, 1)
     return true

[…]
