### 2020-02-04 8d3b164
Update examples and verification experiments
