+++ examples/internal_wave.jl
@@ -69,8 +69,11 @@ nothing # hide
 # use temperature as a buoyancy tracer, and use a small constant viscosity
 # and diffusivity to stabilize the model.
 
+grid = RegularCartesianGrid(size=(Nx, 1, Nx), length=(Lx, Lx, Lx),
+                            topology=(Periodic, Singleton, Bounded))
+
 model = Model(
-        grid = RegularCartesianGrid(size=(Nx, 1, Nx), length=(Lx, Lx, Lx)),
+        grid = grid,
      closure = ConstantIsotropicDiffusivity(ν=1e-6, κ=1e-6),
     coriolis = FPlane(f=f),
      tracers = :b,
