### 2020-02-04 90993ca
Changes notation for non-traditional Coriolis force, updates documentation and tests
