+++ src/Coriolis/f_plane.jl
@@ -57,47 +57,47 @@ A Coriolis implementation that facilitates non-traditional Coriolis terms in the
 and vertical momentum equations along with the traditional Coriolis terms.
 """
 struct NonTraditionalFPlane{FT} <: AbstractRotation
-    f  :: FT
-    f′ :: FT
+    fz :: FT
+    fy :: FT
 end
 
 """
-    NonTraditionalFPlane([FT=Float64;] f=nothing, f′=nothing,
+    NonTraditionalFPlane([FT=Float64;] fz=nothing, fy=nothing,
                                        rotation_rate=Ω_Earth, latitude=nothing)
 
-Returns a parameter object for constant rotation about an arbitrary axis. The two
-perpendicular components of rotation are given by angular frequencies `f/2` and `f′/2`
-and therefore with background vorticities `f` and `f′`, respectively, which can be
-directly specified.
+Returns a parameter object for constant rotation about an axis in the `y-z` plane
+with `y`- and `z`-components `fy/2` and `fz/2`, and the background vorticity is
+`(0, fy, fz)`.
 
-In oceanography `f` and `f′` represent the components of planetary voriticity which
-are perpendicular and parallel to the surface, respectively.
+In oceanography `fz` and `fy` represent the components of planetary voriticity which
+are perpendicular and parallel to the ocean surface in a domain in which `x, y, z` 
+correspond to the directions east, north, and up.
 
-If `f` and `f′` are not specified, they are calculated from `rotation_rate` and `latitude`
-according to the relations `f = 2*rotation_rate*sind(latitude)` and
-`f′ = 2*rotation_rate*cosd(latitude)`, respectively. By default, `rotation_rate`
+If `fz` and `fy` are not specified, they are calculated from `rotation_rate` and `latitude`
+according to the relations `fz = 2*rotation_rate*sind(latitude)` and
+`fy = 2*rotation_rate*cosd(latitude)`, respectively. By default, `rotation_rate`
 is assumed to be Earth's.
 """
-function NonTraditionalFPlane(FT=Float64; f=nothing, f′=nothing, rotation_rate=Ω_Earth, latitude=nothing)
+function NonTraditionalFPlane(FT=Float64; fz=nothing, fy=nothing, rotation_rate=Ω_Earth, latitude=nothing)
 
-    use_f = !isnothing(f) && !isnothing(f′) && isnothing(latitude)
-    use_planet_parameters = !isnothing(latitude) && isnothing(f) && isnothing(f′)
+    use_f = !isnothing(fz) && !isnothing(fy) && isnothing(latitude)
+    use_planet_parameters = !isnothing(latitude) && isnothing(fz) && isnothing(fy)
 
     if (use_planet_parameters && use_f) || (!use_planet_parameters && !use_f)
         throw(ArgumentError("Either both rotation_rate and latitude must be " *
                             "specified, *or* both f and f′ must be specified."))
     elseif use_planet_parameters
-        f  = 2rotation_rate*sind(latitude)
-        f′ = 2rotation_rate*cosd(latitude)
+        fz = 2rotation_rate*sind(latitude)
[…]
