### 2020-02-04 d70f84d
Fixes bug in model setup documentation
