### 2020-02-04 699867f
Revert "Update examples and verification experiments"
This reverts commit 8d3b1647c30cc310ec680a5ba2f6a5169b482012.
