+++ verification/stratified_couette_flow/stratified_couette_flow.jl
@@ -109,12 +109,9 @@ function simulate_stratified_couette_flow(; Nxy, Nz, arch=GPU(), h=1, U_wall=1, 
     ##### Non-dimensional model setup
     #####
 
-    grid = RegularCartesianGrid(size=(Nxy, Nxy, Nz), length=(4π*h, 2π*h, 2h),
-                                topology=(Periodic, Periodic, Bounded))
-
     model = Model(
        architecture = arch,
-               grid = grid,
+               grid = RegularCartesianGrid(size = (Nxy, Nxy, Nz), length = (4π*h, 2π*h, 2h)),
            buoyancy = SeawaterBuoyancy(equation_of_state=LinearEquationOfState(α=1.0, β=0.0)),
             closure = AnisotropicMinimumDissipation(ν=ν, κ=κ),
 boundary_conditions = HorizontallyPeriodicSolutionBCs(u=ubcs, v=vbcs, T=Tbcs),
