+++ examples/langmuir_turbulence.jl
@@ -1,9 +1,9 @@
 # # Langmuir turbulence example
 #
 # This example implements the Langmuir turbulence simulation reported in section
-# 4 of McWilliams, J. C. et al., "Langmuir Turbulence in the ocean," Journal of
+# 4 of McWilliams, J. C. et al., "Langmuir Turbulence in the ocean," Journal of 
 # Fluid Mechanics (1997). This example demonstrates:
-#
+#   
 #   * how to run a simulation with surface wave effects via the Craik-Leibovich
 #     approximation
 #
@@ -17,7 +17,7 @@ using Oceananigans.SurfaceWaves: UniformStokesDrift
 # ## Model parameters
 #
 # Here we use an anisotropic grid with `Nh` horizontal grid points, `Nz` vertical
-# grid points, `Δh` horizontal grid spacing, and `Δz` vertical grid spacing.
+# grid points, `Δh` horizontal grid spacing, and `Δz` vertical grid spacing. 
 # We specify fluxes of buouyancy and momentum that match the specifications in
 # McWilliams and Sullivan (1997) via
 #
@@ -39,7 +39,7 @@ using Oceananigans.SurfaceWaves: UniformStokesDrift
        f = 1e-4     # [s⁻¹] Coriolis parameter
 end_time = 2hour    # [s] End time for the simulation
 
-# Surface wave stokes drift profile:
+# Surface wave stokes drift profile: 
 #
 # $ uˢ(z) = Uˢ exp(2kˢʷ z) $
 #
@@ -61,7 +61,7 @@ const aˢʷ = 0.8      # [m] Surface wave amplitude
 const Uˢ = aˢʷ^2 * kˢʷ * sqrt(g_Earth * kˢʷ)
 
 ## Oceananigans
-∂z_uˢ(z, t) = 2kˢʷ * Uˢ * exp(2kˢʷ * z)
+∂z_uˢ(z, t) = 2kˢʷ * Uˢ * exp(2kˢʷ * z) 
 
 # We use the Stokes drift profile for the initial condition
 uˢ(z) = Uˢ * exp(2kˢʷ * z)
@@ -75,20 +75,19 @@ uˢ(z) = Uˢ * exp(2kˢʷ * z)
 
 u_bcs = HorizontallyPeriodicBCs(top = BoundaryCondition(Flux, Qᵘ))
 
-b_bcs = HorizontallyPeriodicBCs(top = BoundaryCondition(Flux, Qᵇ),
+b_bcs = HorizontallyPeriodicBCs(top = BoundaryCondition(Flux, Qᵇ), 
                                 bottom = BoundaryCondition(Gradient, N²))
 
 # ## Model instantiation
 #
 # We instantiate a horizontally-periodic `Model` on the CPU with on a `RegularCartesianGrid`,
-# using a `FPlane` model for rotation (constant rotation rate), the Anisotropic Minimum
-# Dissipation closure to model the effects of unresolved turbulence, and the previously defined
+# using a `FPlane` model for rotation (constant rotation rate), the Anisotropic Minimum 
+# Dissipation closure to model the effects of unresolved turbulence, and the previously defined 
 # boundary conditions for `u` and `b`.
 
 model = Model(
          architecture = CPU(),
-                 grid = RegularCartesianGrid(size=(Nh, Nh, Nz), length=(Δh*Nh, Δh*Nh, Δz*Nz),
-                                             topology=(Periodic, Periodic, Bounded)),
+                 grid = RegularCartesianGrid(size=(Nh, Nh, Nz), length=(Δh*Nh, Δh*Nh, Δz*Nz)),
              buoyancy = BuoyancyTracer(), tracers = :b,
[…]
