+++ examples/netcdf_ouput_example.jl
@@ -8,9 +8,8 @@ Nx, Ny, Nz = 16, 16, 16       # No. of grid points in x, y, and z, respectively.
 Lx, Ly, Lz = 100, 100, 100    # Length of the domain in x, y, and z, respectively (m).
 tf = 5000                     # Length of the simulation (s)
 
-grid = RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz),
-                            topology=(Periodic, Periodic, Bounded))
-model = Model(grid=grid, closure=ConstantIsotropicDiffusivity())
+model = Model(grid=RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz)),
+              closure=ConstantIsotropicDiffusivity())
 
 # Add a cube-shaped warm temperature anomaly that takes up the middle 50%
 # of the domain volume.
