+++ test/test_diagnostics.jl
@@ -1,5 +1,5 @@
 function horizontal_average_is_correct(arch, FT)
-    grid = RegularCartesianGrid(size=(16, 16, 16), length=(100, 100, 100), topology=DT)
+    grid = RegularCartesianGrid(size=(16, 16, 16), length=(100, 100, 100))
     model = Model(grid=grid, architecture=arch, float_type=FT)
 
     T₀(x, y, z) = 20 + 0.01*z
@@ -13,7 +13,7 @@ function horizontal_average_is_correct(arch, FT)
 end
 
 function nan_checker_aborts_simulation(arch, FT)
-    grid=RegularCartesianGrid(size=(16, 16, 2), length=(1, 1, 1), topology=DT)
+    grid=RegularCartesianGrid(size=(16, 16, 2), length=(1, 1, 1))
     model = Model(grid=grid, architecture=arch, float_type=FT)
 
     # It checks for NaNs in w by default.
@@ -26,14 +26,14 @@ function nan_checker_aborts_simulation(arch, FT)
 end
 
 TestModel(::GPU, FT, ν=1.0, Δx=0.5) =
-    Model(grid = RegularCartesianGrid(FT, size=(16, 16, 16), length=(16Δx, 16Δx, 16Δx), topology=DT),
+    Model(grid = RegularCartesianGrid(FT; size=(16, 16, 16), length=(16Δx, 16Δx, 16Δx)),
        closure = ConstantIsotropicDiffusivity(FT; ν=ν, κ=ν),
   architecture = GPU(),
     float_type = FT
 )
 
 TestModel(::CPU, FT, ν=1.0, Δx=0.5) =
-    Model(grid = RegularCartesianGrid(FT, size=(3, 3, 3), length=(3Δx, 3Δx, 3Δx), topology=DT),
+    Model(grid = RegularCartesianGrid(FT; size=(3, 3, 3), length=(3Δx, 3Δx, 3Δx)),
        closure = ConstantIsotropicDiffusivity(FT; ν=ν, κ=ν),
   architecture = CPU(),
     float_type = FT
