+++ test/regression_tests/thermal_bubble_regression_test.jl
@@ -3,8 +3,7 @@ function run_thermal_bubble_regression_test(arch)
     Lx, Ly, Lz = 100, 100, 100
     Δt = 6
 
-    grid = RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz),
-                                topology=(Periodic, Periodic, Bounded))
+    grid = RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz))
     closure = ConstantIsotropicDiffusivity(ν=4e-2, κ=4e-2)
     model = Model(architecture=arch, grid=grid, closure=closure, coriolis=FPlane(f=1e-4))
 

[…]
