+++ test/test_time_stepping.jl
@@ -1,6 +1,6 @@
 function time_stepping_works_with_closure(arch, FT, Closure)
     # Use halos of size 2 to accomadate time stepping with AnisotropicBiharmonicDiffusivity.
-    grid = RegularCartesianGrid(FT, size=(16, 16, 16), halo=(2, 2, 2), length=(1, 2, 3), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(16, 16, 16), halo=(2, 2, 2), length=(1, 2, 3))
 
     model = Model(grid=grid, architecture=arch, float_type=FT, closure=Closure(FT))
     time_step!(model, 1, 1)
@@ -9,7 +9,7 @@ function time_stepping_works_with_closure(arch, FT, Closure)
 end
 
 function time_stepping_works_with_nonlinear_eos(arch, FT, eos_type)
-    grid = RegularCartesianGrid(FT, size=(16, 16, 16), length=(1, 2, 3), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(16, 16, 16), length=(1, 2, 3))
 
     eos = RoquetIdealizedNonlinearEquationOfState(eos_type)
     b = SeawaterBuoyancy(equation_of_state=eos)
@@ -22,7 +22,7 @@ end
 
 function run_first_AB2_time_step_tests(arch, FT)
     add_ones(args...) = 1.0
-    model = Model(grid=RegularCartesianGrid(FT, size=(16, 16, 16), length=(1, 2, 3), topology=DT),
+    model = Model(grid=RegularCartesianGrid(FT; size=(16, 16, 16), length=(1, 2, 3)),
                   architecture=arch, float_type=FT, forcing=ModelForcing(T=add_ones))
     time_step!(model, 1, 1)
 
@@ -44,7 +44,7 @@ function compute_w_from_continuity(arch, FT)
     Nx, Ny, Nz = 16, 16, 16
     Lx, Ly, Lz = 16, 16, 16
 
-    grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(Lx, Ly, Lz), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(Nx, Ny, Nz), length=(Lx, Ly, Lz))
     bcs = HorizontallyPeriodicSolutionBCs()
 
     u = FaceFieldX(FT, arch, grid)
@@ -86,8 +86,7 @@ function incompressible_in_time(arch, FT, Nt)
     Nx, Ny, Nz = 32, 32, 32
     Lx, Ly, Lz = 10, 10, 10
 
-    grid = RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz), topology=DT)
-    model = Model(grid=grid, architecture=arch, float_type=FT)
+    model = Model(grid=RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz)), architecture=arch, float_type=FT)
 
     grid = model.grid
     u, v, w = model.velocities.u, model.velocities.v, model.velocities.w
@@ -131,11 +130,9 @@ function tracer_conserved_in_channel(arch, FT, Nt)
     νh, κh = 20.0, 20.0
     νv, κv = α*νh, α*κh
 
-    model = ChannelModel(
-        architecture = arch, float_type = FT,
-        grid = RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz), topology=DT),
-        closure = ConstantAnisotropicDiffusivity(νh=νh, νv=νv, κh=κh, κv=κv)
-    )
+    model = ChannelModel(architecture = arch, float_type = FT,
[…]
