### 2020-02-04 4eca90d
Revert "Update grid tests"
This reverts commit 8fef1ccb955dd616992cfcb7c9c7235cc106188f.
