+++ test/test_grids.jl
@@ -1,8 +1,5 @@
-# A Default Topology to use for testing.
-const DT = (Periodic, Periodic, Bounded)
-
 function correct_grid_size_regular(FT)
-    grid = RegularCartesianGrid(FT, size=(4, 6, 8), length=(2π, 4π, 9π), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(4, 6, 8), length=(2π, 4π, 9π))
 
     # Checking ≈ as the grid could be storing Float32 values.
     return (grid.Nx ≈ 4  && grid.Ny ≈ 6  && grid.Nz ≈ 8 &&
@@ -10,17 +7,17 @@ function correct_grid_size_regular(FT)
 end
 
 function correct_halo_size_regular(FT)
-    grid = RegularCartesianGrid(FT, size=(4, 6, 8), length=(2π, 4π, 9π), halo=(1, 2, 3), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(4, 6, 8), length=(2π, 4π, 9π), halo=(1, 2, 3))
     return (grid.Hx == 1  && grid.Hy == 2  && grid.Hz == 3)
 end
 
 function faces_start_at_zero_regular(FT)
-    grid = RegularCartesianGrid(FT, size=(10, 10, 10), length=(2π, 2π, 2π), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(10, 10, 10), length=(2π, 2π, 2π))
     return grid.xF[1] == 0 && grid.yF[1] == 0 && grid.zF[end] == 0
 end
 
 function end_faces_match_grid_length_regular(FT)
-    grid = RegularCartesianGrid(FT, size=(12, 13, 14), length=(π, π^2, π^3), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(12, 13, 14), length=(π, π^2, π^3))
     return (grid.xF[end] - grid.xF[1] ≈ π   &&
             grid.yF[end] - grid.yF[1] ≈ π^2 &&
             grid.zF[end] - grid.zF[1] ≈ π^3)
@@ -28,7 +25,7 @@ end
 
 function ranges_have_correct_length_regular(FT)
     Nx, Ny, Nz = 8, 9, 10
-    grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(1, 1, 1), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(Nx, Ny, Nz), length=(1, 1, 1))
     return (length(grid.xC) == Nx && length(grid.xF) == Nx+1 &&
             length(grid.yC) == Ny && length(grid.yF) == Ny+1 &&
             length(grid.zC) == Nz && length(grid.zF) == Nz+1)
@@ -37,26 +34,24 @@ end
 # See: https://github.com/climate-machine/Oceananigans.jl/issues/480
 function no_roundoff_error_in_ranges_regular(FT)
     Nx, Ny, Nz = 1, 1, 64
-    grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(1, 1, π/2), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(Nx, Ny, Nz), length=(1, 1, π/2))
     return length(grid.zC) == Nz && length(grid.zF) == Nz+1
 end
 
 function grid_properties_are_same_type_regular(FT)
-    grid = RegularCartesianGrid(FT, size=(10, 10, 10), length=(1, 1//7, 2π), topology=DT)
+    grid = RegularCartesianGrid(FT; size=(10, 10, 10), length=(1, 1//7, 2π))
[…]
