### 2020-02-04 81bad3f
Make horizontally periodic the default topology
