+++ src/Grids/regular_cartesian_grid.jl
@@ -34,7 +34,7 @@ struct RegularCartesianGrid{FT, TX, TY, TZ, R} <: AbstractGrid{FT, TX, TY, TZ}
 end
 
 """
-    RegularCartesianGrid([FT=Float64]; size, length, x, y, z)
+    RegularCartesianGrid([FT=Float64]; size, length, topology, x, y, z)
 
 Creates a `RegularCartesianGrid` with `size = (Nx, Ny, Nz)` grid points.
 
@@ -42,6 +42,10 @@ The physical length of the domain can be specified via `x`, `y`, and `z` keyword
 indicating the left and right endpoints of each dimensions, e.g. `x=(-π, π)` or via
 the `length` argument, e.g. `length=(Lx, Ly, Lz)` which specifies the length of each dimension
 in which case 0 ≤ x ≤ Lx, 0 ≤ y ≤ Ly, and -Lz ≤ z ≤ 0.
+
+A grid topology may be specified via a tuple assigning one of `Periodic`, `Bounded, and `Singleton`
+to each dimension. By default, a horizontally periodic grid topology `(Periodic, Periodic, Singleton)`
+is assumed.
 
 Constants are stored using floating point values of type `FT`. By default this is `Float64`.
 Make sure to specify the desired `FT` if not using `Float64`.
@@ -72,7 +76,7 @@ domain: x ∈ [0.0, 8.0], y ∈ [-10.0, 10.0], z ∈ [3.141592653589793, -3.1415
 grid spacing (Δx, Δy, Δz) = (0.25f0, 0.625f0, 0.3926991f0)
 ```
 """
-function RegularCartesianGrid(FT=Float64; size, halo=(1, 1, 1), topology,
+function RegularCartesianGrid(FT=Float64; size, halo=(1, 1, 1), topology=(Periodic, Periodic, Bounded),
                               length=nothing, x=nothing, y=nothing, z=nothing)
 
     # Hack that allows us to use `size` and `length` as keyword arguments but then also
@@ -102,19 +106,19 @@ function RegularCartesianGrid(FT=Float64; size, halo=(1, 1, 1), topology,
     yF = range(y₁, y₂; length=Ny+1)
     zF = range(z₁, z₂; length=Nz+1)
 
-    RegularCartesianGrid{FT, typeof(TX), typeof(TY), typeof(TZ) typeof(xC)}(
+    RegularCartesianGrid{FT, typeof(TX), typeof(TY), typeof(TZ), typeof(xC)}(
         Nx, Ny, Nz, Hx, Hy, Hz, Lx, Ly, Lz, Δx, Δy, Δz, xC, yC, zC, xF, yF, zF)
 end
 
-short_show(grid::RegularCartesianGrid{T}) where T =
-    "RegularCartesianGrid{$T}(Nx=$(grid.Nx), Ny=$(grid.Ny), Nz=$(grid.Nz))"
+short_show(grid::RegularCartesianGrid{FT, TX, TY, TZ}) where {FT, TX, TY, TZ} =
+    "RegularCartesianGrid{$FT, $TX, $TY, $TZ}(Nx=$(grid.Nx), Ny=$(grid.Ny), Nz=$(grid.Nz))"
 
 show_domain(grid) = string("x ∈ [", grid.xF[1], ", ", grid.xF[end], "], ",
                            "y ∈ [", grid.yF[1], ", ", grid.yF[end], "], ",
                            "z ∈ [", grid.zF[1], ", ", grid.zF[end], "]")
 
-show(io::IO, g::RegularCartesianGrid) =
[…]
