### 2020-02-04 8c9c880
Change `Singleton` topology to `Flat`
