### 2020-02-04 6805332
Enable all tests
