### 2020-02-04 fa18aea
Fixes typos in coriolis_forces doc
