### 2020-02-06 65370bf
Add test cases with entropy as prognostic variable
