### 2020-02-06 c25fcf0
Remove Makie from dependencies
