### 2020-02-06 41de96e
Refactoring to support multiple species
