### 2020-02-07 483d296
Fixing minor bugs
