+++ src/microphysics.jl
@@ -1,3 +1,1 @@
-import JULES.Operators: ρᵐ
-
 abstract type AbstractMicrophysics end
