### 2020-02-07 b2d4a46
Possible type stability?
