+++ src/Operators/compressible_operators.jl
@@ -36,13 +36,12 @@ end
 ##### Diffusion
 #####
 
-@inline diffusive_flux_x(i, j, k, grid, κᶠᶜᶜ, ρ, C) = ρ[i,j,k] * κᶠᶜᶜ * Axᵃᵃᶠ(i, j, k, grid) * ∂xᶠᵃᵃ(i, j, k, grid, C_over_ρ, C, ρ)
-@inline diffusive_flux_y(i, j, k, grid, κᶜᶠᶜ, ρ, C) = ρ[i,j,k] * κᶜᶠᶜ * Ayᵃᵃᶠ(i, j, k, grid) * ∂yᵃᶠᵃ(i, j, k, grid, C_over_ρ, C, ρ)
-@inline diffusive_flux_z(i, j, k, grid, κᶜᶜᶠ, ρ, C) = ρ[i,j,k] * κᶜᶜᶠ * Azᵃᵃᵃ(i, j, k, grid) * ∂zᵃᵃᶠ(i, j, k, grid, C_over_ρ, C, ρ)
+@inline diffusive_flux_x(i, j, k, grid, κᶠᶜᶜ, ρ, C) = @inbounds ρ[i,j,k] * κᶠᶜᶜ * Axᵃᵃᶠ(i, j, k, grid) * ∂xᶠᵃᵃ(i, j, k, grid, C_over_ρ, C, ρ)
+@inline diffusive_flux_y(i, j, k, grid, κᶜᶠᶜ, ρ, C) = @inbounds ρ[i,j,k] * κᶜᶠᶜ * Ayᵃᵃᶠ(i, j, k, grid) * ∂yᵃᶠᵃ(i, j, k, grid, C_over_ρ, C, ρ)
+@inline diffusive_flux_z(i, j, k, grid, κᶜᶜᶠ, ρ, C) = @inbounds ρ[i,j,k] * κᶜᶜᶠ * Azᵃᵃᵃ(i, j, k, grid) * ∂zᵃᵃᶠ(i, j, k, grid, C_over_ρ, C, ρ)
 
 @inline function ∂ⱼDᶜⱼ(i, j, k, grid, closure::ConstantIsotropicDiffusivity,
-                        ρ, C, ::Val{tracer_index}, args...) where tracer_index
-
+                       ρ, C, ::Val{tracer_index}, args...) where tracer_index
     @inbounds κ = closure.κ[tracer_index]
     return 1/Vᵃᵃᶜ(i, j, k, grid) * (δxᶜᵃᵃ(i, j, k, grid, diffusive_flux_x, κ, ρ, C) +
                                     δyᵃᶜᵃ(i, j, k, grid, diffusive_flux_y, κ, ρ, C) +
@@ -92,27 +91,27 @@ end
 ##### Viscous dissipation
 #####
 
-@inline strain_rate_tensor_ux(i, j, k, grid, ρ, U) = 1.0/3.0 * ∂xᶜᵃᵃ(i, j, k, grid, U_over_ρ, U, ρ)
+@inline strain_rate_tensor_ux(i, j, k, grid, ρ, U)    = 1.0/3.0 * ∂xᶜᵃᵃ(i, j, k, grid, U_over_ρ, U, ρ)
 @inline strain_rate_tensor_uy(i, j, k, grid, ρ, U, V) = ∂yᵃᶠᵃ(i, j, k, grid, U_over_ρ, U, ρ) + ∂xᶠᵃᵃ(i, j, k, grid, V_over_ρ, V, ρ)
[…]
