+++ src/right_hand_sides.jl
@@ -4,8 +4,7 @@
 
 @inline function FU(i, j, k, grid, coriolis, closure, ρ, Ũ, K)
     @inbounds begin
-        return (- div_ρuũ(i, j, k, grid, ρ, Ũ)
-                - x_f_cross_U(i, j, k, grid, coriolis, Ũ)
+        return (- x_f_cross_U(i, j, k, grid, coriolis, Ũ)
                 + ∂ⱼτ₁ⱼ(i, j, k, grid, closure, ρ, Ũ, K))
     end
 end
@@ -13,21 +12,19 @@ end
 
 @inline function FV(i, j, k, grid, coriolis, closure, ρ, Ũ, K)
     @inbounds begin
-        return (- div_ρvũ(i, j, k, grid, ρ, Ũ)
-                - y_f_cross_U(i, j, k, grid, coriolis, Ũ)
+        return (- y_f_cross_U(i, j, k, grid, coriolis, Ũ)
                 + ∂ⱼτ₂ⱼ(i, j, k, grid, closure, ρ, Ũ, K))
     end
 end
 
 @inline function FW(i, j, k, grid, coriolis, closure, ρ, Ũ, K)
     @inbounds begin
-        return (- div_ρwũ(i, j, k, grid, ρ, Ũ)
-                - z_f_cross_U(i, j, k, grid, coriolis, Ũ)
+        return (- z_f_cross_U(i, j, k, grid, coriolis, Ũ)
                 + ∂ⱼτ₃ⱼ(i, j, k, grid, closure, ρ, Ũ, K))
     end
 end
 
-@inline FC(i, j, k, grid, closure, ρ, C, tracer_index, K̃) =
+@inline FC(i, j, k, grid, closure, ρ, C, ::Val{tracer_index}, K̃) where tracer_index =
     @inbounds ∂ⱼDᶜⱼ(i, j, k, grid, closure, ρ, C, Val(tracer_index), K̃)
 
 @inline function FS(i, j, k, grid, closure, tvar::Entropy, ρ, ρ̃, Ũ, C̃, K̃)
@@ -48,21 +45,24 @@ end
 
 @inline function RU(i, j, k, grid, tvar, ρ, ρ̃, Ũ, C, FU)
     @inbounds begin
-        return (- ∂p∂x(i, j, k, grid, tvar, ρ̃, C)
+        return (- div_ρuũ(i, j, k, grid, ρ, Ũ)
+                - ∂p∂x(i, j, k, grid, tvar, ρ̃, C)
                 + FU[i, j, k])
     end
 end
 
 @inline function RV(i, j, k, grid, tvar, ρ, ρ̃, Ũ, C, FV)
     @inbounds begin
-        return (- ∂p∂y(i, j, k, grid, tvar, ρ̃, C)
+        return (- div_ρvũ(i, j, k, grid, ρ, Ũ)
+                - ∂p∂y(i, j, k, grid, tvar, ρ̃, C)
                 + FV[i, j, k])
     end
 end
 
 @inline function RW(i, j, k, grid, tvar, gravity, ρ, ρ̃, Ũ, C, FW)
     @inbounds begin
-        return (- ∂p∂z(i, j, k, grid, tvar, ρ̃, C)
+        return (- div_ρwũ(i, j, k, grid, ρ, Ũ)
+                - ∂p∂z(i, j, k, grid, tvar, ρ̃, C)
                 - gravity*ρ[i,j,k]
                 + FW[i, j, k])
     end
