### 2020-02-07 9698a1a
Fixes file misnaming in Diagnostics.jl
