+++ src/Diagnostics/Diagnostics.jl
@@ -12,7 +12,7 @@ using Oceananigans: AbstractDiagnostic
 include("diagnostics_kernels.jl")
 include("nan_checker.jl")
 include("horizontal_average.jl")
-include("timeseries.jl")
+include("time_series.jl")
 include("field_maximum.jl")
 include("cfl.jl")
 
