### 2020-02-07 3e7f0cf
Change iteration in pressure gradient calculations to avoid allocation
