### 2020-02-07 948277c
Changes to improve type inference for tracers
