### 2020-02-07 cf81e6c
Stop dispatching on value of tracer indices
