### 2020-02-08 5dedbea
Three-fluid test working but only without scalar diffusion
