+++ verification/three_gas_dry_rising_thermal_bubble/three_gas_dry_rising_thermal_bubble.jl
@@ -30,28 +30,43 @@ grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
 
 model = CompressibleModel(
                       grid = grid,
-                 densities = DryEarth(),
+                 densities = DryEarth3(),
     thermodynamic_variable = PrognosticS(),
-                   closure = ConstantIsotropicDiffusivity(ν=150.0, κ=150.0)
+                   closure = ConstantIsotropicDiffusivity(ν=0.0, κ=0.0)
 )
 
 #####
 ##### Dry thermal bubble perturbation
 #####
 
-gas = model.densities.ρ
+gas = model.densities.ρ₁
 R, cₚ, cᵥ = gas.R, gas.cₚ, gas.cᵥ
 sref, Tref, ρref = gas.s₀, gas.T₀, gas.ρ₀
 g  = model.gravity
 pₛ = 1000hPa
 Tₛ = 300
 
+# Define initial mixing ratios
+q₁(z) = exp(-(4z/Lz)^2)
+q₂(z) = exp(-(4*(z - Lz)/Lz)^2)
+q₃(z) = 1 - q₁(z) - q₂(z)
+
 # Define an approximately hydrostatic background state
 θ₀(x, y, z) = Tₛ
 p₀(x, y, z) = pₛ * (1 - g*z/(cₚ*Tₛ))^(cₚ/R)
 T₀(x, y, z) = Tₛ*(p₀(x, y, z)/pₛ)^(R/cₚ)
 ρ₀(x, y, z) = p₀(x, y, z)/(R*T₀(x, y, z))
-ρs₀(x, y, z) = ρ₀(x, y, z) * (sref + cᵥ*log(T₀(x, y, z)/Tref) - R*log(ρ₀(x, y, z)/ρref))
+ρ₁₀(x, y, z) = q₁(z) * ρ₀(x, y, z)
+ρ₂₀(x, y, z) = q₂(z) * ρ₀(x, y, z)
+ρ₃₀(x, y, z) = q₃(z) * ρ₀(x, y, z)
+function ρs₀(x, y, z)
+    ρs = 0.0
+    T = T₀(x, y, z)
+    for ρ in (ρ₁₀(x, y, z), ρ₂₀(x, y, z), ρ₃₀(x, y, z))
+        ρs += (ρ > 0 ?  ρ * (sref + cᵥ*log(T/Tref) - R*log(ρ/ρref)) : 0.0)
+    end
+    return ρs
+end
 
 # Define the initial density perturbation
 xᶜ, zᶜ = 0km, 2km
@@ -67,17 +82,31 @@ end
 ρᵢ(x, y, z) = ρ₀(x, y, z) + ρ′(x, y, z)
 pᵢ(x, y, z) = p₀(x, y, z)
 Tᵢ(x, y, z) = pᵢ(x, y, z) / (R * ρᵢ(x, y, z))
-ρsᵢ(x, y, z) = ρᵢ(x, y, z) * (sref + cᵥ*log(Tᵢ(x, y, z)/Tref) - R*log(ρᵢ(x, y, z)/ρref))
+ρ₁ᵢ(x, y, z) = q₁(z) * ρᵢ(x, y, z)
+ρ₂ᵢ(x, y, z) = q₂(z) * ρᵢ(x, y, z)
+ρ₃ᵢ(x, y, z) = q₃(z) * ρᵢ(x, y, z)
+function ρsᵢ(x, y, z)
+    ρs = 0.0
+    T = Tᵢ(x, y, z)
+    for ρ in (ρ₁ᵢ(x, y, z), ρ₂ᵢ(x, y, z), ρ₃ᵢ(x, y, z))
[…]
