+++ src/Operators/Operators.jl
@@ -5,7 +5,7 @@ using Oceananigans
 export
     x_f_cross_U, y_f_cross_U, z_f_cross_U,
     hdivᶜᶜᵃ, divᶜᶜᶜ, ∇²,
-    div_flux, ∂ⱼDᶜⱼ, sᶜ∂ⱼDᶜⱼ,
+    div_flux, ∂ⱼDᶜⱼ, ∂ⱼsᶜDᶜⱼ,
     div_ρuũ, div_ρvũ, div_ρwũ,
     ∂ⱼτ₁ⱼ, ∂ⱼτ₂ⱼ, ∂ⱼτ₃ⱼ, Q_dissipation
 
