### 2020-02-08 9a8075a
Add missing density interpolation
