+++ verification/three_gas_dry_rising_thermal_bubble/three_gas_dry_rising_thermal_bubble.jl
@@ -12,6 +12,7 @@ using VideoIO
 using FileIO
 using Oceananigans
 using JULES
+using Oceananigans.Operators: ℑzᵃᵃᶠ
 
 const km = 1000
 const hPa = 100
@@ -32,7 +33,7 @@ model = CompressibleModel(
                       grid = grid,
                  densities = DryEarth3(),
     thermodynamic_variable = PrognosticS(),
-                   closure = ConstantIsotropicDiffusivity(ν=0.0, κ=0.0)
+                   closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
 )
 
 #####
@@ -103,6 +104,9 @@ set!(model.tracers.ρ₃, ρ₃₀)
 set!(model.tracers.ρs, ρs₀)
 update_total_density!(model.total_density, model.grid, model.densities, model.tracers)
 ρʰᵈ = ρ.data[1:Nx, 1, 1:Nz]
+ρ₁ʰᵈ = ρ₁.data[1:Nx, 1, 1:Nz]
+ρ₂ʰᵈ = ρ₂.data[1:Nx, 1, 1:Nz]
+ρ₃ʰᵈ = ρ₃.data[1:Nx, 1, 1:Nz]
 ρsʰᵈ = ρs.data[1:Nx, 1, 1:Nz]
 set!(model.tracers.ρ₁, ρ₁ᵢ)
 set!(model.tracers.ρ₂, ρ₂ᵢ)
@@ -131,9 +135,20 @@ for n in 1:200
     time_step!(model, Δt = Δt, Nt = 50)
 
     CFL = cfl(model, Δt)
-    ρ̄ = sum(ρ.data[1:Nx,1,1:Nz])/(Nx*Nz)
+    ρ̄ = sum(ρ.data[1:Nx,1:Ny,1:Nz])/(Nx*Ny*Nz)
+    ∂tρ₁ = maximum(model.slow_forcings.tracers.ρ₁[1:Nx,1:Ny,1:Nz])
+    ∂tρ₂ = maximum(model.slow_forcings.tracers.ρ₂[1:Nx,1:Ny,1:Nz])
+    ∂tρ₃ = maximum(model.slow_forcings.tracers.ρ₃[1:Nx,1:Ny,1:Nz])
+    ∂tρ = maximum(model.slow_forcings.tracers.ρ₁[1:Nx,1:Ny,1:Nz] .+
+              model.slow_forcings.tracers.ρ₂[1:Nx,1:Ny,1:Nz] .+
+              model.slow_forcings.tracers.ρ₃[1:Nx,1:Ny,1:Nz])
     @printf("t = %.2f s, CFL = %.2e, ρ̄ = %.2e (rel.err. = %.2e)\n",
         model.clock.time, CFL, ρ̄, (ρ̄ - ρ̄ᵢ)/ρ̄)
+    @printf("Maximum mass tendencies from diffusion:\n")
+    @printf("ρ₁: %.2e\n", ∂tρ₁)
+    @printf("ρ₂: %.2e\n", ∂tρ₂)
+    @printf("ρ₃: %.2e\n", ∂tρ₃)
+    @printf("ρ : %.2e\n", ∂tρ)
 
     xC, yC, zC = model.grid.xC ./ km, model.grid.yC ./ km, model.grid.zC ./ km
     xF, yF, zF = model.grid.xF ./ km, model.grid.yF ./ km, model.grid.zF ./ km
@@ -149,17 +164,17 @@ for n in 1:200
 
     ρ₁_title = @sprintf("rho1, t = %d s", round(Int, model.clock.time))
     pρ₁ = heatmap(xC, zC, ρ₁_slice, title=ρ₁_title, fill=true, levels=50,
[…]
