### 2020-02-09 5ed7ee6
Add energy as a thermodynamic variable
