+++ src/Operators/compressible_operators.jl
@@ -11,6 +11,24 @@ using Oceananigans: AbstractGrid
 @inline C_over_ρ(i, j, k, grid, C, ρ) = @inbounds C[i, j, k] / ρ[i, j, k]
 
 #####
+##### Kinetic energy
+#####
+@inline kinetic_energy(i, j, k, grid, Ũ, ρ) = @inbounds 0.5*((ℑxᶜᵃᵃ(i, j, k, grid, Ũ.ρu) / ρ[i, j, k])^2 + (ℑyᵃᶜᵃ(i, j, k, grid, Ũ.ρv) / ρ[i, j, k])^2 + (ℑzᵃᵃᶜ(i, j, k, grid, Ũ.ρw) / ρ[i, j, k])^2)
+
+#####
+##### Pressure work
+#####
+@inline pu(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) = ℑxᶠᵃᵃ(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) * Axᵃᵃᶠ(i, j, k, grid) * U_over_ρ(i, j, k, grid, Ũ.ρu, ρ)
+@inline pv(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) = ℑyᵃᶠᵃ(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) * Ayᵃᵃᶠ(i, j, k, grid) * V_over_ρ(i, j, k, grid, Ũ.ρv, ρ)
+@inline pw(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) = ℑzᵃᵃᶠ(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) * Azᵃᵃᵃ(i, j, k, grid) * W_over_ρ(i, j, k, grid, Ũ.ρw, ρ)
+
+@inline function ∂ⱼpuⱼ(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃)
+    return 1/Vᵃᵃᶜ(i, j, k, grid) * (δxᶜᵃᵃ(i, j, k, grid, pu, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) +
+                                    δyᵃᶜᵃ(i, j, k, grid, pv, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) +
+                                    δzᵃᵃᶜ(i, j, k, grid, pw, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃))
+end
+
+#####
 ##### Coriolis terms
 #####
 
@@ -39,10 +57,12 @@ end
 @inline diffusive_flux_x(i, j, k, grid, κᶠᶜᶜ, ρ, C) = ℑxᶠᵃᵃ(i, j, k, grid, ρ) * κᶠᶜᶜ * Axᵃᵃᶠ(i, j, k, grid) * ∂xᶠᵃᵃ(i, j, k, grid, C_over_ρ, C, ρ)
 @inline diffusive_flux_y(i, j, k, grid, κᶜᶠᶜ, ρ, C) = ℑyᵃᶠᵃ(i, j, k, grid, ρ) * κᶜᶠᶜ * Ayᵃᵃᶠ(i, j, k, grid) * ∂yᵃᶠᵃ(i, j, k, grid, C_over_ρ, C, ρ)
[…]
