### 2020-02-09 0d75c80
Working rising thermal bubble with energy as thermodynamic variable
