+++ src/Operators/compressible_operators.jl
@@ -52,6 +52,8 @@ end
 
 #####
 ##### Diffusion
+##### Note: need a better way to enforce boundary conditions on diffusive fluxes
+##### at rigid boundaries
 #####
 
 @inline diffusive_flux_x(i, j, k, grid, κᶠᶜᶜ, ρ, C) = ℑxᶠᵃᵃ(i, j, k, grid, ρ) * κᶠᶜᶜ * Axᵃᵃᶠ(i, j, k, grid) * ∂xᶠᵃᵃ(i, j, k, grid, C_over_ρ, C, ρ)
@@ -60,9 +62,12 @@ end
 @inline diffusive_tvar_flux_x(i, j, k, grid, κᶠᶜᶜ, tvar_diag, tvar, tracer_index, g, Ũ, ρ̃, C̃, ρ, C) = ℑxᶠᵃᵃ(i, j, k, grid, tvar_diag, tvar, tracer_index, g, Ũ, ρ, ρ̃, C̃) * κᶠᶜᶜ * Axᵃᵃᶠ(i, j, k, grid) * ∂xᶠᵃᵃ(i, j, k, grid, C_over_ρ, C, ρ)
 @inline diffusive_tvar_flux_y(i, j, k, grid, κᶜᶠᶜ, tvar_diag, tvar, tracer_index, g, Ũ, ρ̃, C̃, ρ, C) = ℑyᵃᶠᵃ(i, j, k, grid, tvar_diag, tvar, tracer_index, g, Ũ, ρ, ρ̃, C̃) * κᶜᶠᶜ * Ayᵃᵃᶠ(i, j, k, grid) * ∂yᵃᶠᵃ(i, j, k, grid, C_over_ρ, C, ρ)
 @inline diffusive_tvar_flux_z(i, j, k, grid, κᶜᶜᶠ, tvar_diag, tvar, tracer_index, g, Ũ, ρ̃, C̃, ρ, C) = ℑzᵃᵃᶠ(i, j, k, grid, tvar_diag, tvar, tracer_index, g, Ũ, ρ, ρ̃, C̃) * κᶜᶜᶠ * Azᵃᵃᵃ(i, j, k, grid) * ∂zᵃᵃᶠ(i, j, k, grid, C_over_ρ, C, ρ)
-@inline diffusive_pressure_flux_x(i, j, k, grid, κᶠᶜᶜ, p_diag, tvar, g, Ũ, ρ, ρ̃, C̃) = κᶠᶜᶜ * Axᵃᵃᶠ(i, j, k, grid) * ∂xᶠᵃᵃ(i, j, k, grid, p_diag, tvar, g, Ũ, ρ, ρ̃, C̃)
-@inline diffusive_pressure_flux_y(i, j, k, grid, κᶜᶠᶜ, p_diag, tvar, g, Ũ, ρ, ρ̃, C̃) = κᶜᶠᶜ * Ayᵃᵃᶠ(i, j, k, grid) * ∂yᵃᶠᵃ(i, j, k, grid, p_diag, tvar, g, Ũ, ρ, ρ̃, C̃)
-@inline diffusive_pressure_flux_z(i, j, k, grid, κᶜᶜᶠ, p_diag, tvar, g, Ũ, ρ, ρ̃, C̃) = κᶜᶜᶠ * Azᵃᵃᵃ(i, j, k, grid) * ∂zᵃᵃᶠ(i, j, k, grid, p_diag, tvar, g, Ũ, ρ, ρ̃, C̃)
[…]
