### 2020-02-09 331781f
Working three-component thermal bubble with energy
