+++ verification/three_gas_dry_rising_thermal_bubble_energy/three_gas_dry_rising_thermal_bubble_energy.jl
@@ -1,0 +1,193 @@
+"""
+This example sets up a dry, warm thermal bubble perturbation in a uniform
+lateral mean flow which buoyantly rises. Identical to the verification
+experiment in ../three_gas_dry_rising_thermal_bubble_entropy except that
+energy is used as the prognostic thermodynamic variable rather than entropy.
+"""
+
+using Printf
+using Profile
+using Plots
+using VideoIO
+using FileIO
+using Oceananigans
+using JULES
+using Oceananigans.Operators: ℑzᵃᵃᶠ
+
+const km = 1000
+const hPa = 100
+
+Lx = 20km
+Lz = 10km
+
+Δ = 0.2km  # grid spacing [m]
+
+Nx = Int(Lx/Δ)
+Ny = 1
+Nz = Int(Lz/Δ)
+
+grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
+                            x=(-Lx/2, Lx/2), y=(-Lx/2, Lx/2), z=(0, Lz))
+
+model = CompressibleModel(
+                      grid = grid,
+                 densities = DryEarth3(),
+    thermodynamic_variable = PrognosticE(),
+                   closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
+)
+
+#####
+##### Dry thermal bubble perturbation
+#####
+
+gas = model.densities.ρ₁
+R, cₚ, cᵥ = gas.R, gas.cₚ, gas.cᵥ
+uref, Tref, ρref = gas.u₀, gas.T₀, gas.ρ₀
+g  = model.gravity
+pₛ = 1000hPa
+Tₛ = 300
+
+# Define initial mixing ratios
+q₁(z) = exp(-(4z/Lz)^2)
+q₃(z) = exp(-(4*(z - Lz)/Lz)^2)
+q₂(z) = 1 - q₁(z) - q₃(z)
+
+# Define an approximately hydrostatic background state
+θ₀(x, y, z) = Tₛ
+p₀(x, y, z) = pₛ * (1 - g*z/(cₚ*Tₛ))^(cₚ/R)
+T₀(x, y, z) = Tₛ*(p₀(x, y, z)/pₛ)^(R/cₚ)
+ρ₀(x, y, z) = p₀(x, y, z)/(R*T₀(x, y, z))
+ρ₁₀(x, y, z) = q₁(z) * ρ₀(x, y, z)
+ρ₂₀(x, y, z) = q₂(z) * ρ₀(x, y, z)
+ρ₃₀(x, y, z) = q₃(z) * ρ₀(x, y, z)
+function ρe₀(x, y, z)
+    ρe = 0.0
+    T = T₀(x, y, z)
+    for ρ in (ρ₁₀(x, y, z), ρ₂₀(x, y, z), ρ₃₀(x, y, z))
+        ρe += ρ * (uref + cᵥ*(T - Tref) + g*z)
+    end
+    return ρe
+end
+
+# Define the initial density perturbation
+xᶜ, zᶜ = 0km, 2km
+xʳ, zʳ = 2km, 2km
+L(x, y, z) = sqrt(((x - xᶜ)/xʳ)^2 + ((z - zᶜ)/zʳ)^2)
+function ρ′(x, y, z; θᶜ′ = 2.0)
+    l = L(x, y, z)
+    θ′ = (l <= 1) * θᶜ′ * cos(π/2 * L(x, y, z))^2
+    return -ρ₀(x, y, z) * θ′ / θ₀(x, y, z)
+end
+
[…]
