### 2020-02-09 6327e2f
Change color limits to match three-gas test case
