### 2020-02-10 3c2b23c
Document RegularCartesianGrid defaults in docstring
