### 2020-02-11 60f5280
Create field boundary conditions using information from grid topology
