### 2020-02-11 d22bb3a
Drafting a `Simulation` type and a `run!` simulation function
