### 2020-02-11 0194e95
Improves docstring and refactors so that AMD constants may be set for each tracer individually
