### 2020-02-11 3dbddb0
Adds a show function for AMD viscosity
