### 2020-02-11 61985b3
Bugfixes for show method for Verstappen AMD turbulence closure
