+++ src/TurbulenceClosures/turbulence_closure_implementations/verstappen_anisotropic_minimum_dissipation.jl
@@ -18,8 +18,8 @@ end
 
 const VAMD = VerstappenAnisotropicMinimumDissipation
 
-function Base.show(io::IO, closure::VAMD{FT}) where FT =
-    print(io, "VerstappenAnisotropicMinimumDissipation{$FT} turbulence closure with:\n"
+Base.show(io::IO, closure::VAMD{FT}) where FT =
+    print(io, "VerstappenAnisotropicMinimumDissipation{$FT} turbulence closure with:\n",
               "           Poincaré constant for momentum eddy viscosity Cν: ", closure.Cν, '\n',
               "    Poincaré constant for tracer(s) eddy diffusivit(ies) Cκ: ", closure.Cκ, '\n',
               "                        Buoyancy modification multiplier Cb: ", closure.Cb, '\n',
