### 2020-02-11 d25abdb
Adds more to the docstring for verstappen AMD turbulence closure
