+++ src/TurbulenceClosures/turbulence_closure_implementations/verstappen_anisotropic_minimum_dissipation.jl
@@ -58,15 +58,33 @@ Example
 =======
 
 julia> pretty_diffusive_closure = AnisotropicMinimumDissipation(C=1/2)
+VerstappenAnisotropicMinimumDissipation{Float64} turbulence closure with:
+           Poincaré constant for momentum eddy viscosity Cν: 0.5
+    Poincaré constant for tracer(s) eddy diffusivit(ies) Cκ: 0.5
+                        Buoyancy modification multiplier Cb: 0.0
+                Background diffusivit(ies) for tracer(s), κ: 1.46e-7
+             Background kinematic viscosity for momentum, ν: 1.05e-6
 
 julia> const Δz = 0.5; # grid resolution at surface
 
 julia> surface_enhanced_tracer_C(x, y, z) = 1/12 * (1 + exp((z + Δz/2) / 8Δz))
+surface_enhanced_tracer_C (generic function with 1 method)
 
 julia> fancy_closure = AnisotropicMinimumDissipation(Cκ=surface_enhanced_tracer_C)
+VerstappenAnisotropicMinimumDissipation{Float64} turbulence closure with:
+           Poincaré constant for momentum eddy viscosity Cν: 0.08333333333333333
+    Poincaré constant for tracer(s) eddy diffusivit(ies) Cκ: surface_enhanced_tracer_C
+                        Buoyancy modification multiplier Cb: 0.0
+                Background diffusivit(ies) for tracer(s), κ: 1.46e-7
+             Background kinematic viscosity for momentum, ν: 1.05e-6
 
 julia> tracer_specific_closure = AnisotropicMinimumDissipation(Cκ=(c₁=1/12, c₂=1/6))
-
+VerstappenAnisotropicMinimumDissipation{Float64} turbulence closure with:
+           Poincaré constant for momentum eddy viscosity Cν: 0.08333333333333333
+    Poincaré constant for tracer(s) eddy diffusivit(ies) Cκ: (c₁ = 0.08333333333333333, c₂ = 0.16666666666666666)
+                        Buoyancy modification multiplier Cb: 0.0
+                Background diffusivit(ies) for tracer(s), κ: 1.46e-7
+             Background kinematic viscosity for momentum, ν: 1.05e-6
 
 References
 ==========
