### 2020-02-11 f37da12
Draft some lazy field abstractions
