### 2020-02-11 99727d6
Re-organize densities and thermodynamic variables
