+++ src/models.jl
@@ -1,5 +1,6 @@
-using Oceananigans
+import Base: getproperty
 
+using Oceananigans
 using Oceananigans.Models: AbstractModel, Clock
 using Oceananigans.Forcing: zeroforcing
 
@@ -7,21 +8,21 @@ using Oceananigans.Forcing: zeroforcing
 ##### Definition of a compressible model
 #####
 
-mutable struct CompressibleModel{A, FT, G, M, D, T, TS, TD, C, TC, DF, MP, F, P, SF, RHS, IV, ATS} <: AbstractModel
+mutable struct CompressibleModel{A, FT, G, M, GS, T, TS, D, TD, C, TC, DF, MP, F, SF, RHS, IV, ATS} <: AbstractModel
              architecture :: A
                      grid :: G
                     clock :: Clock{FT}
                   momenta :: M
-                densities :: D
+                    gases :: GS
    thermodynamic_variable :: T
              microphysics :: MP
                   tracers :: TS
+                densities :: D
             total_density :: TD
                  coriolis :: C
                   closure :: TC
             diffusivities :: DF
                   forcing :: F
-               parameters :: P
                   gravity :: FT
             slow_forcings :: SF
          right_hand_sides :: RHS
@@ -42,7 +43,7 @@ function CompressibleModel(;
                float_type = Float64,
                     clock = Clock{float_type}(0, 0),
                   momenta = MomentumFields(architecture, grid),
-                densities = DryEarth(float_type),
+                    gases = DryEarth(float_type),
    thermodynamic_variable = PrognosticEntropy(),
              microphysics = nothing,
             extra_tracers = nothing,
@@ -51,7 +52,6 @@ function CompressibleModel(;
                   closure = ConstantIsotropicDiffusivity(float_type, ν=0.5, κ=0.5),
             diffusivities = TurbulentDiffusivities(architecture, grid, tracernames, closure),
                   forcing = ModelForcing(),
-               parameters = nothing,
                   gravity = g_Earth,
             slow_forcings = ForcingFields(architecture, grid, tracernames),
          right_hand_sides = RightHandSideFields(architecture, grid, tracernames),
@@ -63,11 +63,16 @@ function CompressibleModel(;
     tracers = TracerFields(architecture, grid, tracernames)
     forcing = ModelForcing(tracernames, forcing)
     closure = with_tracers(tracernames, closure)
+    densities = create_density_pointers(gases, tracers,
+        forcings, slow_forcings, right_hand_sides, intermediate_vars)
+    thermodynamic_variable = create_thermodynamic_pointers(
+        thermodynamic_variable, tracers, forcings, slow_forcings,
+        right_hand_sides, intermediate_vars)
     total_density = CellField(architecture, grid)
 
-    return CompressibleModel(architecture, grid, clock, momenta, densities, thermodynamic_variable,
-                             microphysics, tracers, total_density, coriolis, closure, diffusivities,
-                             forcing, parameters, gravity, slow_forcings,
+    return CompressibleModel(architecture, grid, clock, momenta, gases, thermodynamic_variable,
+                             microphysics, tracers, densities, total_density, coriolis, closure, diffusivities,
+                             forcing, gravity, slow_forcings,
                              right_hand_sides, intermediate_vars, acoustic_time_stepper)
 end
 
@@ -85,6 +90,28 @@ function collect_tracers(args...)
         end
     end
     return Tuple(list)
+end
+
+function create_density_pointers(gases, tracers, forcings, slow_forcings,
+    right_hand_sides, intermediate_vars)
+    return ((gas = getproperty(gases, key),
+             ρ = getproperty(tracers, key),
+             forcing = getproperty(forcings, key),
[…]
