### 2020-02-11 d1ee83c
Minor changes to model field organization
