### 2020-02-12 7014037
Allow 2D GPU models
