### 2020-02-14 1d1ae64
Much cleaner `FieldBoundaryConditions` constructor
