+++ src/BoundaryConditions/BoundaryConditions.jl
@@ -15,6 +15,7 @@ export
 using CUDAnative
 
 using Oceananigans.Architectures
+using Oceananigans.Grids
 using Oceananigans.Fields
 
 include("boundary_condition_types.jl")
