### 2020-02-14 f61206d
Generalize `HorizontallyPeriodic` and `Channel` solution BCs.
