+++ src/BoundaryConditions/solution_and_model_boundary_conditions.jl
@@ -52,7 +52,7 @@ function PressureBoundaryConditions(solution_boundary_conditions)
     x = CoordinateBoundaryConditions(PressureBC(ubcs.x.left), PressureBC(ubcs.x.right))
     y = CoordinateBoundaryConditions(PressureBC(ubcs.y.left), PressureBC(ubcs.y.right))
     z = CoordinateBoundaryConditions(PressureBC(ubcs.z.left), PressureBC(ubcs.z.right))
-    return (x=x, y=y, z=z)
+    return FieldBoundaryConditions(x, y, z)
 end
 
 PressureBoundaryConditions(model_boundary_conditions::ModelBoundaryConditions) =
@@ -72,7 +72,7 @@ function DiffusivityBoundaryConditions(solution_boundary_conditions)
     x = CoordinateBoundaryConditions(DiffusivityBC(ubcs.x.left), DiffusivityBC(ubcs.x.right))
     y = CoordinateBoundaryConditions(DiffusivityBC(ubcs.y.left), DiffusivityBC(ubcs.y.right))
     z = CoordinateBoundaryConditions(DiffusivityBC(ubcs.z.left), DiffusivityBC(ubcs.z.right))
-    return (x=x, y=y, z=z)
+    return FieldBoundaryConditions(x, y, z)
 end
 
 DiffusivitiesBoundaryConditions(::Nothing, args...) = nothing
@@ -107,41 +107,10 @@ and may contain some or all of the boundary conditions on `tracers`.
 SolutionBoundaryConditions(tracer_names, proposal_bcs) =
     with_tracers(tracer_names, proposal_bcs, default_tracer_bcs, with_velocities=true)
 
-"""
-    HorizontallyPeriodicSolutionBCs(u=HorizontallyPeriodicBCs(), ...)
-
-Construct `SolutionBoundaryConditions` for a horizontally-periodic model
-configuration with solution fields `u`, `v`, `w`, `T`, and `S` specified by keyword arguments.
-
-By default `HorizontallyPeriodicBCs` are applied to `u`, `v`, `T`, and `S`
-and `HorizontallyPeriodicBCs(top=NoPenetrationBC(), bottom=NoPenetrationBC())` is applied to `w`.
-
-Use `HorizontallyPeriodicBCs` when constructing non-default boundary conditions for `u`, `v`, `w`, `T`, `S`.
-"""
-function HorizontallyPeriodicSolutionBCs(;
-    u = HorizontallyPeriodicBCs(),
-    v = HorizontallyPeriodicBCs(),
-    w = HorizontallyPeriodicBCs(top=NoPenetrationBC(), bottom=NoPenetrationBC()),
-    tracers_boundary_conditions...)
-
-    return merge((u=u, v=v, w=w), tracers_boundary_conditions)
-end
-
-"""
-    ChannelSolutionBCs(u=ChannelBCs(), ...)
-
-Construct `SolutionBoundaryConditions` for a reentrant channel model
-configuration with solution fields `u`, `v`, `w`, `T`, and `S` specified by keyword arguments.
-
-By default `ChannelBCs` are applied to `u`, `v`, `T`, and `S`
-and `ChannelBCs(top=NoPenetrationBC(), bottom=NoPenetrationBC())` is applied to `w`.
-
-Use `ChannelBCs` when constructing non-default boundary conditions for `u`, `v`, `w`, `T`, `S`.
-"""
-function ChannelSolutionBCs(;
-    u = ChannelBCs(),
-    v = ChannelBCs(north=NoPenetrationBC(), south=NoPenetrationBC()),
-    w = ChannelBCs(top=NoPenetrationBC(), bottom=NoPenetrationBC()),
[…]
