### 2020-02-14 ee696d1
Consider field type when constructing field boundary conditions
