### 2020-02-14 479eef6
Run diagnostics and output writers in `run!(simulation)`
