+++ src/Simulation.jl
@@ -5,13 +5,13 @@ export Simulation, run!
 using Oceananigans.Models
 using Oceananigans.Utils
 
-struct Simulation{M, T, SI, ST, W, R, D, O, P}
+mutable struct Simulation{M, T, SI, ST, W, R, D, O, P}
                  model :: M
                     Δt :: T
         stop_iteration :: SI
              stop_time :: ST
        wall_time_limit :: W
-               runtime :: R
+              run_time :: R
            diagnostics :: D
         output_writers :: O
               progress :: P
@@ -33,10 +33,10 @@ function Simulation(model, Δt;
                "= wall time limit = Inf."
    end
 
-   runtime = 0.0
-
+   run_time = 0.0
+   
    return Simulation(model, Δt, stop_iteration, stop_time, wall_time_limit,
-                     runtime, diagnostics, output_writers, progress)
+                     run_time, diagnostics, output_writers, progress)
 end
 
 get_Δt(Δt) = Δt
@@ -57,27 +57,40 @@ function run!(sim)
             stop_simulation = true
       end
 
-      if sim.runtime > sim.wall_time_limit
-            @warn "Simulation will not be run! Simulation run time $(sim.runtime) " *
+      if sim.run_time > sim.wall_time_limit
+            @warn "Simulation will not be run! Simulation run time $(sim.run_time) " *
                   "has exceeded simulation wall time limit $(sim.wall_time_limit)."
             stop_simulation = true
       end
 
+      clock = sim.model.clock
       while !stop_simulation
             time_before = time()
+
+            if clock.iteration == 0
+                [run_diagnostic(sim.model, diag) for diag in values(sim.diagnostics)]
+                [write_output(sim.model, out)    for out  in values(sim.output_writers)]
+            end
+
+            for n in 1:sim.progress_frequency
+                time_step!(model, Δt, euler=n==1)
+
+                [time_to_run(clock, diag) && run_diagnostic(sim.model, diag) for diag in values(diagnostics)]
+                [time_to_run(clock, out)  && write_output(sim.model, out)    for out  in values(output_writers)]
+            end
 
             time_step!(sim.model, Δt=get_Δt(sim.Δt), Nt=sim.progress_frequency,
                        diagnostics=sim.diagnostics, output_writers=sim.output_writers)
 
             Δt isa TimeStepWizard && update_Δt!(Δt, model)
-            progress isa Function && progress(model)
+            progress isa Function && progress(simulation)
 
             time_after = time()
-            sim.runtime += time_after - time_before
+            sim.run_time += time_after - time_before
 
             sim.model.clock.iteration > sim.stop_iteration && stop_simulation = true
             sim.model.clock.time > sim.stop_time && stop_simulation = true
-            sim.runtime > sim.wall_time_limit && stop_simulation = true
+            sim.run_time > sim.wall_time_limit && stop_simulation = true
       end
 end
 
