### 2020-02-14 1647d00
Stop simulation based on list of stop criteria
