+++ src/Simulation.jl
@@ -5,9 +5,10 @@ export Simulation, run!
 using Oceananigans.Models
 using Oceananigans.Utils
 
-mutable struct Simulation{M, T, SI, ST, W, R, D, O, P}
+mutable struct Simulation{M, T, S, SI, ST, W, R, D, O, P}
                  model :: M
                     Δt :: T
+         stop_criteria :: S
         stop_iteration :: SI
              stop_time :: ST
        wall_time_limit :: W
@@ -19,6 +20,7 @@ mutable struct Simulation{M, T, SI, ST, W, R, D, O, P}
 end
 
 function Simulation(model, Δt;
+        stop_criteria = Function[iteration_limit_exceeded, stop_time_exceeded, wall_time_limit_exceeded]
        stop_iteration = Inf,
             stop_time = Inf,
       wall_time_limit = Inf,
@@ -34,64 +36,76 @@ function Simulation(model, Δt;
    end
 
    run_time = 0.0
-   
+
    return Simulation(model, Δt, stop_iteration, stop_time, wall_time_limit,
                      run_time, diagnostics, output_writers, progress)
+end
+
+function stop(simulation)
+    for sc in sim.stop_criteria
+        sc(sim) && return true
+    end
+    return false
+end
+
+function iteration_limit_exceeded(sim)
+    if sim.model.clock.iteration > sim.stop_iteration
+          @warn "Simulation is stopping. Model iteration $(sim.model.clock.iteration) " *
+                "has exceeded simulation stop iteration $(sim.stop_iteration)."
+          return true
+    end
+    return false
+end
+
+function stop_time_exceeded(sim)
+    if sim.model.clock.time > sim.stop_time
+          @warn "Simulation is stopping. Model time $(sim.model.clock.time) " *
+                "has exceeded simulation stop time $(sim.stop_time)."
+          return true
+    end
+    return false
+end
+
+function wall_time_limit_exceeded(sim)
+    if sim.run_time > sim.wall_time_limit
+          @warn "Simulation is stopping. Simulation run time $(sim.run_time) " *
+                "has exceeded simulation wall time limit $(sim.wall_time_limit)."
+          return true
+    end
+    return false
 end
 
 get_Δt(Δt) = Δt
 get_Δt(wizard::TimeStepWizard) = wizard.Δt
 
 function run!(sim)
-      stop_simulation = false
+    model = sim.model
+    clock = model.clock
 
-      if sim.model.clock.iteration > sim.stop_iteration
-            @warn "Simulation will not be run! Model iteration $(sim.model.clock.iteration) " *
-                  "has exceeded simulation stop iteration $(sim.stop_iteration)."
-            stop_simulation = true
-      end
+    while !stop(sim)
+        time_before = time()
 
-      if sim.model.clock.time > sim.stop_time
-            @warn "Simulation will not be run! Model time $(sim.model.clock.time) " *
-                  "has exceeded simulation stop time $(sim.stop_time)."
-            stop_simulation = true
-      end
+        if clock.iteration == 0
+            [run_diagnostic(sim.model, diag) for diag in values(sim.diagnostics)]
+            [write_output(sim.model, out)    for out  in values(sim.output_writers)]
+        end
 
-      if sim.run_time > sim.wall_time_limit
-            @warn "Simulation will not be run! Simulation run time $(sim.run_time) " *
-                  "has exceeded simulation wall time limit $(sim.wall_time_limit)."
-            stop_simulation = true
-      end
+        for n in 1:sim.progress_frequency
+            time_step!(model, Δt, euler=n==1)
 
-      clock = sim.model.clock
-      while !stop_simulation
-            time_before = time()
+            [time_to_run(clock, diag) && run_diagnostic(sim.model, diag) for diag in values(diagnostics)]
[…]
