### 2020-02-14 b967a8f
Count stop criteria evaluation in simulation run time
