### 2020-02-14 b092e6f
Actually enforce no normal flow BCs by default
