### 2020-02-14 5cf5c66
Nuke the `ChannelModel`
