+++ src/Models/Models.jl
@@ -21,7 +21,6 @@ abstract type AbstractModel end
 
 include("clock.jl")
 include("model.jl")
-include("channel_model.jl")
 include("non_dimensional_model.jl")
 include("show_models.jl")
 
