+++ src/Models/channel_model.jl
@@ -1,12 +1,0 @@
-using Oceananigans.BoundaryConditions: ChannelSolutionBCs
-
-"""
-    ChannelModel(; kwargs...)
-
-Construct a `Model` with walls in the y-direction. This is done by imposing
-`FreeSlip` boundary conditions in the y-direction instead of `Periodic`.
-
-kwargs are passed to the regular `Model` constructor.
-"""
-ChannelModel(; boundary_conditions=ChannelSolutionBCs(), kwargs...) =
-    Model(; boundary_conditions=boundary_conditions, kwargs...)
