+++ src/Oceananigans.jl
@@ -35,9 +35,10 @@ export
     # Boundary conditions
     BoundaryCondition,
     Flux, Gradient, Value,
-    CoordinateBoundaryConditions, FieldBoundaryConditions, HorizontallyPeriodicBCs, ChannelBCs,
-    BoundaryConditions, SolutionBoundaryConditions, HorizontallyPeriodicSolutionBCs, ChannelSolutionBCs,
-    BoundaryFunction, getbc, setbc!,
+    CoordinateBoundaryConditions, FieldBoundaryConditions, UVelocityBoundaryConditions,
+    VVelocityBoundaryConditions, WVelocityBoundaryConditions, TracerBoundaryConditions,
+    PressureBoundaryConditions, SolutionBoundaryConditions,
+    BoundaryFunction,
 
     # Time stepping
     time_step!,
