### 2020-02-14 ad6f2ea
Fixes and updates tests
