### 2020-02-15 4f43f97
Update examples and verification tests
