### 2020-02-15 7ad5c8a
Turns out there were two other uses of `Simulation` already
