+++ src/Oceananigans.jl
@@ -9,7 +9,7 @@ export
     CPU, GPU,
 
     # Logging
-    ModelLogger, Diagnostic, Setup, Simulation,
+    ModelLogger, Diagnostic, Setup,
 
     # Grids
     Periodic, Bounded, Flat,
@@ -45,6 +45,10 @@ export
 
     # Models
     Model, ChannelModel, NonDimensionalModel,
+
+    # Simulation
+    Simulation, run!,
+    iteration_limit_exceeded, stop_time_exceeded, wall_time_limit_exceeded,
 
     # Utilities
     prettytime, pretty_filesize,
@@ -150,6 +154,7 @@ include("Models/Models.jl")
 include("Diagnostics/Diagnostics.jl")
 include("OutputWriters/OutputWriters.jl")
 include("TimeSteppers/TimeSteppers.jl")
+include("Simulations.jl")
 include("AbstractOperations/AbstractOperations.jl")
 
 #####
