### 2020-02-15 c0838bd
Fix simulation typos
