### 2020-02-15 83a5a38
Update model setup docs to document grid topologies
