### 2020-02-15 355ea2f
More readable constructors for boundary conditions
