### 2020-02-15 0dc06a4
Update model setup docs for boundary conditions
