+++ docs/src/model_setup/boundary_conditions.md
@@ -1,16 +1,13 @@
 # Boundary conditions
-A boundary condition is applied to each field, dimension, and endpoint. There are left and right (or bottom and top)
-boundary conditions for each of the x, y, and z dimensions so each field is associated with 6 boundary conditions. Each
-of these boundary conditions may be specified individually. Each boundary condition can be specified via a constant
-value, an array, or a function.
+A boundary condition is applied to each field, dimension, and endpoint. There are left and right boundary conditions
+for each of the x, y, and z dimensions so each field is associated with 6 boundary conditions. Each of these boundary
+conditions may be specified individually. Each boundary condition can be specified via a constant value, an array, or
+a function.
+
+The left and right boundary conditions associated with the x-dimension are called west and east, respectively. For the
+y-dimension, left and right are called south and north. For the z-dimension, left and right are called bottom and top.
 
 See [Numerical implementation of boundary conditions](@ref numerical_bcs) for more details.
-
-!!! warning "Consistent boundary conditions"
-    Be careful to ensure that you don't set up a model with inconsistent boundary conditions. For example, periodic
-    boundary conditions should remain imposed on all fields and endpoints for periodic dimensions, and velocities
-    normal to a wall (e.g. vertical velocity w with walls at the top and bottom) must have no-penetration boundary
-    conditions.
 
 ## Types of boundary conditions
 1. [`Periodic`](@ref Periodic)
@@ -20,16 +17,16 @@ See [Numerical implementation of boundary conditions](@ref numerical_bcs) for mo
 5. [`No-penetration`](@ref NoPenetration)
 
 Notice that open boundary conditions and radiation boundary conditions can be imposed via flux or value boundary
-conditions defined by a function or array.
+conditions defined by a function or array. Or alternatively, through a forcing function which is more flexible.
 
 ## Default boundary conditions
-By default, periodic boundary conditions are applied on all fields along periodic dimensions. All other boundary
-conditions are no-flux, except for velocities normal to a wall which get no-penetration boundary conditions.
+By default, periodic boundary conditions are applied on all fields along periodic dimensions. Otherwise tracers
+get no-flux boundary conditions and velocities get free-slip and no-penetration boundary conditions.
 
 ## Boundary condition structures
 Oceananigans uses a hierarchical structure to expressing boundary conditions.
 1. A [`BoundaryCondition`](@ref) is associated with every field, dimension, and endpoint.
-2. Boundary conditions specifying the condition at the left and right endpoints (or top and bottom endpoints) are
+2. Boundary conditions specifying the condition at the left and right endpoints are
    grouped into [`CoordinateBoundaryConditions`](@ref).
 3. A set of three `CoordinateBoundaryConditions` specifying the boundary conditions along the x, y, and z dimensions
    for a single field are grouped into a [`FieldBoundaryConditions`](@ref) named tuple.
@@ -39,13 +36,8 @@ Oceananigans uses a hierarchical structure to expressing boundary conditions.
 Boundary conditions are defined at model construction time by passing a named tuple of `FieldBoundaryConditions`
 specifying boundary conditions on every field: velocities ($u$, $v$, $w$) and all tracers.
 
[…]
