### 2020-02-15 d7753f2
Test default field boundary conditions with different grid topologies
