### 2020-02-15 f098c33
Simulation docstrings
