+++ src/Simulations.jl
@@ -12,12 +12,12 @@ using Oceananigans.Utils
 
 using Oceananigans: AbstractDiagnostic, AbstractOutputWriter
 
-mutable struct Simulation{M, Δ, S, SI, ST, W, R, D, O, P, F}
+mutable struct Simulation{M, Δ, S, T, W, R, D, O, P, F}
                  model :: M
                     Δt :: Δ
          stop_criteria :: S
-        stop_iteration :: SI
-             stop_time :: ST
+        stop_iteration :: Int
+             stop_time :: T
        wall_time_limit :: W
               run_time :: R
            diagnostics :: D
@@ -26,6 +26,35 @@ mutable struct Simulation{M, Δ, S, SI, ST, W, R, D, O, P, F}
     progress_frequency :: F
 end
 
+"""
+    Simulation(model; Δt,
+         stop_criteria = Function[iteration_limit_exceeded, stop_time_exceeded, wall_time_limit_exceeded],
+        stop_iteration = Inf,
+             stop_time = Inf,
+       wall_time_limit = Inf,
+           diagnostics = OrderedDict{Symbol, AbstractDiagnostic}(),
+        output_writers = OrderedDict{Symbol, AbstractOutputWriter}(),
+              progress = nothing,
+    progress_frequency = 1)
+
+Construct an Oceananigans.jl `Simulation` for a `model` with time step `Δt`.
+
+Keyword arguments
+=================
+- `Δt`: Required keyword argument specifying the simulation time step. Can be a `Number`
+  for constant time steps or a `TimeStepWizard` for adaptive time-stepping.
+- `stop_critera`: A list of functions (each taking a single argument, the `simulation`).
+  If any of the functions return `true` when the stop criteria is evaluated the simulation
+  will stop.
+- `stop_iteration`: Stop the simulation after this many iterations.
+- `stop_time`: Stop the simulation once this much model clock time has passed.
+- `wall_time_limit`: Stop the simulation if it's been running for longer than this many
+   seconds of wall clock time.
+- `progress`: A function with a single argument, the `simulation`. Will be called every
+  `progress_frequency` iterations. Useful for logging simulation health.
+- `progress_frequency`: How often to check stop criteria and call `progress` function
+  (in number of iterations).
+"""
 function Simulation(model; Δt,
         stop_criteria = Function[iteration_limit_exceeded, stop_time_exceeded, wall_time_limit_exceeded],
        stop_iteration = Inf,
@@ -34,8 +63,7 @@ function Simulation(model; Δt,
           diagnostics = OrderedDict{Symbol, AbstractDiagnostic}(),
        output_writers = OrderedDict{Symbol, AbstractOutputWriter}(),
              progress = nothing,
-   progress_frequency = 1
-   )
+   progress_frequency = 1)
 
    if stop_iteration == Inf && stop_time == Inf && wall_time_limit == Inf
          @warn "This simulation will run forever as stop iteration = stop time " *
@@ -95,6 +123,12 @@ end
 get_Δt(Δt) = Δt
 get_Δt(wizard::TimeStepWizard) = wizard.Δt
 
+"""
+    run!(simulation)
+
+Run a `simulation` until one of the stop criteria evaluates to true and the
+simulation stops.
+"""
 function run!(sim)
     model = sim.model
     clock = model.clock
