### 2020-02-15 3eecef7
Pretty printing simulations
