### 2020-02-15 05e02fa
Only test Smagorinsky-Lilly time-stepping once
