+++ test/test_time_stepping.jl
@@ -3,7 +3,7 @@ function time_stepping_works_with_closure(arch, FT, Closure)
     grid = RegularCartesianGrid(FT; size=(16, 16, 16), halo=(2, 2, 2), length=(1, 2, 3))
 
     model = Model(grid=grid, architecture=arch, float_type=FT, closure=Closure(FT))
-    time_step!(model, 1)
+    time_step!(model, 1, euler=true)
 
     return true  # Test that no errors/crashes happen when time stepping.
 end
@@ -15,7 +15,7 @@ function time_stepping_works_with_nonlinear_eos(arch, FT, eos_type)
     b = SeawaterBuoyancy(equation_of_state=eos)
 
     model = Model(architecture=arch, float_type=FT, grid=grid, buoyancy=b)
-    time_step!(model, 1)
+    time_step!(model, 1, euler=true)
 
     return true  # Test that no errors/crashes happen when time stepping.
 end
@@ -24,7 +24,7 @@ function run_first_AB2_time_step_tests(arch, FT)
     add_ones(args...) = 1.0
     model = Model(grid=RegularCartesianGrid(FT; size=(16, 16, 16), length=(1, 2, 3)),
                   architecture=arch, float_type=FT, forcing=ModelForcing(T=add_ones))
-    time_step!(model, 1)
+    time_step!(model, 1, euler=true)
 
     # Test that GT = 1 after first time step and that AB2 actually reduced to forward Euler.
     @test all(interior(model.timestepper.Gⁿ.u) .≈ 0)
@@ -72,7 +72,8 @@ function compute_w_from_continuity(arch, FT)
     max_div = maximum(interior(div_u))
     sum_div = sum(interior(div_u))
     abs_sum_div = sum(abs.(interior(div_u)))
-    @info "Velocity divergence after recomputing w [$arch, $FT]: min=$min_div, max=$max_div, sum=$sum_div, abs_sum=$abs_sum_div"
+    @info "Velocity divergence after recomputing w [$(typeof(arch)), $FT]: " *
+          "min=$min_div, max=$max_div, sum=$sum_div, abs_sum=$abs_sum_div"
 
     return all(isapprox.(interior(div_u), 0; atol=5*eps(FT)))
 end
@@ -96,8 +97,8 @@ function incompressible_in_time(arch, FT, Nt)
     # Just add a temperature perturbation so we get some velocity field.
     @. model.tracers.T.data[8:24, 8:24, 8:24] += 0.01
 
-    for _ in 1:Nt
-        time_step!(model, 0.05)
+    for n in 1:Nt
+        time_step!(model, 0.05, euler = n==1)
     end
 
     velocity_div!(grid, u, v, w, div_u)
@@ -106,16 +107,14 @@ function incompressible_in_time(arch, FT, Nt)
     max_div = minimum(interior(div_u))
     sum_div = sum(interior(div_u))
     abs_sum_div = sum(abs.(interior(div_u)))
-    @info "Velocity divergence after $Nt time steps [$arch, $FT]: min=$min_div, max=$max_div, sum=$sum_div, abs_sum=$abs_sum_div"
+    @info "Velocity divergence after $Nt time steps [$(typeof(arch)), $FT]: " *
+          "min=$min_div, max=$max_div, sum=$sum_div, abs_sum=$abs_sum_div"
 
     # We are comparing with 0 so we use absolute tolerances. They are a bit larger than eps(Float64) and eps(Float32)
     # because we are summing over the absolute value of many machine epsilons. A better atol value may be
     # Nx*Ny*Nz*eps(FT) but it's much higher than the observed abs_sum_div.
-    if FT == Float64
-        return isapprox(abs_sum_div, 0; atol=5e-16)
[…]
