### 2020-02-15 c964988
Remove diagnostics and output writers from model
