### 2020-02-15 df297e8
Fix stop criteria
