### 2020-02-15 cb8ed64
Logic for when to take a forward Euler step
