### 2020-02-15 1158991
Update tests to use simulations
