+++ src/Models/show_models.jl
@@ -2,8 +2,8 @@ using Oceananigans.Grids: short_show
 using Oceananigans.Utils: prettytime, ordered_dict_show
 
 """Show the innards of a `Model` in the REPL."""
-Base.show(io::IO, model::Model) =
-    print(io, "Oceananigans.Model on a $(typeof(model.architecture)) architecture ",
+Base.show(io::IO, model::IncompressibleModel) =
+    print(io, "Oceananigans.IncompressibleModel on a $(typeof(model.architecture)) architecture ",
         "(time = $(prettytime(model.clock.time)), iteration = $(model.clock.iteration)) \n",
         "├── grid: $(short_show(model.grid))\n",
         "├── tracers: $(tracernames(model.tracers))\n",
