+++ test/regression_tests/thermal_bubble_regression_test.jl
@@ -5,7 +5,7 @@ function run_thermal_bubble_regression_test(arch)
 
     grid = RegularCartesianGrid(size=(Nx, Ny, Nz), length=(Lx, Ly, Lz))
     closure = ConstantIsotropicDiffusivity(ν=4e-2, κ=4e-2)
-    model = Model(architecture=arch, grid=grid, closure=closure, coriolis=FPlane(f=1e-4))
+    model = IncompressibleModel(architecture=arch, grid=grid, closure=closure, coriolis=FPlane(f=1e-4))
     simulation = Simulation(model, Δt=6, stop_iteration=10)
 
     model.tracers.T.data.parent .= 9.85
