### 2020-02-15 36aa360
Add basic Simulation tests
