### 2020-02-15 6a409e7
Update stratified Couette flow verification experiment to use Simulation
