+++ verification/stratified_couette_flow/stratified_couette_flow.jl
@@ -71,7 +71,8 @@ function (Nu::NusseltNumber)(model)
 end
 
 """
-    simulate_stratified_couette_flow(; Nxy, Nz, h=1, U_wall=1, Re=4250, Pr=0.7, Ri, Ni=10, end_time=1000)
+    simulate_stratified_couette_flow(; Nxy, Nz, h=1, U_wall=1, Re=4250, Pr=0.7,
+                                     Ri, Ni=10, end_time=1000)
 
 Simulate stratified plane Couette flow with `Nxy` grid cells in each horizontal
 direction, `Nz` grid cells in the vertical, in a domain of size (4πh, 2πh, 2h),
@@ -81,7 +82,8 @@ number `Re, Prandtl number `Pr`, and Richardson number `Ri`.
 `Ni` is the number of "intermediate" time steps taken at a time before printing a progress
 statement and updating the time step.
 """
-function simulate_stratified_couette_flow(; Nxy, Nz, arch=GPU(), h=1, U_wall=1, Re=4250, Pr=0.7, Ri, Ni=10, end_time=1000)
+function simulate_stratified_couette_flow(; Nxy, Nz, arch=GPU(), h=1, U_wall=1,
+                                          Re=4250, Pr=0.7, Ri, Ni=10, end_time=1000)
     #####
     ##### Computed parameters
     #####
@@ -182,22 +184,21 @@ function simulate_stratified_couette_flow(; Nxy, Nz, arch=GPU(), h=1, U_wall=1, 
    :kappaT => model -> Array(model.diffusivities.κₑ.T.data.parent),
        :nu => model -> Array(model.diffusivities.νₑ.data.parent))
 
-    field_writer = JLD2OutputWriter(model, fields; dir=base_dir, prefix=prefix * "_fields",
-                                    init=init_save_parameters_and_bcs,
-                                    interval=10, force=true, verbose=true)
-
-    push!(model.output_writers, field_writer)
+    field_writer =
+        JLD2OutputWriter(model, fields, dir=base_dir, prefix=prefix * "_fields",
+                         init=init_save_parameters_and_bcs, interval=10,
+                         force=true, verbose=true)
 
     #####
     ##### Set up profile output writer
     #####
 
-    Uavg = HorizontalAverage(model.velocities.u;       return_type=Array)
-    Vavg = HorizontalAverage(model.velocities.v;       return_type=Array)
-    Wavg = HorizontalAverage(model.velocities.w;       return_type=Array)
-    Tavg = HorizontalAverage(model.tracers.T;          return_type=Array)
-    νavg = HorizontalAverage(model.diffusivities.νₑ;   return_type=Array)
-    κavg = HorizontalAverage(model.diffusivities.κₑ.T; return_type=Array)
+    Uavg = HorizontalAverage(model.velocities.u,       return_type=Array)
+    Vavg = HorizontalAverage(model.velocities.v,       return_type=Array)
+    Wavg = HorizontalAverage(model.velocities.w,       return_type=Array)
+    Tavg = HorizontalAverage(model.tracers.T,          return_type=Array)
+    νavg = HorizontalAverage(model.diffusivities.νₑ,   return_type=Array)
+    κavg = HorizontalAverage(model.diffusivities.κₑ.T, return_type=Array)
 
     profiles = Dict(
          :u => model -> Uavg(model),
@@ -207,10 +208,10 @@ function simulate_stratified_couette_flow(; Nxy, Nz, arch=GPU(), h=1, U_wall=1, 
         :nu => model -> νavg(model),
     :kappaT => model -> κavg(model))
 
[…]
