### 2020-02-15 8fe73c6
Consolidate all functions called fill_halo_regions into fill_halo_regions.jl
