### 2020-02-15 4ca523b
Bugfix in field.jl
