+++ src/Fields/field.jl
@@ -72,7 +72,7 @@ Field(L::Tuple, data::AbstractArray, grid) = Field{L[1], L[2], L[3]}(data, grid)
 Construct a `Field` on architecture `arch` and `grid` at location `L`,
 where `L` is a tuple of `Cell` or `Face` types.
 """
-Field(L::Tuple, arch, grid) = Field{L[1], L[2], L[3]}(zeros(arch, grid), grid)
+Field(L::Tuple, arch, grid) = Field{L[1], L[2], L[3]}(zeros(arch, grid, L), grid)
 
 """
     Field(X, Y, Z, arch, grid)
@@ -88,7 +88,7 @@ Field(X, Y, Z, arch, grid) =  Field((X, Y, Z), arch, grid)
 Return a `Field{Cell, Cell, Cell}` on architecture `arch` and `grid`.
 Used for tracers and pressure fields.
 """
-CellField(T, arch, grid) = Field{Cell, Cell, Cell}(zeros(T, arch, grid), grid)
+CellField(T, arch, grid) = Field{Cell, Cell, Cell}(zeros(T, arch, grid, (Cell, Cell, Cell)), grid)
 
 """
     FaceFieldX([T=eltype(grid)], arch, grid)
@@ -96,7 +96,7 @@ CellField(T, arch, grid) = Field{Cell, Cell, Cell}(zeros(T, arch, grid), grid)
 Return a `Field{Face, Cell, Cell}` on architecture `arch` and `grid`.
 Used for the x-velocity field.
 """
-FaceFieldX(T, arch, grid) = Field{Face, Cell, Cell}(zeros(T, arch, grid), grid)
+FaceFieldX(T, arch, grid) = Field{Face, Cell, Cell}(zeros(T, arch, grid, (Face, Cell, Cell)), grid)
 
 """
     FaceFieldY([T=eltype(grid)], arch, grid)
@@ -104,7 +104,7 @@ FaceFieldX(T, arch, grid) = Field{Face, Cell, Cell}(zeros(T, arch, grid), grid)
 Return a `Field{Cell, Face, Cell}` on architecture `arch` and `grid`.
 Used for the y-velocity field.
 """
-FaceFieldY(T, arch, grid) = Field{Cell, Face, Cell}(zeros(T, arch, grid), grid)
+FaceFieldY(T, arch, grid) = Field{Cell, Face, Cell}(zeros(T, arch, grid, (Cell, Face, Cell)), grid)
 
 """
     FaceFieldZ([T=eltype(grid)], arch, grid)
@@ -112,7 +112,7 @@ FaceFieldY(T, arch, grid) = Field{Cell, Face, Cell}(zeros(T, arch, grid), grid)
 Return a `Field{Cell, Cell, Face}` on architecture `arch` and `grid`.
 Used for the z-velocity field.
 """
-FaceFieldZ(T, arch, grid) = Field{Cell, Cell, Face}(zeros(T, arch, grid), grid)
+FaceFieldZ(T, arch, grid) = Field{Cell, Cell, Face}(zeros(T, arch, grid, (Cell, Cell, Face)), grid)
 
  CellField(arch, grid) = Field((Cell, Cell, Cell), arch, grid)
 FaceFieldX(arch, grid) = Field((Face, Cell, Cell), arch, grid)
@@ -188,6 +188,12 @@ const AbstractCPUField =
 ##### Creating fields by dispatching on architecture
 #####
 
+field_length(::Type{Cell}, N, H) = N + 2H
+field_length(::Type{Face}, N, H) = N + 1 + 2H
+
+field_length(::Cell, N, H) = N + 2H
+field_length(::Face, N, H) = N + 1 + 2H
+
 function OffsetArray(underlying_data, grid)
[…]
