+++ src/BoundaryConditions/BoundaryConditions.jl
@@ -34,7 +34,6 @@ include("fill_halo_regions.jl")
 include("zero_halo_regions.jl")
 
 include("apply_flux_bcs.jl")
-include("apply_flux_periodic_no_penetration_bcs.jl")
 include("apply_value_gradient_bcs.jl")
 
 end
