### 2020-02-15 5651954
Simulation stop criteria can be a callable object
