### 2020-02-15 397c15d
Update examples to use simulations
