### 2020-02-16 c417701
Get rid of `AbstractLocatedField` in favor of just `AbstractField`
