+++ src/Fields/show_fields.jl
@@ -1,8 +1,6 @@
-show_location(X, Y, Z) = string("(", string(typeof(X())), ", ",
-                                     string(typeof(Y())), ", ",
-                                     string(typeof(Z())), ")")
+show_location(X, Y, Z) = "($(typeof(X())), $(typeof(Y())), $(typeof(Z())))"
 
-show_location(field::AbstractLocatedField{X, Y, Z}) where {X, Y, Z} = show_location(X, Y, Z)
+show_location(field::AbstractField{X, Y, Z}) where {X, Y, Z} = show_location(X, Y, Z)
 
 short_show(a) = string(typeof(a))
 shortname(a::Array) = string(typeof(a).name.wrapper)
@@ -15,4 +13,4 @@ show(io::IO, field::Field) =
           "    ├── size: ", size(field.grid), '\n',
           "    └── domain: ", show_domain(field.grid), '\n')
 
-short_show(field::AbstractLocatedField) = string("Field at ", show_location(field))
+short_show(field::AbstractField) = string("Field at ", show_location(field))
