### 2020-02-16 4d80063
Don't allocate CPU array when calling zeros for GPU
