+++ src/Fields/show_fields.jl
@@ -1,16 +1,16 @@
+import Base: show
+
+using Oceananigans.Grids: show_domain
+
 show_location(X, Y, Z) = "($(typeof(X())), $(typeof(Y())), $(typeof(Z())))"
 
 show_location(field::AbstractField{X, Y, Z}) where {X, Y, Z} = show_location(X, Y, Z)
 
-short_show(a) = string(typeof(a))
-shortname(a::Array) = string(typeof(a).name.wrapper)
+short_show(field::Field) = string("Field at ", show_location(field))
 
-show(io::IO, field::Field) =
-    print(io,
-          short_show(field), '\n',
-          "├── data: ", typeof(field.data), '\n',
-          "└── grid: ", typeof(field.grid), '\n',
-          "    ├── size: ", size(field.grid), '\n',
-          "    └── domain: ", show_domain(field.grid), '\n')
-
-short_show(field::AbstractField) = string("Field at ", show_location(field))
+show(io::IO, field::Field{X, Y, Z}) where {X, Y, Z} =
+    print(io, "$(short_show(field))\n",
+          "├── data: $(typeof(field.data))\n",
+          "└── grid: $(typeof(field.grid))\n",
+          "    ├── size: $(size(field.grid))\n",
+          "    └── domain: $(show_domain(field.grid))\n")
