### 2020-02-16 c3fd3ed
The size of fields now depends on both location and grid topology; new x, y, z specific topology functions help field specs
