+++ src/Fields/field.jl
@@ -6,6 +6,7 @@ using OffsetArrays
 
 import Oceananigans.Architectures: architecture
 import Oceananigans.Utils: datatuple
+import Oceananigans.Grids: topology, x_topology, y_topology, z_topology
 
 using Oceananigans.Architectures
 using Oceananigans.Grids
@@ -83,50 +84,91 @@ where each of `X, Y, Z` is `Cell` or `Face`.
 Field(X, Y, Z, arch, grid) =  Field((X, Y, Z), arch, grid)
 
 """
-    CellField([T=eltype(grid)], arch, grid)
+    CellField([FT=eltype(grid)], arch, grid)
 
 Return a `Field{Cell, Cell, Cell}` on architecture `arch` and `grid`.
 Used for tracers and pressure fields.
 """
-CellField(T, arch, grid) = Field{Cell, Cell, Cell}(zeros(T, arch, grid, (Cell, Cell, Cell)), grid)
+CellField(FT, arch, grid) = Field{Cell, Cell, Cell}(zeros(FT, arch, grid, (Cell, Cell, Cell)), grid)
 
 """
-    FaceFieldX([T=eltype(grid)], arch, grid)
+    FaceFieldX([FT=eltype(grid)], arch, grid)
 
 Return a `Field{Face, Cell, Cell}` on architecture `arch` and `grid`.
 Used for the x-velocity field.
 """
-FaceFieldX(T, arch, grid) = Field{Face, Cell, Cell}(zeros(T, arch, grid, (Face, Cell, Cell)), grid)
+FaceFieldX(FT, arch, grid) = Field{Face, Cell, Cell}(zeros(FT, arch, grid, (Face, Cell, Cell)), grid)
 
 """
-    FaceFieldY([T=eltype(grid)], arch, grid)
+    FaceFieldY([FT=eltype(grid)], arch, grid)
 
 Return a `Field{Cell, Face, Cell}` on architecture `arch` and `grid`.
 Used for the y-velocity field.
 """
-FaceFieldY(T, arch, grid) = Field{Cell, Face, Cell}(zeros(T, arch, grid, (Cell, Face, Cell)), grid)
+FaceFieldY(FT, arch, grid) = Field{Cell, Face, Cell}(zeros(FT, arch, grid, (Cell, Face, Cell)), grid)
 
 """
-    FaceFieldZ([T=eltype(grid)], arch, grid)
+    FaceFieldZ([FT=eltype(grid)], arch, grid)
 
 Return a `Field{Cell, Cell, Face}` on architecture `arch` and `grid`.
 Used for the z-velocity field.
 """
-FaceFieldZ(T, arch, grid) = Field{Cell, Cell, Face}(zeros(T, arch, grid, (Cell, Cell, Face)), grid)
+FaceFieldZ(FT, arch, grid) = Field{Cell, Cell, Face}(zeros(FT, arch, grid, (Cell, Cell, Face)), grid)
 
  CellField(arch, grid) = Field((Cell, Cell, Cell), arch, grid)
 FaceFieldX(arch, grid) = Field((Face, Cell, Cell), arch, grid)
 FaceFieldY(arch, grid) = Field((Cell, Face, Cell), arch, grid)
 FaceFieldZ(arch, grid) = Field((Cell, Cell, Face), arch, grid)
 
-location(a) = nothing
-location(::AbstractLocatedField{X, Y, Z}) where {X, Y, Z} = (X, Y, Z)
+#####
+##### Functions for querying fields
+#####
 
+location(a) = nothing
+
+"Returns the location `(X, Y, Z)` of an `AbstractLocatedField{X, Y, Z}`."
[…]
