+++ src/Fields/field.jl
@@ -16,8 +16,8 @@ using Oceananigans.Utils
 """
     AbstractField{X, Y, Z, A, G}
 
-Abstract supertype for fields located at `(X, Y, Z)`, stored on an architecture `A`,
-and defined on a grid `G`.
+Abstract supertype for fields located at `(X, Y, Z)` with data stored in a container
+of type `A`. The field is defined on a grid `G`.
 """
 abstract type AbstractField{X, Y, Z, A, G} end
 
@@ -38,7 +38,9 @@ struct Face end
 """
     Field{X, Y, Z, A, G} <: AbstractField{X, Y, Z, A, G}
 
-A field defined at the location (`X`, `Y`, `Z`) which can be either `Cell` or `Face`.
+A field defined at the location (`X`, `Y`, `Z`), each of which can be either `Cell`
+or `Face`, and with data stored in a container of type `A` (typically an array).
+The field is defined on a grid `G`.
 """
 struct Field{X, Y, Z, A, G} <: AbstractField{X, Y, Z, A, G}
     data :: A
@@ -52,70 +54,63 @@ Adapt.adapt_structure(to, field::Field{X, Y, Z}) where {X, Y, Z} =
     Field{X, Y, Z}(adapt(to, data), field.grid)
 
 """
-	Field(L::Tuple, data::AbstractArray, grid)
+    Field(L::Tuple, arch, grid, [data=zeros(arch, grid)])
 
-Construct a `Field` on `grid` using the array `data` with location defined by the tuple `L`
-of length 3 whose elements are `Cell` or `Face`.
+Construct a `Field` on some architecture `arch` and a `grid` with some `data`.
+The field's location is defined by a tuple `L` of length 3 whose elements are
+`Cell` or `Face`.
 """
-Field(L::Tuple, data::AbstractArray, grid) = Field{L[1], L[2], L[3]}(data, grid)
-
-"""
-    Field(L::Tuple, arch, grid)
-
-Construct a `Field` on architecture `arch` and `grid` at location `L`,
-where `L` is a tuple of `Cell` or `Face` types.
-"""
-Field(L::Tuple, arch, grid) = Field{L[1], L[2], L[3]}(zeros(arch, grid), grid)
+Field(L::Tuple, arch, grid, data=zeros(arch, grid)) = Field{L[1], L[2], L[3]}(data, grid)
 
 """
-    Field(X, Y, Z, arch, grid)
+    Field(X, Y, Z, arch, grid, [data=zeros(arch, grid)])
 
-Construct a `Field` on architecture `arch` and `grid` at location `X`, `Y`, `Z`,
-where each of `X, Y, Z` is `Cell` or `Face`.
+Construct a `Field` on some architecture `arch` and a `grid` with some `data`.
+The field's location is defined by `X`, `Y`, `Z` where each is either `Cell` or `Face`.
 """
-Field(X, Y, Z, arch, grid) =  Field((X, Y, Z), arch, grid)
-
-"""
-    CellField([T=eltype(grid)], arch, grid)
-
-Return a `Field{Cell, Cell, Cell}` on architecture `arch` and `grid`.
-Used for tracers and pressure fields.
-"""
-CellField(T, arch, grid) = Field{Cell, Cell, Cell}(zeros(T, arch, grid), grid)
+Field(X, Y, Z, arch, grid, data=zeros(arch, grid)) =  Field((X, Y, Z), arch, grid, data)
 
 """
-    FaceFieldX([T=eltype(grid)], arch, grid)
+    CellField(arch::AbstractArchitecture, grid, [data=zeros(arch, grid)])
 
[…]
