### 2020-02-16 da78c1d
Regression tests pass as long as as old convention for fields is accounted for
