+++ test/regression_tests/thermal_bubble_regression_test.jl
@@ -51,15 +51,20 @@ function run_thermal_bubble_regression_test(arch)
     Sᶜ = ds["S"][:, :, :, end]
 
     field_names = ["u", "v", "w", "T", "S"]
-    fields = [interior(model.velocities.u), interior(model.velocities.v), interior(model.velocities.w),
-              interior(model.tracers.T), interior(model.tracers.S)]
+
+    test_fields = (interior(model.velocities.u), 
+                   interior(model.velocities.v), 
+                   interior(model.velocities.w)[:, :, 1:Nz],
+                   interior(model.tracers.T), 
+                   interior(model.tracers.S))
+
     correct_fields = [uᶜ, vᶜ, wᶜ, Tᶜ, Sᶜ]
-    summarize_regression_test(field_names, fields, correct_fields)
+    summarize_regression_test(field_names, test_fields, correct_fields)
 
     # Now test that the model state matches the regression output.
     @test all(Array(interior(model.velocities.u)) .≈ uᶜ)
     @test all(Array(interior(model.velocities.v)) .≈ vᶜ)
-    @test all(Array(interior(model.velocities.w)) .≈ wᶜ)
+    @test all(Array(interior(model.velocities.w)[:, :, 1:Nz]) .≈ wᶜ)
     @test all(Array(interior(model.tracers.T))    .≈ Tᶜ)
     @test all(Array(interior(model.tracers.S))    .≈ Sᶜ)
 
