### 2020-02-16 2d3f440
Refactor field tuples so we can pass in data arrays
