+++ src/Fields/show_fields.jl
@@ -10,7 +10,7 @@ short_show(field::Field) = string("Field at ", show_location(field))
 
 show(io::IO, field::Field{X, Y, Z}) where {X, Y, Z} =
     print(io, "$(short_show(field))\n",
-          "├── data: $(typeof(field.data))\n",
+          "├── data: $(typeof(field.data)), size: $(size(field.data))\n",
           "└── grid: $(typeof(field.grid))\n",
           "    ├── size: $(size(field.grid))\n",
-          "    └── domain: $(show_domain(field.grid))\n")
+          "    └── domain: $(show_domain(field.grid))")
