### 2020-02-16 4f689e2
Refactors the xnodes, ynodes, etc functionality for new field convention
