### 2020-02-16 78b17dc
Dispatch insanity to get xnodes, ynodes etc working
