+++ src/Fields/field.jl
@@ -202,9 +202,10 @@ of `f` along `x, y, z`.
                                                                    interior_indices(Z, z_topology(f), f.grid.Nz))
 
 "Returns a reference (not a view) to the interior points of `field.data.parent.`"
-@inline interiorparent(f::Field) = @inbounds f.data.parent[interior_parent_indices(X, x_topology(f), f.grid.Nx, f.grid.Hx),
-                                                           interior_parent_indices(Y, y_topology(f), f.grid.Ny, f.grid.Hy),
-                                                           interior_parent_indices(Z, z_topology(f), f.grid.Nz, f.grid.Hz)]
+@inline interiorparent(f::Field{X, Y, Z}) where {X, Y, Z} = 
+    @inbounds f.data.parent[interior_parent_indices(X, x_topology(f), f.grid.Nx, f.grid.Hx),
+                            interior_parent_indices(Y, y_topology(f), f.grid.Ny, f.grid.Hy),
+                            interior_parent_indices(Z, z_topology(f), f.grid.Nz, f.grid.Hz)]
 
 iterate(f::Field, state=1) = iterate(f.data, state)
 
@@ -222,17 +223,17 @@ iterate(f::Field, state=1) = iterate(f.data, state)
 @inline znode(k, ψ::Field{X, Y, Z}) where {X, Y, Z} = znode(Z, k, ψ.grid)
 
 # Dispatch insanity
-xnodes(::Type{Cell}, topo, grid) = reshape(ψ.grid.xC, ψ.grid.Nx, 1, 1)
-ynodes(::Type{Cell}, topo, grid) = reshape(ψ.grid.yC, 1, ψ.grid.Ny, 1)
-znodes(::Type{Cell}, topo, grid) = reshape(ψ.grid.zC, 1, 1, ψ.grid.Nz)
+xnodes(::Type{Cell}, topo, grid) = reshape(grid.xC, grid.Nx, 1, 1)
+ynodes(::Type{Cell}, topo, grid) = reshape(grid.yC, 1, grid.Ny, 1)
+znodes(::Type{Cell}, topo, grid) = reshape(grid.zC, 1, 1, grid.Nz)
 
-xnodes(::Type{Face}, topo, grid) = reshape(ψ.grid.xF[1:end-1], ψ.grid.Nx, 1, 1)
-ynodes(::Type{Face}, topo, grid) = reshape(ψ.grid.yF[1:end-1], 1, ψ.grid.Ny, 1)
-znodes(::Type{Face}, topo, grid) = reshape(ψ.grid.zF[1:end-1], 1, 1, ψ.grid.Nz)
+xnodes(::Type{Face}, topo, grid) = reshape(grid.xF[1:end-1], grid.Nx, 1, 1)
+ynodes(::Type{Face}, topo, grid) = reshape(grid.yF[1:end-1], 1, grid.Ny, 1)
+znodes(::Type{Face}, topo, grid) = reshape(grid.zF[1:end-1], 1, 1, grid.Nz)
 
-xnodes(::Type{Face}, ::Bounded, grid) = reshape(ψ.grid.xF, ψ.grid.Nx+1, 1, 1)
-ynodes(::Type{Face}, ::Bounded, grid) = reshape(ψ.grid.yF, 1, ψ.grid.Ny+1, 1)
-znodes(::Type{Face}, ::Bounded, grid) = reshape(ψ.grid.zF, 1, 1, ψ.grid.Nz+1)
+xnodes(::Type{Face}, ::Bounded, grid) = reshape(grid.xF, grid.Nx+1, 1, 1)
+ynodes(::Type{Face}, ::Bounded, grid) = reshape(grid.yF, 1, grid.Ny+1, 1)
[…]
