### 2020-02-16 5024b6c
Use a list of required properties to be checkpointed
