### 2020-02-16 d8cabdb
Move tracer field construction to model constructor
