### 2020-02-16 4d49032
Use TendencyFields instead of Tendencies for consistency
