+++ src/Oceananigans.jl
@@ -101,17 +101,24 @@ abstract type AbstractPoissonSolver end
 """
     AbstractDiagnostic
 
-Abstract supertype for types that compute diagnostic information from the current model
-state.
+Abstract supertype for diagnostics that compute information from the current
+model state.
 """
 abstract type AbstractDiagnostic end
 
 """
     AbstractOutputWriter
 
-Abstract supertype for types that perform input and output.
+Abstract supertype for output writers that write data to disk.
 """
 abstract type AbstractOutputWriter end
+
+"""
+    AbstractTimeStepper
+
+Abstract supertype for time steppers.
+"""
+abstract type AbstractTimeStepper end
 
 #####
 ##### Place-holder functions
