+++ src/OutputWriters/OutputWriters.jl
@@ -8,6 +8,7 @@ export
 
 using Oceananigans
 using Oceananigans.Grids
+using Oceananigans.Fields
 using Oceananigans.Architectures
 
 using Oceananigans: AbstractOutputWriter, @hascuda
