### 2020-02-16 3ceb021
Simplify checkpointer and make it work for large models
