+++ src/OutputWriters/checkpointer.jl
@@ -3,14 +3,13 @@
 
 An output writer for checkpointing models to a JLD2 file from which models can be restored.
 """
-mutable struct Checkpointer{I, T, P, A} <: AbstractOutputWriter
+mutable struct Checkpointer{I, T, P} <: AbstractOutputWriter
          frequency :: I
           interval :: T
           previous :: Float64
                dir :: String
             prefix :: String
         properties :: P
-    has_array_refs :: A
              force :: Bool
            verbose :: Bool
 end
@@ -52,15 +51,13 @@ function Checkpointer(model; frequency=nothing, interval=nothing, dir=".",
     validate_interval(frequency, interval)
 
     # Certain properties are required for `restore_from_checkpoint` to work.
-    required_properties = (:grid, :architecture, :velocities, :tracers)
+    required_properties = (:grid, :architecture, :velocities, :tracers, :timestepper)
     for rp in required_properties
         if rp ∉ properties
             @warn "$rp is required for checkpointing. It will be added to checkpointed properties"
             push!(properties, rp)
         end
     end
-
-    has_array_refs = Symbol[]
 
     for p in properties
         p isa Symbol || @error "Property $p to be checkpointed must be a Symbol."
@@ -70,45 +67,31 @@ function Checkpointer(model; frequency=nothing, interval=nothing, dir=".",
             @warn "model.$p contains a function somewhere in its hierarchy and will not be checkpointed."
             filter!(e -> e != p, properties)
         end
-
-        has_reference(AbstractField, getproperty(model, p)) && push!(has_array_refs, p)
     end
 
     mkpath(dir)
-    return Checkpointer(frequency, interval, 0.0, dir, prefix, properties, has_array_refs, force, verbose)
+    return Checkpointer(frequency, interval, 0.0, dir, prefix, properties, force, verbose)
 end
 
 function write_output(model, c::Checkpointer)
     filepath = joinpath(c.dir, c.prefix * "_iteration" * string(model.clock.iteration) * ".jld2")
     c.verbose && @info "Checkpointing to file $filepath..."
 
-    t0 = time_ns()
+    t1 = time_ns()
     jldopen(filepath, "w") do file
         file["checkpointed_properties"] = c.properties
-        file["has_array_refs"] = c.has_array_refs
-
-        serializeproperties!(file, model, filter(e -> !(e ∈ c.has_array_refs), c.properties))
-        saveproperties!(file, model, filter(e -> e ∈ c.has_array_refs, c.properties))
+        serializeproperties!(file, model, c.properties)
     end
 
-    t1, sz = time_ns(), filesize(filepath)
-    c.verbose && @info "Checkpointing done: time=$(prettytime((t1-t0)/1e9)), size=$(pretty_filesize(sz))"
+    t2, sz = time_ns(), filesize(filepath)
+    c.verbose && @info "Checkpointing done: time=$(prettytime((t2-t1)/1e9)), size=$(pretty_filesize(sz))"
 end
 
+# This is the default name used in the simulation.output_writers ordered dict.
 defaultname(::Checkpointer, nelems) = :checkpointer
+
 _arr(::CPU, a) = a
 _arr(::GPU, a) = CuArray(a)
-
-function restore_fields!(model, file, arch, fieldset; location="$fieldset")
-    if fieldset == :timestepper
-        restore_fields!(model.timestepper, file, arch, :Gⁿ; location="timestepper/Gⁿ")
-        restore_fields!(model.timestepper, file, arch, :G⁻; location="timestepper/G⁻")
[…]
