### 2020-02-16 50cad40
Fixing tests for new field convention
