### 2020-02-16 b513abd
More general assumption in horizontal average constructor
