### 2020-02-16 756c050
Gotta reconstruct the offset arrays
