+++ test/test_output_writers.jl
@@ -157,8 +157,8 @@ function run_thermal_bubble_checkpointer_tests(arch)
         time_step!(restored_model, Δt, euler=false)
     end
 
-    rm("checkpoint0.jld2", force=true)
-    rm("checkpoint5.jld2", force=true)
+    rm("checkpoint_iteration0.jld2", force=true)
+    rm("checkpoint_iteration5.jld2", force=true)
 
     # Now the true_model and restored_model should be identical.
     @test all(restored_model.velocities.u.data     .≈ true_model.velocities.u.data)
