### 2020-02-16 ead0439
Fix typo in restoring time stepper from checkpoint
