+++ src/Oceananigans.jl
@@ -15,6 +15,15 @@ export
     Periodic, Bounded, Flat,
     RegularCartesianGrid, VerticallyStretchedCartesianGrid,
 
+    # Boundary conditions
+    BoundaryCondition,
+    Flux, Value, Gradient,
+    FluxBoundaryCondition, ValueBoundaryCondition, GradientBoundaryCondition,
+    CoordinateBoundaryConditions, FieldBoundaryConditions, UVelocityBoundaryConditions,
+    VVelocityBoundaryConditions, WVelocityBoundaryConditions, TracerBoundaryConditions,
+    PressureBoundaryConditions, SolutionBoundaryConditions,
+    BoundaryFunction,
+
     # Fields and field manipulation
     Field, CellField, XFaceField, YFaceField, ZFaceField,
     interior, set!,
@@ -31,15 +40,6 @@ export
 
     # Surface waves via Craik-Leibovich equations
     SurfaceWaves,
-
-    # Boundary conditions
-    BoundaryCondition,
-    Flux, Value, Gradient,
-    FluxBoundaryCondition, ValueBoundaryCondition, GradientBoundaryCondition,
-    CoordinateBoundaryConditions, FieldBoundaryConditions, UVelocityBoundaryConditions,
-    VVelocityBoundaryConditions, WVelocityBoundaryConditions, TracerBoundaryConditions,
-    PressureBoundaryConditions, SolutionBoundaryConditions,
-    BoundaryFunction,
 
     # Time stepping
     time_step!,
@@ -136,13 +136,13 @@ end
 include("Utils/Utils.jl")
 include("Logger.jl")
 include("Grids/Grids.jl")
-include("Fields/Fields.jl")
 include("Operators/Operators.jl")
+include("BoundaryConditions/BoundaryConditions.jl")
+include("Fields/Fields.jl")
 include("Coriolis/Coriolis.jl")
 include("Buoyancy/Buoyancy.jl")
 include("SurfaceWaves.jl")
 include("TurbulenceClosures/TurbulenceClosures.jl")
-include("BoundaryConditions/BoundaryConditions.jl")
 include("Solvers/Solvers.jl")
 include("Forcing/Forcing.jl")
 include("Models/Models.jl")
@@ -159,12 +159,12 @@ include("AbstractOperations/AbstractOperations.jl")
 using .Architectures
 using .Utils
 using .Grids
+using .BoundaryConditions
 using .Fields
 using .Coriolis
 using .Buoyancy
 using .SurfaceWaves
 using .TurbulenceClosures
-using .BoundaryConditions
 using .Solvers
 using .Forcing
 using .Models
