### 2020-02-16 3fa7951
We don't actually use fields in GPU kernels so don't adapt them
