### 2020-02-16 70169bf
Boundary conditions now a field property
