### 2020-02-16 f7db6a9
Better pretty printing for fields
