### 2020-02-16 2b64add
Upgrade field tuple constructors to accept fields as optional kwargs
