### 2020-02-16 31296d0
Version of field tuple constructors that accepts boundary conditions
