+++ src/Fields/field_tuples.jl
@@ -11,6 +11,25 @@ VelocityFields(arch, grid;
 ) = (u=u, v=v, w=w)
 
 """
+    VelocityFields(arch, grid, bcs::NamedTuple)
+
+Return a NamedTuple with fields `u`, `v`, `w` initialized on the architecture `arch`
+and `grid`. Boundary conditions `bcs` may be specified via a named tuple of
+`FieldBoundaryCondition`s.
+"""
+function VelocityFields(arch, grid, bcs::NamedTuple)
+    u_bcs = :u ∈ keys(bcs) ? bcs[:u] : UVelocityBoundaryConditions(grid)
+    v_bcs = :v ∈ keys(bcs) ? bcs[:v] : VVelocityBoundaryConditions(grid)
+    w_bcs = :w ∈ keys(bcs) ? bcs[:w] : WVelocityBoundaryConditions(grid)
+
+    u = XFaceField(arch, grid, u_bcs, zeros(arch, grid))
+    v = YFaceField(arch, grid, v_bcs, zeros(arch, grid))
+    w = ZFaceField(arch, grid, w_bcs, zeros(arch, grid))
+
+    return (u=u, v=v, w=w)
+end
+
+"""
     TracerFields(arch, grid, tracer_names; kwargs...)
 
 Return a NamedTuple with tracer fields specified by `tracer_names` initialized as
@@ -18,19 +37,33 @@ Return a NamedTuple with tracer fields specified by `tracer_names` initialized a
 keyword arguments `kwargs` for each field.
 """
 function TracerFields(arch, grid, tracer_names; kwargs...)
-    for c in keys(kwargs)
-        c ∉ tracer_names && @warn "$c field passed but $c is not in tracer_names."
-    end
-
-    tracer_fields = Tuple(c ∈ keys(kwargs) ?
-                          kwargs[c] :
-                          CellField(arch, grid, TracerBoundaryConditions(grid), zeros(arch, grid))
-                          for c in tracer_names)
+    tracer_fields =
+        Tuple(c ∈ keys(kwargs) ?
+              kwargs[c] :
+              CellField(arch, grid, TracerBoundaryConditions(grid), zeros(arch, grid))
+              for c in tracer_names)
 
     return NamedTuple{tracer_names}(tracer_fields)
 end
 
-TracerFields(arch, grid, ::Union{Tuple{}, Nothing}; kwargs...) = NamedTuple{()}(())
+"""
+    TracerFields(arch, grid, tracer_names, bcs::NamedTuple)
+
+Return a NamedTuple with tracer fields specified by `tracer_names` initialized as
+`CellField`s on the architecture `arch` and `grid`. Boundary conditions `bcs` may
+be specified via a named tuple of `FieldBoundaryCondition`s.
+"""
+function TracerFields(arch, grid, tracer_names, bcs::NamedTuple)
+    tracer_fields =
+        Tuple(c ∈ keys(bcs) ?
+              CellField(arch, grid, bcs[c],                         zeros(arch, grid)) :
+              CellField(arch, grid, TracerBoundaryConditions(grid), zeros(arch, grid))
+              for c in tracer_names)
+
+    return NamedTuple{tracer_names}(tracer_fields)
+end
+
+TracerFields(arch, grid, ::Union{Tuple{}, Nothing}; kwargs...) = NamedTuple()
 TracerFields(arch, grid, tracer::Symbol; kwargs...) = TracerFields(arch, grid, tuple(tracer); kwargs...)
 TracerFields(arch, grid, tracers::NamedTuple; kwargs...) = tracers
 
[…]
