### 2020-02-16 476eeec
Remove BCs from model properties and refactor model constructor
