+++ src/Models/incompressible_model.jl
@@ -7,28 +7,26 @@ using Oceananigans.Buoyancy: validate_buoyancy
 using Oceananigans.TurbulenceClosures: ν₀, κ₀, with_tracers
 
 mutable struct IncompressibleModel{TS, E, A<:AbstractArchitecture, G, T, B, R, SW, U, C, Φ, F,
-                                   BCS, S, K, Θ} <: AbstractModel
-
-           architecture :: A         # Computer `Architecture` on which `Model` is run
-                   grid :: G         # Grid of physical points on which `Model` is solved
-                  clock :: Clock{T}  # Tracks iteration number and simulation time of `Model`
-               buoyancy :: B         # Set of parameters for buoyancy model
-               coriolis :: R         # Set of parameters for the background rotation rate of `Model`
-          surface_waves :: SW        # Set of parameters for surfaces waves via the Craik-Leibovich approximation
-             velocities :: U         # Container for velocity fields `u`, `v`, and `w`
-                tracers :: C         # Container for tracer fields
-              pressures :: Φ         # Container for hydrostatic and nonhydrostatic pressure
-                forcing :: F         # Container for forcing functions defined by the user
-                closure :: E         # Diffusive 'turbulence closure' for all model fields
-    boundary_conditions :: BCS       # Container for 3d bcs on all fields
-            timestepper :: TS        # Object containing timestepper fields and parameters
-        pressure_solver :: S         # Pressure/Poisson solver
-          diffusivities :: K         # Container for turbulent diffusivities
-             parameters :: Θ         # Container for arbitrary user-defined parameters
+                                   S, K, Θ} <: AbstractModel
+       architecture :: A         # Computer `Architecture` on which `Model` is run
+               grid :: G         # Grid of physical points on which `Model` is solved
+              clock :: Clock{T}  # Tracks iteration number and simulation time of `Model`
+           buoyancy :: B         # Set of parameters for buoyancy model
+           coriolis :: R         # Set of parameters for the background rotation rate of `Model`
+      surface_waves :: SW        # Set of parameters for surfaces waves via the Craik-Leibovich approximation
+         velocities :: U         # Container for velocity fields `u`, `v`, and `w`
+            tracers :: C         # Container for tracer fields
+          pressures :: Φ         # Container for hydrostatic and nonhydrostatic pressure
+            forcing :: F         # Container for forcing functions defined by the user
+            closure :: E         # Diffusive 'turbulence closure' for all model fields
+        timestepper :: TS        # Object containing timestepper fields and parameters
+    pressure_solver :: S         # Pressure/Poisson solver
+      diffusivities :: K         # Container for turbulent diffusivities
+         parameters :: Θ         # Container for arbitrary user-defined parameters
 end
 
 """
-    IncompressibleModel(;
+   IncompressibleModel(;
                    grid,
            architecture = CPU(),
              float_type = Float64,
@@ -39,13 +37,15 @@ end
                coriolis = nothing,
           surface_waves = nothing,
                 forcing = ModelForcing(),
-    boundary_conditions = HorizontallyPeriodicSolutionBCs(),
+    boundary_conditions = NamedTuple(),
              parameters = nothing,
-             velocities = VelocityFields(architecture, grid),
-              pressures = PressureFields(architecture, grid),
+             velocities = VelocityFields(architecture, grid, boundary_conditions),
+          tracer_fields = TracerFields(architecture, grid, tracers, boundary_conditions),
+              pressures = PressureFields(architecture, grid, boundary_conditions),
[…]
