### 2020-02-16 d48351e
Nuke solution and model boundary conditions
