+++ src/OutputWriters/output_writer_utils.jl
@@ -1,5 +1,5 @@
 using Oceananigans.Fields: AbstractField
-using Oceananigans.BoundaryConditions: bctype, CoordinateBoundaryConditions, ModelBoundaryConditions
+using Oceananigans.BoundaryConditions: bctype, CoordinateBoundaryConditions
 using Oceananigans.TimeSteppers: AdamsBashforthTimeStepper
 
 #####
@@ -50,7 +50,6 @@ saveproperties!(file, structure, ps) = [saveproperty!(file, "$p", getproperty(st
 # unless they need to be converted (basically CuArrays only).
 
 serializeproperty!(file, location, p) = (file[location] = p)
-serializeproperty!(file, location, p::ModelBoundaryConditions) = (file[location] = p)
 serializeproperty!(file, location, p::Union{AbstractArray, AbstractField}) = saveproperty!(file, location, p)
 serializeproperty!(file, location, p::Union{NamedTuple, AdamsBashforthTimeStepper}) = saveproperty!(file, location, p)
 serializeproperty!(file, location, p::Function) = @warn "Cannot serialize Function property into $location"
