+++ src/TurbulenceClosures/TurbulenceClosures.jl
@@ -13,7 +13,7 @@ export
   RozemaAnisotropicMinimumDissipation,
   VerstappenAnisotropicMinimumDissipation,
 
-  TurbulentDiffusivities,
+  DiffusivityFields,
   calculate_diffusivities!,
 
   ∇_κ_∇c,
@@ -148,8 +148,8 @@ function calculate_diffusivities!(K_tuple::Tuple, arch, grid, closure_tuple::Tup
     return nothing
 end
 
-TurbulentDiffusivities(arch::AbstractArchitecture, grid::AbstractGrid, tracers, closure_tuple::Tuple) =
-    Tuple(TurbulentDiffusivities(arch, grid, tracers, closure) for closure in closure_tuple)
+DiffusivityFields(arch::AbstractArchitecture, grid::AbstractGrid, tracers, closure_tuple::Tuple) =
+    Tuple(DiffusivityFields(arch, grid, tracers, closure) for closure in closure_tuple)
 
 with_tracers(tracers, closure_tuple::Tuple) =
     Tuple(with_tracers(tracers, closure) for closure in closure_tuple)
@@ -158,8 +158,8 @@ with_tracers(tracers, closure_tuple::Tuple) =
 ##### Include module code
 #####
 
-# Fallback constructor for diffusivity types without precomputed diffusivities:
-TurbulentDiffusivities(arch::AbstractArchitecture, grid::AbstractGrid, args...) = nothing
+# Fallback constructor for diffusivity types without precomputed diffusivities.
+DiffusivityFields(arch::AbstractArchitecture, grid::AbstractGrid, args...) = nothing
 
 include("turbulence_closure_utils.jl")
 include("closure_operators.jl")
