+++ src/TurbulenceClosures/turbulence_closure_implementations/rozema_anisotropic_minimum_dissipation.jl
@@ -44,7 +44,7 @@ const Δxᶜᶜᶠ = Δx
 const Δyᶜᶜᶠ = Δy
 const Δzᶜᶜᶠ = Δz
 
-function TurbulentDiffusivities(arch::AbstractArchitecture, grid::AbstractGrid, tracers, ::RAMD)
+function DiffusivityFields(arch::AbstractArchitecture, grid::AbstractGrid, tracers, ::RAMD)
     νₑ = CellField(arch, grid)
     κₑ = TracerFields(arch, grid, tracers)
     return (νₑ=νₑ, κₑ=κₑ)
