+++ src/TurbulenceClosures/turbulence_closure_implementations/verstappen_anisotropic_minimum_dissipation.jl
@@ -49,7 +49,7 @@ Keyword arguments
 
 By default: `C = Cν = Cκ` = 1/12, which is appropriate for a finite-volume method employing a
 second-order advection scheme, `Cb` = 0, which terms off the buoyancy modification term,
-the molecular viscosity of seawater at 20 deg C and 35 psu is used for `ν`, and 
+the molecular viscosity of seawater at 20 deg C and 35 psu is used for `ν`, and
 the molecular diffusivity of heat in seawater at 20 deg C and 35 psu is used for `κ`.
 
 `Cν` or `Cκ` may be constant numbers, or functions of `x, y, z`.
@@ -113,7 +113,7 @@ end
 ##### Constructor
 #####
 
-function TurbulentDiffusivities(arch::AbstractArchitecture, grid::AbstractGrid, tracers, ::VAMD)
+function DiffusivityFields(arch::AbstractArchitecture, grid::AbstractGrid, tracers, ::VAMD)
     νₑ = CellField(arch, grid)
     κₑ = TracerFields(arch, grid, tracers)
     return (νₑ=νₑ, κₑ=κₑ)
