### 2020-02-16 567c469
Clean up `field.jl` a bit
