### 2020-02-16 7a8bd3c
Update DiffusivityFields with ability to control fields and BCs
