+++ src/TurbulenceClosures/turbulence_closure_implementations/verstappen_anisotropic_minimum_dissipation.jl
@@ -109,13 +109,17 @@ function with_tracers(tracers, closure::VerstappenAnisotropicMinimumDissipation{
     return VerstappenAnisotropicMinimumDissipation{FT}(closure.Cν, Cκ, closure.Cb, closure.ν, κ)
 end
 
-#####
-##### Constructor
-#####
+function DiffusivityFields(
+    arch::AbstractArchitecture, grid::AbstractGrid, tracers, ::VAMD;
+    νₑ = CellField(arch, grid, DiffusivityBoundaryConditions(grid), zeros(arch, grid)), kwargs...)
+    κₑ = TracerFields(arch, grid, tracers; kwargs...)
+    return (νₑ=νₑ, κₑ=κₑ)
+end
 
-function DiffusivityFields(arch::AbstractArchitecture, grid::AbstractGrid, tracers, ::VAMD)
-    νₑ = CellField(arch, grid)
-    κₑ = TracerFields(arch, grid, tracers)
+function DiffusivityFields(arch::AbstractArchitecture, grid::AbstractGrid, tracers, bcs::NamedTuple, ::VAMD)
+    νₑ_bcs = :νₑ ∈ keys(bcs) ? bcs[:νₑ] : DiffusivityBoundaryConditions(grid)
+    νₑ = CellField(arch, grid, νₑ_bcs, zeros(arch, grid))
+    κₑ = TracerFields(arch, grid, tracers, bcs)
     return (νₑ=νₑ, κₑ=κₑ)
 end
 
