### 2020-02-16 8842bd0
Separate `diffusivity_fields.jl` file to reduce code repetition
