### 2020-02-16 f8d3b04
Need to split boundary condition tests into two now
