+++ test/test_time_stepping_bcs.jl
@@ -1,0 +1,179 @@
+function test_z_boundary_condition_simple(arch, FT, fldname, bctype, bc, Nx, Ny)
+    Nz = 16
+    grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(0.1, 0.2, 0.3))
+
+    bc = BoundaryCondition(bctype, bc)
+    fieldbcs = TracerBoundaryConditions(grid, top=bc)
+    modelbcs = SolutionBoundaryConditions(grid; Dict(fldname=>fieldbcs)...)
+
+    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT, boundary_conditions=modelbcs)
+
+    time_step!(model, 1e-16, euler=true)
+
+    return model isa IncompressibleModel
+end
+
+function test_z_boundary_condition_top_bottom_alias(arch, FT, fldname)
+    N, val = 16, 1.0
+    grid = RegularCartesianGrid(FT, size=(N, N, N), length=(0.1, 0.2, 0.3))
+
+    top_bc    = BoundaryCondition(Value,  val)
+    bottom_bc = BoundaryCondition(Value, -val)
+    fieldbcs = TracerBoundaryConditions(grid, top=top_bc, bottom=bottom_bc)
+    modelbcs = SolutionBoundaryConditions(grid; Dict(fldname=>fieldbcs)...)
+
+    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT, boundary_conditions=modelbcs)
+
+    bcs = getfield(model.boundary_conditions.solution, fldname)
+
+    time_step!(model, 1e-16, euler=true)
+
+    return getbc(bcs.z.top) == val && getbc(bcs.z.bottom) == -val
+end
+
+function test_z_boundary_condition_array(arch, FT, fldname)
+    Nx = Ny = Nz = 16
+
+    bcarray = rand(FT, Nx, Ny)
+
+    if arch == GPU()
+        bcarray = CuArray(bcarray)
+    end
+
+    grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(0.1, 0.2, 0.3))
+
+    value_bc = BoundaryCondition(Value, bcarray)
+    fieldbcs = TracerBoundaryConditions(grid, top=value_bc)
+    modelbcs = SolutionBoundaryConditions(grid; Dict(fldname=>fieldbcs)...)
+
+    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT, boundary_conditions=modelbcs)
+
+    bcs = getfield(model.boundary_conditions.solution, fldname)
+
+    time_step!(model, 1e-16, euler=true)
+
+    return bcs.z.top[1, 2] == bcarray[1, 2]
+end
+
+function test_flux_budget(arch, FT, fldname)
+    N, κ, Lz = 16, 1, 0.7
+    grid = RegularCartesianGrid(FT, size=(N, N, N), length=(1, 1, Lz))
+
+    bottom_flux = FT(0.3)
+    flux_bc = BoundaryCondition(Flux, bottom_flux)
+    fieldbcs = TracerBoundaryConditions(grid, bottom=flux_bc)
+    modelbcs = SolutionBoundaryConditions(grid; Dict(fldname=>fieldbcs)...)
+
+    closure = ConstantIsotropicDiffusivity(FT, ν=κ, κ=κ)
+    model = IncompressibleModel(grid=grid, closure=closure, architecture=arch,
+                                float_type=FT, buoyancy=nothing, boundary_conditions=modelbcs)
+
+    if fldname ∈ (:u, :v, :w)
+        field = getfield(model.velocities, fldname)
+    else
[…]
