### 2020-02-16 949a79d
Simplify halo filling since fields store their own boundary conditions
