### 2020-02-16 2a406d1
Further simplify halo filling as field also stores its own grid
