+++ src/TimeSteppers/TimeSteppers.jl
@@ -62,24 +62,35 @@ boundary_condition_function_arguments(model) =
 #####
 
 """
+    calculate_explicit_substep!(tendencies, velocities, tracers, pressures, diffusivities, model)
+
+Calculate the initial and explicit substep of the two-step fractional step method with pressure correction.
+"""
+function calculate_explicit_substep!(tendencies, velocities, tracers, pressures, diffusivities, model)
+    time_step_precomputations!(diffusivities, pressures, velocities, tracers, model)
+    calculate_tendencies!(tendencies, velocities, tracers, pressures, diffusivities, model)
+    return nothing
+end
+
+"""
     time_step_precomputations!(diffusivities, pressures, velocities, tracers, model)
 
 Perform precomputations necessary for an explicit timestep or substep.
 """
 function time_step_precomputations!(diffusivities, pressures, velocities, tracers, model)
 
-    fill_halo_regions!(merge(velocities, tracers), model.boundary_conditions.solution, model.architecture,
-                       model.grid, boundary_condition_function_arguments(model)...)
+    fill_halo_regions!(merge(model.velocities, model.tracers), model.architecture,
+                       boundary_condition_function_arguments(model)...)
 
-    calculate_diffusivities!(diffusivities, model.architecture, model.grid, model.closure, model.buoyancy,
-                             velocities, tracers)
+    calculate_diffusivities!(diffusivities, model.architecture, model.grid, model.closure,
+                             model.buoyancy, velocities, tracers)
 
-    fill_halo_regions!(diffusivities, model.boundary_conditions.diffusivities, model.architecture, model.grid)
+    fill_halo_regions!(model.diffusivities, model.architecture)
 
     @launch(device(model.architecture), config=launch_config(model.grid, :xy),
             update_hydrostatic_pressure!(pressures.pHY′, model.grid, model.buoyancy, tracers))
 
-    fill_halo_regions!(pressures.pHY′, model.boundary_conditions.pressure, model.architecture, model.grid)
+    fill_halo_regions!(model.pressures.pHY′, model.architecture)
 
     return nothing
 end
@@ -92,12 +103,16 @@ contribution from non-hydrostatic pressure.
 """
 function calculate_tendencies!(tendencies, velocities, tracers, pressures, diffusivities, model)
 
-    calculate_interior_source_terms!(tendencies, model.architecture, model.grid, model.coriolis, model.buoyancy,
-                                     model.surface_waves, model.closure, velocities, tracers, pressures.pHY′,
-                                     diffusivities, model.forcing, model.parameters, model.clock.time)
+    calculate_interior_source_terms!(
+        tendencies, model.architecture, model.grid, model.coriolis, model.buoyancy,
+        model.surface_waves, model.closure, velocities, tracers, pressures.pHY′,
+        diffusivities, model.forcing, model.parameters, model.clock.time
+    )
 
-    calculate_boundary_source_terms!(tendencies, model.boundary_conditions.solution, model.architecture,
-                                     model.grid, boundary_condition_function_arguments(model)...)
+    calculate_boundary_source_terms!(
+        model.timestepper.Gⁿ, model.architecture,
+        model.grid, boundary_condition_function_arguments(model)...
+    )
 
     return nothing
 end
@@ -108,36 +123,19 @@ end
 Calculate the (nonhydrostatic) pressure correction associated `tendencies`, `velocities`, and step size `Δt`.
 """
 function calculate_pressure_correction!(nonhydrostatic_pressure, Δt, tendencies, velocities, model)
-    velocity_tendencies = (u=tendencies.u, v=tendencies.v, w=tendencies.w)
[…]
