+++ src/BoundaryConditions/apply_flux_bcs.jl
@@ -7,39 +7,34 @@ using Oceananigans.Utils: @loop_xy, @loop_xz, launch_config
 ##### Inhomogeneous Value and Gradient boundary conditions are handled by filling halos.
 #####
 
-# Avoid some computation / memory accesses for Value, Gradient, Periodic, NoPenetration,
-# and no-flux boundary conditions --- every boundary condition that does *not* prescribe
-# a non-trivial flux.
-const NotFluxBC = Union{VBC, GBC, PBC, NPBC, NFBC}
-apply_z_bcs!(Gc, arch, grid, ::NotFluxBC, ::NotFluxBC, args...) = nothing
-apply_y_bcs!(Gc, arch, grid, ::NotFluxBC, ::NotFluxBC, args...) = nothing
-
 """
-    apply_z_bcs!(Gc, arch, grid, bottom_bc, top_bc, args...)
+    apply_z_bcs!(Gc, arch, grid, args...)
 
 Apply flux boundary conditions to a field `c` by adding the associated flux divergence to
 the source term `Gc` at the top and bottom.
 """
-function apply_z_bcs!(Gc, arch, grid, bottom_bc, top_bc, args...)
+function apply_z_bcs!(Gc, arch, grid, args...)
     @launch(device(arch), config=launch_config(grid, :xy),
-            _apply_z_bcs!(Gc, grid, bottom_bc, top_bc, args...))
+            _apply_z_bcs!(Gc, grid, Gc.boundary_conditions.z.left,
+                          Gc.boundary_conditions.z.right, args...))
     return nothing
 end
 
 """
-    apply_y_bcs!(Gc, arch, grid, left_bc, right_bc, args...)
+    apply_y_bcs!(Gc, arch, grid, args...)
 
 Apply flux boundary conditions to a field `c` by adding the associated flux divergence to
 the source term `Gc` at the left and right.
 """
-function apply_y_bcs!(Gc, arch, grid, left_bc, right_bc, args...)
+function apply_y_bcs!(Gc, arch, grid, args...)
     @launch(device(arch), config=launch_config(grid, :xz),
-            _apply_y_bcs!(Gc, grid, left_bc, right_bc, args...))
+            _apply_y_bcs!(Gc, grid, Gc.boundary_conditions.y.left,
+                          Gc.boundary_conditions.y.right, args...))
     return nothing
 end
 
 """
-    _apply_z_bcs!(Gc, grid, top_bc, bottom_bc, args...)
+    _apply_z_bcs!(Gc, grid, bottom_bc, top_bc, args...)
 
 Apply a top and/or bottom boundary condition to variable `c`.
 """
