+++ src/Fields/Fields.jl
@@ -15,5 +15,6 @@ include("field.jl")
 include("set!.jl")
 include("field_tuples.jl")
 include("show_fields.jl")
+include("boundary_function.jl")
 
 end
