### 2020-02-17 2a4724b
Get rid of redundant grid argument
