### 2020-02-17 1439e81
Gotta pass both the field and tendency to apply flux BCs
