### 2020-02-17 3ad92b0
Properly set diffusivity boundary conditions from model constructor
