+++ test/test_time_stepping_bcs.jl
@@ -3,10 +3,11 @@ function test_z_boundary_condition_simple(arch, FT, fldname, bctype, bc, Nx, Ny)
     grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(0.1, 0.2, 0.3))
 
     bc = BoundaryCondition(bctype, bc)
-    fieldbcs = TracerBoundaryConditions(grid, top=bc)
-    modelbcs = SolutionBoundaryConditions(grid; Dict(fldname=>fieldbcs)...)
+    field_bcs = TracerBoundaryConditions(grid, top=bc)
+    model_bcs = NamedTuple{(fldname,)}((field_bcs,))
 
-    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT, boundary_conditions=modelbcs)
+    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT,
+                                boundary_conditions=model_bcs)
 
     time_step!(model, 1e-16, euler=true)
 
@@ -19,15 +20,21 @@ function test_z_boundary_condition_top_bottom_alias(arch, FT, fldname)
 
     top_bc    = BoundaryCondition(Value,  val)
     bottom_bc = BoundaryCondition(Value, -val)
-    fieldbcs = TracerBoundaryConditions(grid, top=top_bc, bottom=bottom_bc)
-    modelbcs = SolutionBoundaryConditions(grid; Dict(fldname=>fieldbcs)...)
+    field_bcs = TracerBoundaryConditions(grid, top=top_bc, bottom=bottom_bc)
+    model_bcs = NamedTuple{(fldname,)}((field_bcs,))
 
-    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT, boundary_conditions=modelbcs)
-
-    bcs = getfield(model.boundary_conditions.solution, fldname)
+    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT,
+                                boundary_conditions=model_bcs)
 
     time_step!(model, 1e-16, euler=true)
 
+    if fldname ∈ (:u, :v, :w)
+        field = getfield(model.velocities, fldname)
+    else
+        field = getfield(model.tracers, fldname)
+    end
+
+    bcs = field.boundary_conditions
     return getbc(bcs.z.top) == val && getbc(bcs.z.bottom) == -val
 end
 
@@ -43,15 +50,21 @@ function test_z_boundary_condition_array(arch, FT, fldname)
     grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(0.1, 0.2, 0.3))
 
     value_bc = BoundaryCondition(Value, bcarray)
-    fieldbcs = TracerBoundaryConditions(grid, top=value_bc)
-    modelbcs = SolutionBoundaryConditions(grid; Dict(fldname=>fieldbcs)...)
+    field_bcs = TracerBoundaryConditions(grid, top=value_bc)
+    model_bcs = NamedTuple{(fldname,)}((field_bcs,))
 
-    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT, boundary_conditions=modelbcs)
-
-    bcs = getfield(model.boundary_conditions.solution, fldname)
+    model = IncompressibleModel(grid=grid, architecture=arch, float_type=FT,
+                                boundary_conditions=model_bcs)
 
     time_step!(model, 1e-16, euler=true)
 
+    if fldname ∈ (:u, :v, :w)
+        field = getfield(model.velocities, fldname)
+    else
+        field = getfield(model.tracers, fldname)
+    end
+
+    bcs = field.boundary_conditions
[…]
