### 2020-02-17 11c9cea
Get tuples of closures working again
