+++ src/BoundaryConditions/fill_halo_regions.jl
@@ -10,7 +10,7 @@ fill_halo_regions!(::Nothing, args...) = nothing
 Fill halo regions for each field in the tuple `fields` according to their boundary
 conditions, possibly recursing into `fields` if it is a nested tuple-of-tuples.
 """
-function fill_halo_regions!(fields::NamedTuple, arch, args...)
+function fill_halo_regions!(fields::Union{Tuple,NamedTuple}, arch, args...)
     for field in fields
         fill_halo_regions!(field, arch, args...)
     end
