+++ test/test_turbulence_closures.jl
@@ -22,7 +22,6 @@ function test_constant_isotropic_diffusivity_fluxdiv(FT=Float64; ν=FT(0.3), κ=
           arch = CPU()
        closure = ConstantIsotropicDiffusivity(FT, κ=(T=κ, S=κ), ν=ν)
           grid = RegularCartesianGrid(FT, size=(3, 1, 4), length=(3, 1, 4))
-           bcs = SolutionBoundaryConditions((:T, :S), SolutionBoundaryConditions(grid))
     velocities = VelocityFields(arch, grid)
        tracers = TracerFields(arch, grid, (:T, :S))
 
@@ -36,8 +35,9 @@ function test_constant_isotropic_diffusivity_fluxdiv(FT=Float64; ν=FT(0.3), κ=
         interior(T)[:, 1, k] .= [0, -1, 0]
     end
 
+    fill_halo_regions!(merge(velocities, tracers), arch)
+
     U, C = datatuples(velocities, tracers)
-    fill_halo_regions!(merge(U, C), bcs, arch, grid)
 
     return (   ∇_κ_∇c(2, 1, 3, grid, closure, C.T, Val(1)) == 2κ &&
             ∂ⱼ_2ν_Σ₁ⱼ(2, 1, 3, grid, closure, U) == 2ν &&
@@ -49,8 +49,8 @@ function test_anisotropic_diffusivity_fluxdiv(FT=Float64; νh=FT(0.3), κh=FT(0.
           arch = CPU()
        closure = ConstantAnisotropicDiffusivity(FT, νh=νh, νv=νv, κh=(T=κh, S=κh), κv=(T=κv, S=κv))
           grid = RegularCartesianGrid(FT, size=(3, 1, 4), length=(3, 1, 4))
-           bcs = SolutionBoundaryConditions((:T, :S), SolutionBoundaryConditions(grid))
-      buoyancy = SeawaterBuoyancy(FT, gravitational_acceleration=1, equation_of_state=LinearEquationOfState(FT))
+           eos = LinearEquationOfState(FT)
+      buoyancy = SeawaterBuoyancy(FT, gravitational_acceleration=1, equation_of_state=eos)
     velocities = VelocityFields(arch, grid)
        tracers = TracerFields(arch, grid, (:T, :S))
 
@@ -72,8 +72,9 @@ function test_anisotropic_diffusivity_fluxdiv(FT=Float64; νh=FT(0.3), κh=FT(0.
     interior(T)[:, 1, 3] .= [0, -4, 0]
     interior(T)[:, 1, 4] .= [0,  1, 0]
 
+    fill_halo_regions!(merge(velocities, tracers), arch)
+
     U, C = datatuples(velocities, tracers)
-    fill_halo_regions!(merge(U, C), bcs, arch, grid)
 
     return (   ∇_κ_∇c(2, 1, 3, grid, closure, C.T, Val(1)) == 8κh + 10κv &&
             ∂ⱼ_2ν_Σ₁ⱼ(2, 1, 3, grid, closure, U) == 2νh + 4νv &&
@@ -86,13 +87,12 @@ function test_calculate_diffusivities(arch, closurename, FT=Float64; kwargs...)
           closure = getproperty(TurbulenceClosures, closurename)(FT, kwargs...)
           closure = with_tracers(tracernames, closure)
              grid = RegularCartesianGrid(FT, size=(3, 3, 3), length=(3, 3, 3))
-    diffusivities = TurbulentDiffusivities(arch, grid, tracernames, closure)
+    diffusivities = DiffusivityFields(arch, grid, tracernames, closure)
          buoyancy = BuoyancyTracer()
        velocities = VelocityFields(arch, grid)
           tracers = TracerFields(arch, grid, tracernames)
 
[…]
