+++ verification/stratified_couette_flow/stratified_couette_flow.jl
@@ -118,7 +118,7 @@ function simulate_stratified_couette_flow(; Nxy, Nz, arch=GPU(), h=1, U_wall=1,
                        grid = grid,
                    buoyancy = SeawaterBuoyancy(equation_of_state=LinearEquationOfState(α=1.0, β=0.0)),
                     closure = AnisotropicMinimumDissipation(ν=ν, κ=κ),
-        boundary_conditions = SolutionBoundaryConditions(grid, u=ubcs, v=vbcs, T=Tbcs),
+        boundary_conditions = (u=ubcs, v=vbcs, T=Tbcs),
                  parameters = parameters
     )
 
