### 2020-02-17 b8f68e9
Start updating checkpointer
