### 2020-02-17 1ad7bbe
Temporary hack to calculate potential energy in halo regions
