### 2020-02-17 de78002
Make use of `model.thermodynamic_fields` in time stepping
