+++ test/runtests.jl
@@ -3,8 +3,6 @@ using Test
 using Oceananigans
 using JULES
 
-using JULES: VaporPlaceholder, VaporLiquidIcePlaceholder
-
 @testset "JULES" begin
     include("test_models.jl")
     include("test_time_stepping.jl")
