### 2020-02-17 04afea2
Update  3-gas dry rising thermal bubble with energy verification
