### 2020-02-18 75499db
Allow for `DateTime` clocks
