### 2020-02-19 9187570
Check that field data matches grid size with halos
