### 2020-02-19 7fd5d60
Add test for setting fields with Arrays and CuArrays
