+++ test/test_fields.jl
@@ -57,6 +57,14 @@ end
             for fieldtype in fieldtypes, val in vals
                 @test correct_field_value_was_set(arch, grid, fieldtype, val)
             end
+
+	    for fieldtype in fieldtypes
+		field = fieldtype(arch, grid)
+		A = rand(FT, N...)
+		arch isa GPU && (A = CuArray(A))
+		set!(field, A)
+		@test field.data[2, 4, 6] == A[2, 4, 6]
+	    end
         end
     end
 end
