### 2020-02-19 7d26516
Convert CRLF line endings to LF in test_fields.jl
