+++ test/test_fields.jl
@@ -58,13 +58,13 @@ end
                 @test correct_field_value_was_set(arch, grid, fieldtype, val)
             end
 
-	    for fieldtype in fieldtypes
-		field = fieldtype(arch, grid)
-		A = rand(FT, N...)
-		arch isa GPU && (A = CuArray(A))
-		set!(field, A)
-		@test field.data[2, 4, 6] == A[2, 4, 6]
-	    end
+            for fieldtype in fieldtypes
+                field = fieldtype(arch, grid)
+                A = rand(FT, N...)
+                arch isa GPU && (A = CuArray(A))
+                set!(field, A)
+                @test field.data[2, 4, 6] == A[2, 4, 6]
+            end
         end
     end
 end
