### 2020-02-20 537c626
Give checkpointing its own section in model setup docs
