### 2020-02-20 48e5424
Update diagnostic and output writer model setup docs to use Simulations
