+++ docs/src/model_setup/output_writers.md
@@ -3,9 +3,9 @@ Saving model data to disk can be done in a flexible manner using output writers.
 implemented are a NetCDF output writer (relying on [NCDatasets.jl](https://github.com/Alexander-Barth/NCDatasets.jl))
 and a JLD2 output writer (relying on [JLD2.jl](https://github.com/JuliaIO/JLD2.jl)).
 
-Output writers are stored as a list of output writers in `model.output_writers`. Output writers can be specified at
-model creation time or be specified at any later time and appended (or assigned with a key value pair) to
-`model.output_writers`.
+Output writers are stored as a list of output writers in `simulation.output_writers`. Output writers can be specified
+at model creation time or be specified at any later time and appended (or assigned with a key value pair) to
+`simulation.output_writers`.
 
 ## NetCDF output writer
 Model data can be saved to NetCDF files along with associated metadata. The NetCDF output writer is generally used by
@@ -16,6 +16,7 @@ slices) along with output attributes
 ```@example
 Nx = Ny = Nz = 16
 model = Model(grid=RegularCartesianGrid(size=(Nx, Ny, Nz), length=(1, 1, 1)))
+simulation = Simulation(model, Δt=12, stop_time=3600)
 
 fields = Dict(
     "u" => model.velocities.u,
@@ -27,12 +28,14 @@ output_attributes = Dict(
     "T" => Dict("longname" => "Temperature", "units" => "C")
 )
 
-model.output_writers[:field_writer] = NetCDFOutputWriter(model, fields; filename="output_fields.nc",
-                                                         interval=6hour, output_attributes=output_attributes)
+simulation.output_writers[:field_writer] =
+    NetCDFOutputWriter(model, fields; filename="output_fields.nc",
+                       interval=6hour, output_attributes=output_attributes)
 
-model.output_writers[:surface_slice_writer] = NetCDFOutputWriter(model, fields; filename="output_surface_xy_slice.nc",
-                                                                 interval=5minute, output_attributes=output_attributes,
-                                                                 zC=Nz, zF=Nz)
+simulation.output_writers[:surface_slice_writer] =
+    NetCDFOutputWriter(model, fields; filename="output_surface_xy_slice.nc",
+                       interval=5minute, output_attributes=output_attributes,
+                       zC=Nz, zF=Nz)
 ```
 
 See [`NetCDFOutputWriter`](@ref) for more details and options.
@@ -47,6 +50,7 @@ of the function will be saved to the JLD2 file. For example, to write out 3D fie
 of T every 1 hour of simulation time to a file called `some_data.jld2`
 ```@example
 model = Model(grid=RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1)))
+simulation = Simulation(model, Δt=12, stop_time=3600)
 
 function init_save_some_metadata(file, model)
     file["author"] = "Chim Riggles"
@@ -62,9 +66,9 @@ outputs = Dict(
     :T_avg => model -> T_avg(model)
 )
 
-jld2_writer = JLD2OutputWriter(model, outputs; init=init_save_some_metadata, interval=1hour, prefix="some_data")
+jld2_writer = JLD2OutputWriter(model, outputs, init=init_save_some_metadata, interval=1hour, prefix="some_data")
 
-push!(model.output_writers, jld2_writer)
+push!(simulation.output_writers, jld2_writer)
 ```
 
 See [`JLD2OutputWriter`](@ref) for more details and options.
