### 2020-02-20 bb2517d
Update packages and only support Julia 1.3+
