+++ Project.toml
@@ -34,7 +34,7 @@ JLD2 = "^0.1.2, ^1"
 NCDatasets = "^0.9.3, ^1"
 OffsetArrays = "^0.11.1, ^1"
 OrderedCollections = "^1.1"
-julia = "^1.1"
+julia = "^1.3"
 
 [extras]
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
