### 2020-02-20 249f659
Update README and docs to inform users of Julia 1.3+ requirement
