+++ README.md
@@ -70,16 +70,16 @@ Our goal is to develop a friendly and intuitive package allowing users to focus 
 You can install the latest version of Oceananigans using the built-in package manager (accessed by pressing `]` in the Julia command prompt) to add the package and instantiate/build all depdendencies
 ```julia
 julia>]
-(v1.1) pkg> add Oceananigans
-(v1.1) pkg> instantiate
+(v1.3) pkg> add Oceananigans
+(v1.3) pkg> instantiate
 ```
 We recommend installing Oceananigans with the built-in Julia package manager, because this installs a stable, tagged release. Oceananigans.jl can be updated to the latest tagged release from the package manager by typing
 ```julia
-(v1.1) pkg> update Oceananigans
+(v1.3) pkg> update Oceananigans
 ```
 At this time, updating should be done with care, as Oceananigans is under rapid development and breaking changes to the user API occur often. But if anything does happen, please open an issue!
 
-**Note**: Oceananigans requires at least Julia v1.1 to run correctly.
+**Note**: Oceananigans requires at least Julia v1.3 to run. Installing Oceananigans with an older version of Julia will install an older version of Oceananigans (the latest version compatible with your version of Julia).
 
 ## Running your first model
 Let's initialize a 3D model with 100×100×50 grid points on a 2×2×1 km domain and simulate it for 1 time step of 60 seconds.
