### 2020-02-20 5f52aa4
Update simple examples in README
