### 2020-02-20 b42af96
GitLab CI should now run tests using Julia 1.3
