### 2020-02-21 8a2c69b
Add docstring example for TracerFields
