### 2020-02-21 0001584
Simplify TendencyFields constructor
