### 2020-02-21 2d5cca4
Revert "Draft some lazy field abstractions"
This reverts commit f37da129ee02e6165af0d3bb6c76a78f0d82f50b.
