### 2020-02-21 c3dda10
Benchmark both energy and entropy models
