+++ verification/three_gas_dry_rising_thermal_bubble_entropy/three_gas_dry_rising_thermal_bubble_entropy.jl
@@ -32,7 +32,7 @@ grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
 model = CompressibleModel(
                       grid = grid,
                  densities = DryEarth3(),
-    thermodynamic_variable = PrognosticS(),
+    thermodynamic_variable = Entropy(),
                    closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
 )
 
