+++ verification/dry_rising_thermal_bubble_energy/dry_rising_thermal_bubble_energy.jl
@@ -31,7 +31,7 @@ grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
 model = CompressibleModel(
                       grid = grid,
                  densities = DryEarth(),
-    thermodynamic_variable = PrognosticE(),
+    thermodynamic_variable = Energy(),
                    closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
 )
 
