+++ src/time_stepping.jl
@@ -38,7 +38,7 @@ function time_step!(model::CompressibleModel; Δt, Nt=1)
     time = model.clock.time
     coriolis = model.coriolis
     closure = model.closure
-    tvar = values(model.thermodynamic_variable)[1]
+    tvar = model.thermodynamic_variable
     microphysics = model.microphysics
     forcing = model.forcing
     params = model.parameters
