### 2020-02-21 c90e714
Update tests and run performance benchmark at the end
