+++ test/runtests.jl
@@ -3,9 +3,10 @@ using Test
 using Oceananigans
 using JULES
 
-using JULES: VaporPlaceholder, VaporLiquidIcePlaceholder
-
 @testset "JULES" begin
     include("test_models.jl")
     include("test_time_stepping.jl")
 end
+
+include("../benchmarks/benchmark_static_atmosphere.jl")
+println()
