+++ src/Operators/Operators.jl
@@ -1,6 +1,10 @@
 module Operators
 
 using Oceananigans
+using Oceananigans.Grids
+using Oceananigans.Operators
+
+using Oceananigans: AbstractGrid
 
 export
     x_f_cross_U, y_f_cross_U, z_f_cross_U,
