+++ src/Operators/compressible_operators.jl
@@ -1,6 +1,3 @@
-using Oceananigans.Operators
-using Oceananigans: AbstractGrid
-
 #####
 ##### Convert conservative variables to primitive variables
 #####
@@ -13,11 +10,16 @@ using Oceananigans: AbstractGrid
 #####
 ##### Kinetic energy
 #####
-@inline kinetic_energy(i, j, k, grid, Ũ, ρ) = @inbounds 0.5*((ℑxᶜᵃᵃ(i, j, k, grid, Ũ.ρu) / ρ[i, j, k])^2 + (ℑyᵃᶜᵃ(i, j, k, grid, Ũ.ρv) / ρ[i, j, k])^2 + (ℑzᵃᵃᶜ(i, j, k, grid, Ũ.ρw) / ρ[i, j, k])^2)
+
+@inline kinetic_energy(i, j, k, grid::AbstractGrid{FT}, Ũ, ρ) where FT =
+    @inbounds FT(0.5) * ((ℑxᶜᵃᵃ(i, j, k, grid, Ũ.ρu) / ρ[i, j, k])^2 +
+                         (ℑyᵃᶜᵃ(i, j, k, grid, Ũ.ρv) / ρ[i, j, k])^2 +
+                         (ℑzᵃᵃᶜ(i, j, k, grid, Ũ.ρw) / ρ[i, j, k])^2)
 
 #####
 ##### Pressure work
 #####
+
 @inline pu(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) = ℑxᶠᵃᵃ(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) * Axᵃᵃᶠ(i, j, k, grid) * U_over_ρ(i, j, k, grid, Ũ.ρu, ρ)
 @inline pv(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) = ℑyᵃᶠᵃ(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) * Ayᵃᵃᶠ(i, j, k, grid) * V_over_ρ(i, j, k, grid, Ũ.ρv, ρ)
 @inline pw(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) = ℑzᵃᵃᶠ(i, j, k, grid, diagnose_p, tvar, g, Ũ, ρ, ρ̃, C̃) * Azᵃᵃᵃ(i, j, k, grid) * W_over_ρ(i, j, k, grid, Ũ.ρw, ρ)
@@ -52,16 +54,17 @@ end
 
 #####
 ##### Diffusion
-##### Note: need a better way to enforce boundary conditions on diffusive fluxes
-##### at rigid boundaries
+##### FIXME: need a better way to enforce boundary conditions on diffusive fluxes at rigid boundaries
 #####
 
 @inline diffusive_flux_x(i, j, k, grid, κᶠᶜᶜ, ρ, C) = ℑxᶠᵃᵃ(i, j, k, grid, ρ) * κᶠᶜᶜ * Axᵃᵃᶠ(i, j, k, grid) * ∂xᶠᵃᵃ(i, j, k, grid, C_over_ρ, C, ρ)
[…]
