+++ test/test_time_stepping.jl
@@ -5,7 +5,7 @@
         @info "  Testing time stepping with energy thermodynamic variable..."
 
         grid = RegularCartesianGrid(size=(16, 16, 16), halo=(2, 2, 2), length=(1, 1, 1))
-        model = CompressibleModel(grid = grid, densities = DryEarth(),
+        model = CompressibleModel(grid = grid, gases = DryEarth(),
                                   thermodynamic_variable = Energy())
         time_step!(model, Δt=1)
         @test model isa CompressibleModel
@@ -15,7 +15,7 @@
         @info "  Testing time stepping with entropy thermodynamic variable..."
 
         grid = RegularCartesianGrid(size=(16, 16, 16), halo=(2, 2, 2), length=(1, 1, 1))
-        model = CompressibleModel(grid = grid, densities = DryEarth(),
+        model = CompressibleModel(grid = grid, gases = DryEarth(),
                                   thermodynamic_variable = Entropy())
         time_step!(model, Δt=1)
         @test model isa CompressibleModel
