### 2020-02-21 e0d11d8
Get rid of model.parameters
