+++ src/time_stepping.jl
@@ -41,7 +41,6 @@ function time_step!(model::CompressibleModel; Δt, Nt=1)
     tvar = model.thermodynamic_variable
     microphysics = model.microphysics
     forcing = model.forcing
-    params = model.parameters
 
     Ũ  = model.momenta
     ρ = model.total_density
@@ -72,7 +71,7 @@ function time_step!(model::CompressibleModel; Δt, Nt=1)
         fill_halo_regions!(ρ.data, hpbcs, arch, grid)
         fill_halo_regions!(datatuple(merge(Ũ, C̃)), hpbcs, arch, grid)
         fill_halo_regions!(Ũ.ρw.data, hpbcs_np, arch, grid)
-        compute_slow_forcings!(F, grid, tvar, g, coriolis, closure, Ũ, ρ, ρ̃, C̃, K̃, forcing, time, params)
+        compute_slow_forcings!(F, grid, tvar, g, coriolis, closure, Ũ, ρ, ρ̃, C̃, K̃, forcing, time)
         fill_halo_regions!(F.ρw.data, hpbcs_np, arch, grid)
 
         # RK3 time-stepping
