### 2020-02-21 e56c79f
Revert "Try JULIA_NUM_THREADS=4 with FFTW patch"
This reverts commit 50522e4faab5f669023c330886fc792879e3f333.
