+++ .gitlab-ci.yml
@@ -1,6 +1,6 @@
 variables:
   CI_IMAGE_TAG: 'cuda'
-  JULIA_NUM_THREADS: '4'
+  JULIA_NUM_THREADS: '1'
 
 # See: https://github.com/JuliaGPU/gitlab-ci
 include:
