### 2020-02-21 aadd241
Build docs on Travis using Julia 1.3
