+++ .travis.yml
@@ -21,7 +21,7 @@ after_success:
 jobs:
   include:
     - stage: "Documentation"
-      julia: 1.2
+      julia: 1.3
       os: linux
       script:
         - julia --project=docs/ -e 'using Pkg; Pkg.instantiate(); Pkg.add("Documenter"); Pkg.add("Literate"); Pkg.add("Plots"); Pkg.develop(PackageSpec(path=pwd()))'
