### 2020-02-21 4986c23
Legend for right hand side functions
