### 2020-02-21 4f9a2ae
Simplify FieldBoundaryConditions constructor
