+++ src/BoundaryConditions/field_boundary_conditions.jl
@@ -14,63 +14,53 @@ Construct a `FieldBoundaryConditions` using a `CoordinateBoundaryCondition` for 
 """
 FieldBoundaryConditions(x, y, z) = FieldBoundaryConditions((x, y, z))
 
-default_tracer_bc(::Grids.Periodic) = PeriodicBoundaryCondition()
-default_tracer_bc(::Bounded)  = NoFluxBoundaryCondition()
-default_tracer_bc(::Flat)     = PeriodicBoundaryCondition()
+DefaultBoundaryCondition(::Union{Grids.Periodic, Flat}, loc) = PeriodicBoundaryCondition()
 
-# Right now it seems all defaults are the same except for no-normal
-# flow velocity BCs (which are treated in default_bc below), but
-# that might change in the future.
-const default_velocity_bc = default_tracer_bc
-const default_pressure_bc = default_tracer_bc
-const default_diffusivity_bc = default_tracer_bc
-
-"""
-    default_bc(grid, field, dim)
-
-Returns the default boundary condition for a `field` on some `grid` and along some dimension `dim`.
-"""
-function default_bc(grid, field, dim)
-    top = topology(grid)[dim]
-
-    # No normal flow velocity boundary conditions by default.
-    top isa Bounded && field == :u && dim == 1 && return NoPenetrationBoundaryCondition()
-    top isa Bounded && field == :v && dim == 2 && return NoPenetrationBoundaryCondition()
-    top isa Bounded && field == :w && dim == 3 && return NoPenetrationBoundaryCondition()
-
-    field in (:u, :v, :w) && (field = :velocity)
-    default_field_bc = Symbol(:default_, field, :_bc)
-    return @eval $default_field_bc($top)
-end
+DefaultBoundaryCondition(::Bounded, ::Cell) = NoFluxBoundaryCondition()
+DefaultBoundaryCondition(::Bounded, ::Face) = NoPenetrationBoundaryCondition()
 
 function validate_bcs(topology, left_bc, right_bc, default_bc, left_name, right_name, dir)
     if topology isa Periodic && (left_bc != default_bc || right_bc != default_bc)
-        e = "Cannot specify $left_name or $right_name boundary conditions with $topology topology in $dir-direction."
+        e = "Cannot specify $left_name or $right_name boundary conditions with " *
+            "$topology topology in $dir-direction."
         throw(ArgumentError(e))
     end
     return true
 end
 
 """
-    FieldBoundaryConditions(grid::AbstractGrid; field_type,
-                            west=default_bc(grid, field_type, 1), east=default_bc(grid, field_type, 1),
-                            south=default_bc(grid, field_type, 2), north=default_bc(grid, field_type, 2),
-                            bottom=default_bc(grid, field_type, 3), top=default_bc(grid, field_type, 3))
+    FieldBoundaryConditions(grid, loc;
+          east = DefaultBoundaryCondition(topology(grid)[1], loc[1]),
+          west = DefaultBoundaryCondition(topology(grid)[1], loc[1]),
+         south = DefaultBoundaryCondition(topology(grid)[2], loc[2]),
+         north = DefaultBoundaryCondition(topology(grid)[2], loc[2]),
+        bottom = DefaultBoundaryCondition(topology(grid)[3], loc[3]),
+           top = DefaultBoundaryCondition(topology(grid)[3], loc[3]))
 
-Construct `FieldBoundaryConditions` for a field with type `field_type` (choices are :u, :v, :w, :tracer,
[…]
