### 2020-02-21 502e16c
More expository model constructor and docstring
