+++ src/Models/incompressible_model.jl
@@ -37,7 +37,9 @@ end
                coriolis = nothing,
           surface_waves = nothing,
                 forcing = ModelForcing(),
-    boundary_conditions = NamedTuple(),
+    boundary_conditions = (u=UVelocityBoundaryConditions(grid),
+                           v=VVelocityBoundaryConditions(grid),
+                           w=WVelocityBoundaryConditions(grid)),
              parameters = nothing,
              velocities = VelocityFields(architecture, grid, boundary_conditions),
           tracer_fields = TracerFields(architecture, grid, tracernames(tracers), boundary_conditions),
@@ -73,7 +75,9 @@ function IncompressibleModel(;
                coriolis = nothing,
           surface_waves = nothing,
                 forcing = ModelForcing(),
-    boundary_conditions = NamedTuple(),
+    boundary_conditions = (u=UVelocityBoundaryConditions(grid),
+                           v=VVelocityBoundaryConditions(grid),
+                           w=WVelocityBoundaryConditions(grid)),
              parameters = nothing,
              velocities = VelocityFields(architecture, grid, boundary_conditions),
           tracer_fields = TracerFields(architecture, grid, tracernames(tracers), boundary_conditions),
