### 2020-02-21 140fb93
Update checkpointer now that BCs are inside fields
