+++ src/OutputWriters/output_writer_utils.jl
@@ -1,5 +1,5 @@
 using Oceananigans.Fields: AbstractField
-using Oceananigans.BoundaryConditions: bctype, CoordinateBoundaryConditions
+using Oceananigans.BoundaryConditions: bctype, CoordinateBoundaryConditions, FieldBoundaryConditions
 using Oceananigans.TimeSteppers: AdamsBashforthTimeStepper
 
 #####
@@ -10,16 +10,16 @@ ext(fw::AbstractOutputWriter) = throw("Extension for $(typeof(fw)) is not implem
 
 # When saving stuff to disk like a JLD2 file, `saveproperty!` is used, which
 # converts Julia objects to language-agnostic objects.
-saveproperty!(file, location, p::Number)        = file[location] = p
+saveproperty!(file, location, p::Union{Number,Array}) = file[location] = p
 saveproperty!(file, location, p::AbstractRange) = file[location] = collect(p)
 saveproperty!(file, location, p::AbstractArray) = file[location] = Array(p)
-# saveproperty!(file, location, p::AbstractField) = file[location] = Array(p.data.parent)
 saveproperty!(file, location, p::Function) = @warn "Cannot save Function property into $location"
 
-saveproperty!(file, location, p::Tuple) = [saveproperty!(file, location * "/$i", p[i]) for i in 1:length(p)]
+saveproperty!(file, location, p::Tuple) =
+    [saveproperty!(file, location * "/$i", p[i]) for i in 1:length(p)]
 
-saveproperty!(file, location, p) = [saveproperty!(file, location * "/$subp", getproperty(p, subp))
-                                    for subp in propertynames(p)]
+saveproperty!(file, location, p) =
+    [saveproperty!(file, location * "/$subp", getproperty(p, subp)) for subp in propertynames(p)]
 
 # Special saveproperty! so boundary conditions are easily readable outside julia.
 function saveproperty!(file, location, cbcs::CoordinateBoundaryConditions)
@@ -41,8 +41,15 @@ saveproperties!(file, structure, ps) = [saveproperty!(file, "$p", getproperty(st
 # When checkpointing, `serializeproperty!` is used, which serializes objects
 # unless they need to be converted (basically CuArrays only).
 serializeproperty!(file, location, p) = (file[location] = p)
+serializeproperty!(file, location, p::FieldBoundaryConditions) = (file[location] = p)
 serializeproperty!(file, location, p::AbstractArray) = saveproperty!(file, location, p)
 serializeproperty!(file, location, p::Function) = @warn "Cannot serialize Function property into $location"
+
+function serializeproperty!(file, location, p::Field{LX, LY, LZ}) where {LX, LY, LZ}
+    serializeproperty!(file, location * "/location", (LX(), LY(), LZ()))
+    serializeproperty!(file, location * "/data", p.data.parent)
+    serializeproperty!(file, location * "/boundary_conditions", p.boundary_conditions)
+end
 
 # Special serializeproperty! for AB2 time stepper struct used by the checkpointer so
 # it only saves the fields and not the tendency BCs or χ value (as they can be
@@ -52,14 +59,21 @@ function serializeproperty!(file, location, ts::AdamsBashforthTimeStepper)
     serializeproperty!(file, location * "/G⁻", ts.G⁻)
 end
 
-serializeproperties!(file, structure, ps) = [serializeproperty!(file, "$p", getproperty(structure, p)) for p in ps]
+serializeproperty!(file, location, p::NamedTuple) =
+    [serializeproperty!(file, location * "/$subp", getproperty(p, subp)) for subp in keys(p)]
+
+serializeproperties!(file, structure, ps) =
[…]
