### 2020-02-22 871ca40
Fixes to get GPU tests to pass
