+++ test/test_pressure_solvers.jl
@@ -74,7 +74,7 @@ function poisson_ppn_planned_div_free_gpu(FT, Nx, Ny, Nz)
     arch = GPU()
     grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(1.0, 2.5, 3.6))
     pbcs = PressureBoundaryConditions(grid)
-    solver = PressureSolver(arch, grid, fbcs)
+    solver = PressureSolver(arch, grid, pbcs)
 
     RHS = rand(Nx, Ny, Nz)
     RHS .= RHS .- mean(RHS)
@@ -111,7 +111,7 @@ function poisson_pnn_planned_div_free_gpu(FT, Nx, Ny, Nz)
     arch = GPU()
     grid = RegularCartesianGrid(FT, size=(Nx, Ny, Nz), length=(1.0, 2.5, 3.6), topology=(Periodic, Bounded, Bounded))
     pbcs = PressureBoundaryConditions(grid)
-    solver = PressureSolver(arch, grid, fbcs)
+    solver = PressureSolver(arch, grid, pbcs)
 
     RHS = rand(Nx, Ny, Nz)
     RHS .= RHS .- mean(RHS)
