### 2020-02-22 27a384d
Make sure to fill halo regions for total density
