### 2020-02-22 2b57d9e
Consolidate into one dry rising thermal bubble verification experiment
