+++ verification/dry_rising_thermal_bubble_entropy/dry_rising_thermal_bubble_entropy.jl
@@ -1,149 +1,0 @@
-"""
-This example sets up a dry, warm thermal bubble perturbation in a uniform
-lateral mean flow which buoyantly rises. Identical to the verification
-experiment in ../dry_rising_thermal_bubble except that entropy is used as
-the prognostic thermodynamic variable rather than potential temperature.
-"""
-
-using Printf
-using Profile
-using Plots
-using VideoIO
-using FileIO
-using Oceananigans
-using JULES
-
-const km = 1000
-const hPa = 100
-
-Lx = 20km
-Lz = 10km
-
-Δ = 0.2km  # grid spacing [m]
-
-Nx = Int(Lx/Δ)
-Ny = 1
-Nz = Int(Lz/Δ)
-
-grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
-                            x=(-Lx/2, Lx/2), y=(-Lx/2, Lx/2), z=(0, Lz))
-
-model = CompressibleModel(
-                      grid = grid,
-                 gases = DryEarth(),
-    thermodynamic_variable = Entropy(),
-                   closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
-)
-
-#####
-##### Dry thermal bubble perturbation
-#####
-
-gas = model.gases.ρ
-R, cₚ, cᵥ = gas.R, gas.cₚ, gas.cᵥ
-sref, Tref, ρref = gas.s₀, gas.T₀, gas.ρ₀
-g  = model.gravity
-pₛ = 1000hPa
-Tₛ = 300
-
-# Define an approximately hydrostatic background state
-θ₀(x, y, z) = Tₛ
-p₀(x, y, z) = pₛ * (1 - g*z/(cₚ*Tₛ))^(cₚ/R)
-T₀(x, y, z) = Tₛ*(p₀(x, y, z)/pₛ)^(R/cₚ)
-ρ₀(x, y, z) = p₀(x, y, z)/(R*T₀(x, y, z))
-ρs₀(x, y, z) = ρ₀(x, y, z) * (sref + cᵥ*log(T₀(x, y, z)/Tref) - R*log(ρ₀(x, y, z)/ρref))
-
-# Define the initial density perturbation
-xᶜ, zᶜ = 0km, 2km
-xʳ, zʳ = 2km, 2km
-L(x, y, z) = sqrt(((x - xᶜ)/xʳ)^2 + ((z - zᶜ)/zʳ)^2)
-function ρ′(x, y, z; θᶜ′ = 2.0)
-    l = L(x, y, z)
-    θ′ = (l <= 1) * θᶜ′ * cos(π/2 * L(x, y, z))^2
-    return -ρ₀(x, y, z) * θ′ / θ₀(x, y, z)
-end
-
-# Define initial state
-ρᵢ(x, y, z) = ρ₀(x, y, z) + ρ′(x, y, z)
-pᵢ(x, y, z) = p₀(x, y, z)
-Tᵢ(x, y, z) = pᵢ(x, y, z) / (R * ρᵢ(x, y, z))
-ρsᵢ(x, y, z) = ρᵢ(x, y, z) * (sref + cᵥ*log(Tᵢ(x, y, z)/Tref) - R*log(ρᵢ(x, y, z)/ρref))
-
-# Set initial state after saving perturbation-free background
-ρ, ρs = model.total_density, model.tracers.ρs
-xC, zC = grid.xC, grid.zC
-set!(model.tracers.ρ, ρ₀)
-set!(model.tracers.ρs, ρs₀)
-update_total_density!(model.total_density, model.grid, model.gases, model.tracers)
[…]

[…]
