### 2020-02-22 b217afa
Consolidate into one 3-gas rising thermal bubble verification experiment
