### 2020-02-22 b93a65a
Consolidate it one nonlinear density current verification experiment
