+++ verification/nonlinear_density_current/nonlinear_density_current.jl
@@ -12,17 +12,19 @@ using Printf
 using Plots
 using VideoIO
 using FileIO
-using Oceananigans
 using JULES
-using JULES: Π
+using Oceananigans
 
-const km = 1000
-const hPa = 100
+using Oceananigans.Fields: interiorparent
+interiorxz(field) = dropdims(interiorparent(field), dims=2)
+
+const km = 1000.0
+const hPa = 100.0
 
 Lx = 51.2km
 Lz = 6.4km
 
-Δ = 100  # grid spacing [m]
+Δ = 100.0  # grid spacing [m]
 
 Nx = Int(Lx/Δ)
 Ny = 1
@@ -30,27 +32,42 @@ Nz = Int(Lz/Δ)
 
 grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
                             x=(-Lx/2, Lx/2), y=(-Lx/2, Lx/2), z=(0, Lz))
+tvar = Energy()
+# tvar = Entropy()
+
+model = CompressibleModel(
+                      grid = grid,
+                     gases = DryEarth(),
+    thermodynamic_variable = tvar,
+                   closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
+)
 
 #####
 ##### Initial perturbation
 #####
 
-gas = IdealGas()
-Rᵈ, cₚ, cᵥ = gas.Rᵈ, gas.cₚ, gas.cᵥ
-g  = 9.80665
+gas = model.gases.ρ
+R, cₚ, cᵥ = gas.R, gas.cₚ, gas.cᵥ
+g  = model.gravity
 pₛ = 1000hPa
-Tₛ = 300
+Tₛ = 300.0
 
 # Define an approximately hydrostatic background state
 θ₀(x, y, z) = Tₛ
-p₀(x, y, z) = pₛ * (1 - g*z/(cₚ*Tₛ))^(cₚ/Rᵈ)
-T₀(x, y, z) = Tₛ*(p₀(x, y, z)/pₛ)^(Rᵈ/cₚ)
-ρ₀(x, y, z) = p₀(x, y, z)/(Rᵈ*T₀(x, y, z))
+p₀(x, y, z) = pₛ * (1 - g*z / (cₚ*Tₛ))^(cₚ/R)
+T₀(x, y, z) = Tₛ * (p₀(x, y, z)/pₛ)^(R/cₚ)
+ρ₀(x, y, z) = p₀(x, y, z) / (R*T₀(x, y, z))
+
+# Define both energy and entropy
+uᵣ, Tᵣ, ρᵣ, sᵣ = gas.u₀, gas.T₀, gas.ρ₀, gas.s₀  # Reference values
+ρe₀(x, y, z) = ρ₀(x, y, z) * (uᵣ + cᵥ * (T₀(x, y, z) - Tᵣ) + g*z)
+ρs₀(x, y, z) = ρ₀(x, y, z) * (sᵣ + cᵥ * log(T₀(x, y, z)/Tᵣ) - R * log(ρ₀(x, y, z)/ρᵣ))
 
 # Define the initial density perturbation
 xᶜ, zᶜ = 0km, 2km
 xʳ, zʳ = 2km, 2km
 L(x, y, z) = sqrt(((x - xᶜ)/xʳ)^2 + ((z - zᶜ)/zʳ)^2)
+
 function ρ′(x, y, z; θᶜ′ = -15.0)
     l = L(x, y, z)
[…]
