+++ test/test_time_stepping.jl
@@ -7,7 +7,7 @@
         grid = RegularCartesianGrid(size=(16, 16, 16), halo=(2, 2, 2), length=(1, 1, 1))
         model = CompressibleModel(grid = grid, gases = DryEarth(),
                                   thermodynamic_variable = Energy())
-        time_step!(model, Δt=1)
+        time_step!(model, 1)
         @test model isa CompressibleModel
     end
 
@@ -17,7 +17,7 @@
         grid = RegularCartesianGrid(size=(16, 16, 16), halo=(2, 2, 2), length=(1, 1, 1))
         model = CompressibleModel(grid = grid, gases = DryEarth(),
                                   thermodynamic_variable = Entropy())
-        time_step!(model, Δt=1)
+        time_step!(model, 1)
         @test model isa CompressibleModel
     end
 end
