### 2020-02-22 ba1ff66
Fix show method for simulations
