### 2020-02-22 236531c
Allow simulations to hold parameters
