### 2020-02-22 ff5ff0b
Clean up and pretty up some show methods
