### 2020-02-22 47f3378
Add a show method for CompressibleModel
