### 2020-02-22 714a7ba
Update thermal bubble verification to use functions and simulations
