+++ verification/dry_rising_thermal_bubble/dry_rising_thermal_bubble.jl
@@ -16,174 +16,203 @@ interiorxz(field) = dropdims(interiorparent(field), dims=2)
 const km = 1000.0
 const hPa = 100.0
 
-Lx = 20km
-Lz = 10km
-Δ  = 0.2km  # grid spacing [m]
+function simulate_dry_rising_thermal_bubble(; thermodynamic_variable, end_time=1000.0, make_plots=true)
+    tvar = thermodynamic_variable
 
-Nx = Int(Lx/Δ)
-Ny = 1
-Nz = Int(Lz/Δ)
+    Lx = 20km
+    Lz = 10km
+    Δ  = 0.2km  # grid spacing [m]
 
-grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
-                            x=(-Lx/2, Lx/2), y=(-Lx/2, Lx/2), z=(0, Lz))
+    Nx = Int(Lx/Δ)
+    Ny = 1
+    Nz = Int(Lz/Δ)
 
-tvar = Energy()
-# tvar = Entropy()
+    grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
+                                x=(-Lx/2, Lx/2), y=(-Lx/2, Lx/2), z=(0, Lz))
 
-model = CompressibleModel(
-                      grid = grid,
-                     gases = DryEarth(),
-    thermodynamic_variable = tvar,
-                   closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
-)
+    model = CompressibleModel(
+                          grid = grid,
+                         gases = DryEarth(),
+        thermodynamic_variable = tvar,
+                       closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
+    )
 
-#####
-##### Dry thermal bubble perturbation
-#####
+    #####
+    ##### Dry thermal bubble perturbation
+    #####
 
-gas = model.gases.ρ
-R, cₚ, cᵥ = gas.R, gas.cₚ, gas.cᵥ
-g  = model.gravity
-pₛ = 1000hPa
-Tₛ = 300.0
+    gas = model.gases.ρ
+    R, cₚ, cᵥ = gas.R, gas.cₚ, gas.cᵥ
+    g  = model.gravity
+    pₛ = 1000hPa
+    Tₛ = 300.0
 
-# Define an approximately hydrostatic background state
-θ₀(x, y, z) = Tₛ
-p₀(x, y, z) = pₛ * (1 - g*z / (cₚ*Tₛ))^(cₚ/R)
-T₀(x, y, z) = Tₛ * (p₀(x, y, z)/pₛ)^(R/cₚ)
-ρ₀(x, y, z) = p₀(x, y, z) / (R*T₀(x, y, z))
+    # Define an approximately hydrostatic background state
+    θ₀(x, y, z) = Tₛ
+    p₀(x, y, z) = pₛ * (1 - g*z / (cₚ*Tₛ))^(cₚ/R)
+    T₀(x, y, z) = Tₛ * (p₀(x, y, z)/pₛ)^(R/cₚ)
+    ρ₀(x, y, z) = p₀(x, y, z) / (R*T₀(x, y, z))
 
-# Define both energy and entropy
-uᵣ, Tᵣ, ρᵣ, sᵣ = gas.u₀, gas.T₀, gas.ρ₀, gas.s₀  # Reference values
-ρe₀(x, y, z) = ρ₀(x, y, z) * (uᵣ + cᵥ * (T₀(x, y, z) - Tᵣ) + g*z)
[…]
