### 2020-02-22 e137b29
Update three gas verification to use functions and simulations
