+++ verification/three_gas_dry_rising_thermal_bubble/three_gas_dry_rising_thermal_bubble.jl
@@ -12,163 +12,200 @@ interiorxz(field) = dropdims(interiorparent(field), dims=2)
 const km = 1000.0
 const hPa = 100.0
 
-Lx = 20km
-Lz = 10km
-Δ  = 0.2km  # grid spacing [m]
+function simulate_three_gas_dry_rising_thermal_bubble(;
+        thermodynamic_variable, end_time=1000.0, make_plots=true)
 
-Nx = Int(Lx/Δ)
-Ny = 1
-Nz = Int(Lz/Δ)
+    tvar = thermodynamic_variable
 
-grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
-                            x=(-Lx/2, Lx/2), y=(-Lx/2, Lx/2), z=(0, Lz))
+    Lx = 20km
+    Lz = 10km
+    Δ  = 0.2km  # grid spacing [m]
 
-tvar = Energy()
-# tvar = Entropy()
+    Nx = Int(Lx/Δ)
+    Ny = 1
+    Nz = Int(Lz/Δ)
 
-model = CompressibleModel(
-                      grid = grid,
-                     gases = DryEarth3(),
-    thermodynamic_variable = tvar,
-                   closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
-)
+    grid = RegularCartesianGrid(size=(Nx, Ny, Nz), halo=(2, 2, 2),
+                                x=(-Lx/2, Lx/2), y=(-Lx/2, Lx/2), z=(0, Lz))
 
-#####
-##### Dry thermal bubble perturbation
-#####
+    model = CompressibleModel(
+                          grid = grid,
+                         gases = DryEarth3(),
+        thermodynamic_variable = tvar,
+                       closure = ConstantIsotropicDiffusivity(ν=75.0, κ=75.0)
+    )
 
-gas = model.gases.ρ₁
-R, cₚ, cᵥ = gas.R, gas.cₚ, gas.cᵥ
-g  = model.gravity
-pₛ = 1000hPa
-Tₛ = 300.0
+    #####
+    ##### Dry thermal bubble perturbation
+    #####
 
-# Define initial mixing ratios
-q₁(z) = exp(-(4z/Lz)^2)
-q₃(z) = exp(-(4*(z - Lz)/Lz)^2)
-q₂(z) = 1 - q₁(z) - q₃(z)
+    gas = model.gases.ρ₁
+    R, cₚ, cᵥ = gas.R, gas.cₚ, gas.cᵥ
+    g  = model.gravity
+    pₛ = 1000hPa
+    Tₛ = 300.0
 
-# Define an approximately hydrostatic background state
-θ₀(x, y, z) = Tₛ
-p₀(x, y, z) = pₛ * (1 - g*z / (cₚ*Tₛ))^(cₚ/R)
-T₀(x, y, z) = Tₛ * (p₀(x, y, z)/pₛ)^(R/cₚ)
-ρ₀(x, y, z) = p₀(x, y, z) / (R*T₀(x, y, z))
+    # Define initial mixing ratios
+    q₁(z) = exp(-(4z/Lz)^2)
+    q₃(z) = exp(-(4*(z - Lz)/Lz)^2)
+    q₂(z) = 1 - q₁(z) - q₃(z)
 
-ρ₁₀(x, y, z) = q₁(z) * ρ₀(x, y, z)
-ρ₂₀(x, y, z) = q₂(z) * ρ₀(x, y, z)
-ρ₃₀(x, y, z) = q₃(z) * ρ₀(x, y, z)
+    # Define an approximately hydrostatic background state
+    θ₀(x, y, z) = Tₛ
[…]
