### 2020-02-23 207b8e6
Add regression tests from the verification experiments
