+++ test/runtests.jl
@@ -1,11 +1,13 @@
 using Test
 
+using JLD2
 using Oceananigans
 using JULES
 
 @testset "JULES" begin
     include("test_models.jl")
     include("test_time_stepping.jl")
+    include("test_regression.jl")
 end
 
 include("../benchmarks/benchmark_static_atmosphere.jl")
