### 2020-02-23 bbeea56
Function to summarized regression tests
