+++ test/test_regression.jl
@@ -1,5 +1,20 @@
 using Oceananigans.Fields: interiorparent
 
+function summarize_regression_test(field_names, fields, correct_fields)
+    for (field_name, φ, φ_c) in zip(field_names, fields, correct_fields)
+        Δ = Array(φ) .- φ_c
+
+        Δ_min      = minimum(Δ)
+        Δ_max      = maximum(Δ)
+        Δ_mean     = mean(Δ)
+        Δ_abs_mean = mean(abs, Δ)
+        Δ_std      = std(Δ)
+
+        @info(@sprintf("Δ%s: min=%.6e, max=%.6e, mean=%.6e, absmean=%.6e, std=%.6e",
+                       field_name, Δ_min, Δ_max, Δ_mean, Δ_abs_mean, Δ_std))
+    end
+end
+
 @testset "Regression" begin
     @info "Testing regression..."
 
@@ -30,6 +45,24 @@ using Oceananigans.Fields: interiorparent
             # end
 
             file = jldopen(regression_filepath, "r")
+
+            field_names = [:ρ, :ρu, :ρv, :ρw]
+            fields = [interior(model.total_density), interior(model.momenta.ρu),
+                      interior(model.momenta.ρv), interior(model.momenta.ρw)]
+            correct_fields = [file["ρ"], file["ρu"], file["ρv"], file["ρw"]]
+
+            if tvar isa Energy
+                push!(field_names, :ρe)
+                push!(fields, interior(model.tracers.ρe))
+                push!(correct_fields, file["ρe"])
+            elseif tvar isa Entropy
+                push!(field_names, :ρs)
+                push!(fields, interior(model.tracers.ρs))
+                push!(correct_fields, file["ρs"])
+            end
+
+            summarize_regression_test(field_names, fields, correct_fields)
+
             @test all(interior(model.total_density) .≈ file["ρ"])
             @test all(interior(model.momenta.ρu)    .≈ file["ρu"])
             @test all(interior(model.momenta.ρv)    .≈ file["ρv"])
@@ -72,6 +105,27 @@ using Oceananigans.Fields: interiorparent
             # end
 
             file = jldopen(regression_filepath, "r")
+
+            field_names = [:ρ, :ρu, :ρv, :ρw, :ρ₁, :ρ₂, :ρ₃]
+            fields = [interior(model.total_density), interior(model.momenta.ρu),
+                      interior(model.momenta.ρv), interior(model.momenta.ρw),
+                      interior(model.tracers.ρ₁), interior(model.tracers.ρ₂),
+                      interior(model.tracers.ρ₃)]
+            correct_fields = [file["ρ"], file["ρu"], file["ρv"], file["ρw"],
+                              file["ρ₁"], file["ρ₂"], file["ρ₃"]]
+
+            if tvar isa Energy
+                push!(field_names, :ρe)
+                push!(fields, interior(model.tracers.ρe))
+                push!(correct_fields, file["ρe"])
+            elseif tvar isa Entropy
+                push!(field_names, :ρs)
+                push!(fields, interior(model.tracers.ρs))
+                push!(correct_fields, file["ρs"])
+            end
+
+            summarize_regression_test(field_names, fields, correct_fields)
+
             @test all(interior(model.total_density) .≈ file["ρ"])
             @test all(interior(model.momenta.ρu)    .≈ file["ρu"])
             @test all(interior(model.momenta.ρv)    .≈ file["ρv"])
