### 2020-02-23 c60bf08
Reorder properties in `CompressibleModel`
