### 2020-02-23 e28b4be
Start standardizing variable names and function signature order
