+++ src/pressure_gradients.jl
@@ -4,6 +4,6 @@ using JULES.Operators: ∂xᶠᵃᵃ, ∂yᵃᶠᵃ, ∂zᵃᵃᶠ
 #### Pressure gradient ∇p terms for entropy S = ρs
 ####
 
-@inline ∂p∂x(i, j, k, grid, tvar, gravity, momenta, total_density, densities, tracers) = ∂xᶠᵃᵃ(i, j, k, grid, diagnose_p, tvar, gravity, momenta, total_density, densities, tracers)
-@inline ∂p∂y(i, j, k, grid, tvar, gravity, momenta, total_density, densities, tracers) = ∂yᵃᶠᵃ(i, j, k, grid, diagnose_p, tvar, gravity, momenta, total_density, densities, tracers)
-@inline ∂p∂z(i, j, k, grid, tvar, gravity, momenta, total_density, densities, tracers) = ∂zᵃᵃᶠ(i, j, k, grid, diagnose_p, tvar, gravity, momenta, total_density, densities, tracers)
+@inline ∂p∂x(i, j, k, grid, tvar, gravity, momenta, total_density, gases, tracers) = ∂xᶠᵃᵃ(i, j, k, grid, diagnose_p, tvar, gases, gravity, total_density, momenta, tracers)
+@inline ∂p∂y(i, j, k, grid, tvar, gravity, momenta, total_density, gases, tracers) = ∂yᵃᶠᵃ(i, j, k, grid, diagnose_p, tvar, gases, gravity, total_density, momenta, tracers)
+@inline ∂p∂z(i, j, k, grid, tvar, gravity, momenta, total_density, gases, tracers) = ∂zᵃᵃᶠ(i, j, k, grid, diagnose_p, tvar, gases, gravity, total_density, momenta, tracers)
