+++ src/Operators/compressible_operators.jl
@@ -1,33 +1,43 @@
 #####
+##### Legend:
+##### ρ  -> density fields
+##### ρŨ -> momentum fields
+##### ρC̃ -> conservative tracer fields
+##### K̃  -> diffusivity fields
+##### Ũ  -> velocity fields
+##### C̃  -> tracer fields
+#####
+
+#####
 ##### Convert conservative variables to primitive variables
 #####
 
-@inline U_over_ρ(i, j, k, grid, U, ρ) = @inbounds U[i, j, k] / ℑxᶠᵃᵃ(i, j, k, grid, ρ)
-@inline V_over_ρ(i, j, k, grid, V, ρ) = @inbounds V[i, j, k] / ℑyᵃᶠᵃ(i, j, k, grid, ρ)
-@inline W_over_ρ(i, j, k, grid, W, ρ) = @inbounds W[i, j, k] / ℑzᵃᵃᶠ(i, j, k, grid, ρ)
-@inline C_over_ρ(i, j, k, grid, C, ρ) = @inbounds C[i, j, k] / ρ[i, j, k]
+@inline U_over_ρ(i, j, k, grid, ρ, ρu) = @inbounds ρu[i, j, k] / ℑxᶠᵃᵃ(i, j, k, grid, ρ)
+@inline V_over_ρ(i, j, k, grid, ρ, ρv) = @inbounds ρv[i, j, k] / ℑyᵃᶠᵃ(i, j, k, grid, ρ)
+@inline W_over_ρ(i, j, k, grid, ρ, ρw) = @inbounds ρw[i, j, k] / ℑzᵃᵃᶠ(i, j, k, grid, ρ)
+@inline C_over_ρ(i, j, k, grid, ρ, ρc) = @inbounds ρc[i, j, k] / ρ[i, j, k]
 
 #####
 ##### Kinetic energy
 #####
 
-@inline kinetic_energy(i, j, k, grid::AbstractGrid{FT}, Ũ, ρ) where FT =
-    @inbounds FT(0.5) * ((ℑxᶜᵃᵃ(i, j, k, grid, Ũ.ρu) / ρ[i, j, k])^2 +
-                         (ℑyᵃᶜᵃ(i, j, k, grid, Ũ.ρv) / ρ[i, j, k])^2 +
-                         (ℑzᵃᵃᶜ(i, j, k, grid, Ũ.ρw) / ρ[i, j, k])^2)
+@inline kinetic_energy(i, j, k, grid::AbstractGrid{FT}, ρ, ρũ) where FT =
+    @inbounds FT(0.5) * ((ℑxᶜᵃᵃ(i, j, k, grid, ρũ.ρu) / ρ[i, j, k])^2 +
+                         (ℑyᵃᶜᵃ(i, j, k, grid, ρũ.ρv) / ρ[i, j, k])^2 +
+                         (ℑzᵃᵃᶜ(i, j, k, grid, ρũ.ρw) / ρ[i, j, k])^2)
 
 #####
 ##### Pressure work
 #####
 
[…]

[…]
