### 2020-02-23 b37f852
Rely on Oceananigans for Coriolis terms
