### 2020-02-23 5dc5025
Standardize operators
