+++ src/Operators/Operators.jl
@@ -7,10 +7,9 @@ using Oceananigans.Operators
 using Oceananigans: AbstractGrid
 
 export
-    x_f_cross_U, y_f_cross_U, z_f_cross_U,
     kinetic_energy,
     hdivᶜᶜᵃ, divᶜᶜᶜ, ∇²,
-    div_flux, ∂ⱼpuⱼ, ∂ⱼDᶜⱼ, ∂ⱼtᶜDᶜⱼ, ∂ⱼDᵖⱼ,
+    div_uc, ∂ⱼpuⱼ, ∂ⱼDᶜⱼ, ∂ⱼtᶜDᶜⱼ, ∂ⱼDᵖⱼ,
     div_ρuũ, div_ρvũ, div_ρwũ,
     ∂ⱼτ₁ⱼ, ∂ⱼτ₂ⱼ, ∂ⱼτ₃ⱼ, Q_dissipation
 
