### 2020-02-23 250ae60
Almost done standarizing variable names and function signatures
