+++ src/thermodynamics.jl
@@ -43,14 +43,14 @@ struct IdealGas{FT} <: AbstractEquationOfState
     u₀ :: FT
 end
 
-function EarthN₂O₂(FT = Float64; T₀ = T₀, p₀ = p₀, s₀ = s₀, u₀ = u₀)
+function EarthN₂O₂(FT=Float64; T₀=T₀, p₀=p₀, s₀=s₀, u₀=u₀)
     # Calculate abundance-weighted molar masses and DOFs
     M = (η_N₂_earth * M_N₂ + η_O₂_earth * M_O₂)/(η_N₂_earth + η_O₂_earth)
     dof = (η_N₂_earth * dof_N₂ + η_O₂_earth * dof_O₂)/(η_N₂_earth + η_O₂_earth)
-    return IdealGas(FT, M, dof; T₀ = T₀, p₀ = p₀, s₀ = s₀, u₀ = u₀)
+    return IdealGas(FT, M, dof; T₀=T₀, p₀=p₀, s₀=s₀, u₀=u₀)
 end
 
-function IdealGas(FT, M, dof; T₀ = T₀, p₀ = p₀, s₀ = s₀, u₀ = u₀)
+function IdealGas(FT, M, dof; T₀=T₀, p₀=p₀, s₀=s₀, u₀=u₀)
     R = R⁰/M
     cᵥ = R*dof/2
     cₚ = cᵥ + R
@@ -71,70 +71,69 @@ struct Energy  <: AbstractThermodynamicVariable end
 ##### Thermodynamic state diagnostics
 #####
 
-@inline function diagnose_ρs(i, j, k, grid::AbstractGrid{FT}, tracer_index, tvar, gases, gravity, total_density, momenta, tracers) where FT
+@inline function diagnose_ρs(i, j, k, grid::AbstractGrid{FT}, tracer_index, tvar, gases, gravity, ρ, ρũ, ρc̃) where FT
     @inbounds begin
-        T = diagnose_T(i, j, k, grid, tvar, gases, gravity, total_density, momenta, tracers)
-        ρ = tracers[tracer_index].data[i, j, k]
+        T = diagnose_T(i, j, k, grid, tvar, gases, gravity, ρ, ρũ, ρc̃)
+        ρᵅ = ρc̃[tracer_index].data[i, j, k]
         gas = gases[tracer_index-1]
-        return ρ > zero(FT) ? ρ*(gas.s₀ + gas.cᵥ*log(T/gas.T₀) - gas.R*log(ρ/gas.ρ₀)) : zero(FT)
+        return ρᵅ > zero(FT) ? ρᵅ*(gas.s₀ + gas.cᵥ*log(T/gas.T₀) - gas.R*log(ρᵅ/gas.ρ₀)) : zero(FT)
     end
 end
 
-@inline function diagnose_T(i, j, k, grid::AbstractGrid{FT}, tvar::Entropy, gases, gravity, total_density, momenta, tracers) where FT
+@inline function diagnose_T(i, j, k, grid::AbstractGrid{FT}, tvar::Entropy, gases, gravity, ρ, ρũ, ρc̃) where FT
     @inbounds begin
-        numerator = tracers.ρs.data[i, j, k]
+        numerator = ρc̃.ρs.data[i, j, k]
         denominator = zero(FT)
         for (gas_index, gas) in enumerate(gases)
-            ρ = tracers[gas_index+1].data[i, j, k]
-            numerator += ρ > 0 ? (ρ*gas.R*log(ρ/gas.ρ₀) - ρ*gas.s₀) : zero(FT)
-            denominator += ρ*gas.cᵥ
+            ρᵅ = ρc̃[gas_index+1].data[i, j, k]
[…]
