### 2020-02-23 8d68e69
Clean up main time stepping loop
