+++ src/time_stepping_kernels.jl
@@ -15,17 +15,17 @@ update_total_density!(model) =
 """
 Slow forcings include viscous dissipation, diffusion, and Coriolis terms.
 """
-function compute_slow_forcings!(F, grid, tvar, gases, gravity, coriolis, closure, ρ, ρũ, ρc̃, K̃, forcing, time)
+function compute_slow_forcings!(F̃, grid, tvar, gases, gravity, coriolis, closure, ρ, ρũ, ρc̃, K̃, forcing, time)
     @inbounds begin
         for k in 1:grid.Nz, j in 1:grid.Ny, i in 1:grid.Nx
-            F.ρu[i, j, k] = FU(i, j, k, grid, coriolis, closure, ρ, ρũ, K̃) + forcing.u(i, j, k, grid, time, ρũ, ρc̃, nothing)
-            F.ρv[i, j, k] = FV(i, j, k, grid, coriolis, closure, ρ, ρũ, K̃) + forcing.v(i, j, k, grid, time, ρũ, ρc̃, nothing)
-            F.ρw[i, j, k] = FW(i, j, k, grid, coriolis, closure, ρ, ρũ, K̃) + forcing.w(i, j, k, grid, time, ρũ, ρc̃, nothing)
+            F̃.ρu[i, j, k] = FU(i, j, k, grid, coriolis, closure, ρ, ρũ, K̃) + forcing.u(i, j, k, grid, time, ρũ, ρc̃, nothing)
+            F̃.ρv[i, j, k] = FV(i, j, k, grid, coriolis, closure, ρ, ρũ, K̃) + forcing.v(i, j, k, grid, time, ρũ, ρc̃, nothing)
+            F̃.ρw[i, j, k] = FW(i, j, k, grid, coriolis, closure, ρ, ρũ, K̃) + forcing.w(i, j, k, grid, time, ρũ, ρc̃, nothing)
         end
 
         for (tracer_index, ρc_name) in enumerate(propertynames(ρc̃))
             ρc   = getproperty(ρc̃, ρc_name)
-            F_ρc = getproperty(F.tracers, ρc_name)
+            F_ρc = getproperty(F̃.tracers, ρc_name)
 
             for k in 1:grid.Nz, j in 1:grid.Ny, i in 1:grid.Nx
                 F_ρc[i, j, k] = FC(i, j, k, grid, closure, tracer_index, ρ, ρc, K̃)
@@ -33,7 +33,7 @@ function compute_slow_forcings!(F, grid, tvar, gases, gravity, coriolis, closure
         end
 
         for k in 1:grid.Nz, j in 1:grid.Ny, i in 1:grid.Nx
-            F.tracers[1].data[i, j, k] += FT(i, j, k, grid, closure, tvar, gases, gravity, ρ, ρũ, ρc̃, K̃)
+            F̃.tracers[1].data[i, j, k] += FT(i, j, k, grid, closure, tvar, gases, gravity, ρ, ρũ, ρc̃, K̃)
         end
 
     end
@@ -42,26 +42,26 @@ end
 """
 Fast forcings include advection, pressure gradient, and buoyancy terms.
 """
-function compute_right_hand_sides!(R, grid, tvar, gases, gravity, ρ, ρũ, ρc̃, F)
+function compute_right_hand_sides!(R̃, grid, tvar, gases, gravity, ρ, ρũ, ρc̃, F̃)
     @inbounds begin
         for k in 1:grid.Nz, j in 1:grid.Ny, i in 1:grid.Nx
[…]
