### 2020-02-24 52fcacd
Add named tuples for lazy velocity and lazy tracer fields
Co-Authored-By: Tristan Abbott <tabbott0312@gmail.com>
