+++ src/lazy_fields.jl
@@ -1,42 +1,44 @@
+using Base: @propagate_inbounds
 import Base: getindex
 
-struct LazyPrimitiveField{X, Y, Z, A, G, C, D} <: AbstractLocatedField{X, Y, Z, A, G}
-    architecture :: A
-    grid :: G
+using Oceananigans.Operators: ℑxᶠᵃᵃ, ℑyᵃᶠᵃ, ℑzᵃᵃᶠ
+using Oceananigans.Fields
+
+struct LazyPrimitiveField{X, Y, Z, A, G, C, D} <: AbstractField{X, Y, Z, A, G}
+          architecture :: A
+                  grid :: G
     conservative_field :: C
-    density :: D
+               density :: D
 end
 
-@inline getindex(f::LazyPrimitiveField{Cell, Cell, Cell}, inds...) =
-    @inbounds f.conservative_field[inds...] / f.density[inds...]
+LazyPrimitiveField(LX, LY, LZ, arch, grid, ρϕ, ρ) =
+    LazyPrimitiveField{LX, LY, LZ, typeof(arch), typeof(grid), typeof(ρϕ), typeof(ρ)}(arch, grid, ρϕ, ρ)
 
-@inline getindex(f::LazyPrimitiveField{Face, Cell, Cell}, inds...) =
-    @inbounds f.conservative_field[inds..] / ℑxᶠᵃᵃ(inds..., f.grid, f.density)
+@inline @propagate_inbounds getindex(f::LazyPrimitiveField{Cell, Cell, Cell}, I...) =
+    @inbounds f.conservative_field[I...] / f.density[I...]
 
-@inline getindex(f::LazyPrimitiveField{Cell, Face, Cell}, inds...) =
-    @inbounds f.conservative_field[inds..] / ℑyᵃᶠᵃ(inds..., f.grid, f.density)
+@inline @propagate_inbounds getindex(f::LazyPrimitiveField{Face, Cell, Cell}, I...) =
+    @inbounds f.conservative_field[I...] / ℑxᶠᵃᵃ(I..., f.grid, f.density)
 
-@inline getindex(f::LazyPrimitiveField{Cell, Cell, Face}, inds...) =
-    @inbounds f.conservative_field[inds..] / ℑzᵃᵃᶠ(inds..., f.grid, f.density)
+@inline @propagate_inbounds getindex(f::LazyPrimitiveField{Cell, Face, Cell}, I...) =
+    @inbounds f.conservative_field[I...] / ℑyᵃᶠᵃ(I..., f.grid, f.density)
 
-struct LazyTotalDensityField{X, Y, Z, A, G, D} <: AbstractLocatedField{X, Y, Z, A, G}
-    architecture :: A
-    grid :: G
-    densities :: D
+@inline @propagate_inbounds getindex(f::LazyPrimitiveField{Cell, Cell, Face}, I...) =
+    @inbounds f.conservative_field[I...] / ℑzᵃᵃᶠ(I..., f.grid, f.density)
+
+LazyVelocityFields(arch, grid, ρ, ρũ) =
+    (u = LazyPrimitiveField(Face, Cell, Cell, arch, grid, ρũ.ρu.data, ρ.data),
+     v = LazyPrimitiveField(Cell, Face, Cell, arch, grid, ρũ.ρv.data, ρ.data),
+     w = LazyPrimitiveField(Cell, Cell, Face, arch, grid, ρũ.ρw.data, ρ.data))
+
+function LazyTracerFields(arch, grid, ρ, ρc̃)
+    c_names = [filter(c -> c != 'ρ', string(c)) for c in keys(ρc̃)]
+    c_names = filter(s -> s != "", c_names) .|> Symbol |> Tuple  # Don't include the ρ tracer.
+
[…]
