### 2020-02-24 aa7515b
Add tests for primitive lazy fields
