+++ test/runtests.jl
@@ -8,6 +8,7 @@ using JULES
 
 @testset "JULES" begin
     include("test_models.jl")
+    include("test_lazy_fields.jl")
     include("test_time_stepping.jl")
     include("test_regression.jl")
 end
