### 2020-02-24 13f2411
Concretely type NetCDFOutputWriter and clean it up a bit
