+++ src/OutputWriters/netcdf_output_writer.jl
@@ -6,20 +6,17 @@ using Oceananigans.Utils: validate_interval
 """
     netcdf_spatial_dimensions(::field)
 
-Returns the dimensions associated with a field.
+Returns the NetCDF dimensions associated with a field.
 
 Examples
 ========
-
 julia> netcdf_spatial_dimensions(model.velocities.u)
 ("xF", "yC", "zC")
 
 julia> netcdf_spatial_dimensions(model.tracers.T)
 ("xC", "yC", "zC")
 """
-function netcdf_spatial_dimensions(field::Field{LX, LY, LZ}) where {LX, LY, LZ}
-    xdim(LX), ydim(LY), zdim(LZ)
-end
+netcdf_spatial_dimensions(::Field{LX, LY, LZ}) where {LX, LY, LZ} = xdim(LX), ydim(LY), zdim(LZ)
 
 xdim(::Type{Face}) = "xF"
 xdim(::Type{Cell}) = "xC"
@@ -41,11 +38,10 @@ Writes a grid.nc file that contains all the dimensions of the domain.
 
 Keyword arguments
 =================
-
-    - `filename::String`  : File name to be saved under
-    - `mode::String`      : Netcdf file is opened in either clobber ("c") or append ("a") mode (Default: "c" )
-    - `compression::Int`  : Defines the compression level of data (0-9, default 0)
-    - `attributes::Dict`  : Attributes (default: Dict())
+- `filename`  : File name to be saved under.
+- `mode`: NetCDF file is opened in either clobber ("c") or append ("a") mode. Default: "c".
+- `compression`: Defines the compression level of data from 0-9. Default: 0.
+- `attributes`: Global attributes. Default: Dict().
 """
 function write_grid(model; filename="./grid.nc", mode="c",
                     compression=0, attributes=Dict(), slice_kw...)
@@ -86,27 +82,34 @@ function write_grid(model; filename="./grid.nc", mode="c",
     return nothing
 end
 
+const default_output_attributes =
+    Dict("u" => Dict("longname" => "Velocity in the x-direction", "units" => "m/s"),
+         "v" => Dict("longname" => "Velocity in the y-direction", "units" => "m/s"),
+         "w" => Dict("longname" => "Velocity in the z-direction", "units" => "m/s"),
+         "T" => Dict("longname" => "Temperature", "units" => "K"),
+         "S" => Dict("longname" => "Salinity",    "units" => "g/kg"))
 
 """
     NetCDFOutputWriter <: AbstractOutputWriter
 
 An output writer for writing to NetCDF files.
 """
-mutable struct NetCDFOutputWriter <: AbstractOutputWriter
+mutable struct NetCDFOutputWriter{D, O, I, F, S} <: AbstractOutputWriter
              filename :: String
-              dataset :: Any
-              outputs :: Dict
-             interval :: Union{Nothing, AbstractFloat}
-            frequency :: Union{Nothing, Int}
+              dataset :: D
+              outputs :: O
+             interval :: I
+            frequency :: F
               clobber :: Bool
-               slices :: Dict
+               slices :: S
    len_time_dimension :: Int
              previous :: Float64
 end
 
 """
     NetCDFOutputWriter(model, outputs; interval=nothing, frequency=nothing, filename=".",
-                                   clobber=true, global_attributes=Dict(), output_attributes=nothing, slice_kw...)
+                       clobber=true, global_attributes=Dict(), output_attributes=nothing,
+                       slice_kw...)
 
 Construct a `NetCDFOutputWriter` that writes `label, field` pairs in `outputs` (which should
 be a `Dict`) to a NC file, where `label` is a symbol that labels the output and `field` is
@@ -114,54 +117,57 @@ a field from the model (e.g. `model.velocities.u`).
 
[…]
