### 2020-02-24 d31f19c
Get rid of read_output function
