### 2020-02-24 77a6ede
Clarify `write_grid_and_attributes` function
