### 2020-02-24 ff2b20e
Arbitrary output functions in NetCDF output writer
