+++ src/OutputWriters/netcdf_output_writer.jl
@@ -107,24 +107,25 @@ mutable struct NetCDFOutputWriter{D, O, I, F, S} <: AbstractOutputWriter
 end
 
 """
-    NetCDFOutputWriter(model, outputs; interval=nothing, frequency=nothing, filename=".",
-                       clobber=true, global_attributes=Dict(), output_attributes=nothing,
-                       slice_kw...)
+    NetCDFOutputWriter(model, outputs; frequency=nothing, interval=nothing, filename=".",
+                       global_attributes=Dict(), output_attributes=Dict(), dimensions=Dict(),
+                       clobber=true, compression=0, slice_kw...)
 
-Construct a `NetCDFOutputWriter` that writes `label, field` pairs in `outputs` (which should
-be a `Dict`) to a NC file, where `label` is a symbol that labels the output and `field` is
-a field from the model (e.g. `model.velocities.u`).
+Construct a `NetCDFOutputWriter` that writes `(label, output)` pairs in `outputs` (which should
+be a `Dict`) to a NetCDF file, where `label` is a string that labels the output and `output` is
+either field from the model (e.g. `model.velocities.u`) or a function `f(model)` that returns
+something to be written to disk.
 
 Keyword arguments
 =================
-- `filename`: Filepath to save output to. Default: "." (current working directory).
 - `frequency`: Save output every `n` model iterations.
 - `interval`: Save output every `t` units of model clock time.
-- `clobber`: Remove existing files if their filenames conflict. Default: `true`.
-- `compression`: Determines the compression level of data (0-9, default 0)
+- `filename`: Filepath to save output to. Default: "." (current working directory).
 - `global_attributes`: Dict of model properties to save with every file (deafult: Dict())
 - `output_attributes`: Dict of attributes to be saved with each field variable (reasonable
-  defaults are provided for velocities, temperature, and salinity)
+  defaults are provided for velocities, temperature, and salinity).
+- `clobber`: Remove existing files if their filenames conflict. Default: `true`.
+- `compression`: Determines the compression level of data (0-9, default 0)
 - `slice_kw`: `dimname = Union{OrdinalRange, Integer}` will slice the dimension `dimname`.
   All other keywords are ignored. E.g. `xC = 3:10` will only produce output along the dimension
   `xC` between indices 3 and 10 for all fields with `xC` as one of their dimensions. `xC = 1`
@@ -133,12 +134,11 @@ Keyword arguments
 """
 
 function NetCDFOutputWriter(model, outputs; interval=nothing, frequency=nothing, filename=".",
-                            clobber=true, global_attributes=Dict(), output_attributes=Dict(),
-                            compression=0, slice_kw...)
-
-    validate_interval(interval, frequency)
+                            global_attributes=Dict(), output_attributes=Dict(), dimensions=Dict(),
+                            clobber=true, compression=0, slice_kw...)
 
     mode = clobber ? "c" : "a"
+    validate_interval(interval, frequency)
 
     # Initiates the output file with dimensions
     write_grid_and_attributes(model; filename=filename, compression=compression,
@@ -149,8 +149,6 @@ function NetCDFOutputWriter(model, outputs; interval=nothing, frequency=nothing,
 
     # Creates an unliimited dimension "time"
     defDim(dataset, "time", Inf)
-    sync(dataset)
-
     defVar(dataset, "time", Float64, ("time",))
     sync(dataset)
 
@@ -165,23 +163,29 @@ function NetCDFOutputWriter(model, outputs; interval=nothing, frequency=nothing,
     end
 
     # Initiates empty variables for fields from the user-supplied variable outputs
-    for (fieldname, field) in outputs
[…]
