+++ src/OutputWriters/OutputWriters.jl
@@ -3,7 +3,7 @@ module OutputWriters
 export
     write_output,
     JLD2OutputWriter, FieldOutput, FieldOutputs,
-    NetCDFOutputWriter, write_grid,
+    NetCDFOutputWriter, write_grid_and_attributes,
     Checkpointer, restore_from_checkpoint
 
 using Oceananigans
