### 2020-02-24 d5a92db
Testing NetCDF function outputs are correct
