+++ test/test_output_writers.jl
@@ -79,6 +79,65 @@ function run_thermal_bubble_netcdf_tests(arch)
     @test all(S_sliced .≈ Array(interiorparent(model.tracers.S))[xC_slice, yC_slice, zC_slice])
 end
 
+function run_netcdf_function_output_tests(arch)
+    N = 16
+    L = 1
+    model = IncompressibleModel(grid=RegularCartesianGrid(size=(N, N, N), length=(L, 2L, 3L)))
+    simulation = Simulation(model, Δt=1.25, stop_iteration=3)
+
+    # Define scalar, vector, 2D slice, and 3D field outputs
+    f(model) = model.clock.time^2
+    g(model) = @. model.clock.time * exp(model.grid.zC)
+    h(model) = @. model.clock.time * sin(model.grid.xC) * cos(model.grid.yC')
+
+    outputs = Dict("scalar" => f, "profile" => g, "slice" => h)
+    dims = Dict("scalar" => (), "profile" => ("zC",), "slice" => ("xC", "yC"))
+
+    output_attributes = Dict(
+        "scalar"  => Dict("longname" => "Some scalar", "units" => "bananas"),
+        "profile" => Dict("longname" => "Some vertical profile", "units" => "watermelons"),
+        "slice"   => Dict("longname" => "Some slice", "units" => "mushrooms")
+    )
+
+    global_attributes = Dict("location" => "Bay of Fundy", "onions" => 7)
+
+    simulation.output_writers[:fruits] =
+        NetCDFOutputWriter(
+            model, outputs; frequency=1, filename="test_function_outputs.nc", dimensions=dims,
+            global_attributes=global_attributes, output_attributes=output_attributes)
+
+    run!(simulation)
+    close(simulation.output_writers[:fruits])
+
+    ds = Dataset("test_function_outputs.nc", "r")
+
+    @test ds.attrib["location"] == "Bay of Fundy"
+    @test ds.attrib["onions"] == 7
+
+    @test length(ds["time"]) == 4
+    @test ds["time"][:] == [1.25i for i in 0:3]
+
+    @test ds["scalar"].attrib["longname"] == "Some scalar"
+    @test ds["scalar"].attrib["units"] == "bananas"
+    @test ds["scalar"][:] == [(1.25i)^2 for i in 0:3]
+    @test dimnames(ds["scalar"]) == ("time",)
+
+    @test ds["profile"].attrib["longname"] == "Some vertical profile"
+    @test ds["profile"].attrib["units"] == "watermelons"
+    @test ds["profile"][:, end] == @. 3.75 * exp(model.grid.zC)
+    @test size(ds["profile"]) == (N, 4)
+    @test dimnames(ds["profile"]) == ("zC", "time")
+
+    @test ds["slice"].attrib["longname"] == "Some slice"
+    @test ds["slice"].attrib["units"] == "mushrooms"
+    @test ds["slice"][:, :, end] == @. 3.75 * sin(model.grid.xC) * cos(model.grid.yC')
+    @test size(ds["slice"]) == (N, N, 4)
+    @test dimnames(ds["slice"]) == ("xC", "yC", "time")
+
+    close(ds)
+    return nothing
+end
+
 function run_jld2_file_splitting_tests(arch)
     model = IncompressibleModel(grid=RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1)))
     simulation = Simulation(model, Δt=1, stop_iteration=10)
@@ -185,6 +244,7 @@ end
          @testset "NetCDF [$(typeof(arch))]" begin
              @info "  Testing NetCDF output writer [$(typeof(arch))]..."
              run_thermal_bubble_netcdf_tests(arch)
+             run_netcdf_function_output_tests(arch)
          end
 
         @testset "JLD2 [$(typeof(arch))]" begin
