### 2020-02-24 dcdb323
No need for `len_time_dimension` property for NetCDF output writer
