### 2020-02-24 35807d0
Adds a short_show method for OffsetArrays
