### 2020-02-24 82e0276
Fixes tests for correct field size
