### 2020-02-24 a96c8f8
Properly use isapprox for zero values
