### 2020-02-25 5a574b8
Makes checkpointer know about field location
