### 2020-02-25 c588527
Many bugfix in fields and grids associated with merge conflict resolution
