+++ src/Fields/field_tuples.jl
@@ -4,11 +4,11 @@
 Return a NamedTuple with fields `u`, `v`, `w` initialized on the architecture `arch`
 and `grid`. Fields may be passed via the optional keyword arguments `u`, `v`, and `w`.
 """
-VelocityFields(arch, grid;
-    u = XFaceField(arch, grid, UVelocityBoundaryConditions(grid), zeros(arch, grid)),
-    v = YFaceField(arch, grid, VVelocityBoundaryConditions(grid), zeros(arch, grid)),
-    w = ZFaceField(arch, grid, WVelocityBoundaryConditions(grid), zeros(arch, grid))
-) = (u=u, v=v, w=w)
+function VelocityFields(arch, grid; u = XFaceField(arch, grid, UVelocityBoundaryConditions(grid)),
+                                    v = YFaceField(arch, grid, VVelocityBoundaryConditions(grid)),
+                                    w = ZFaceField(arch, grid, WVelocityBoundaryConditions(grid)))
+    return (u=u, v=v, w=w)
+end
 
 """
     VelocityFields(arch, grid, bcs::NamedTuple)
@@ -22,9 +22,9 @@ function VelocityFields(arch, grid, bcs::NamedTuple)
     v_bcs = :v ∈ keys(bcs) ? bcs[:v] : VVelocityBoundaryConditions(grid)
     w_bcs = :w ∈ keys(bcs) ? bcs[:w] : WVelocityBoundaryConditions(grid)
 
-    u = XFaceField(arch, grid, u_bcs, zeros(arch, grid))
-    v = YFaceField(arch, grid, v_bcs, zeros(arch, grid))
-    w = ZFaceField(arch, grid, w_bcs, zeros(arch, grid))
+    u = XFaceField(arch, grid, u_bcs)
+    v = YFaceField(arch, grid, v_bcs)
+    w = ZFaceField(arch, grid, w_bcs)
 
     return (u=u, v=v, w=w)
 end
@@ -50,7 +50,7 @@ function TracerFields(arch, grid, tracer_names; kwargs...)
     tracer_fields =
         Tuple(c ∈ keys(kwargs) ?
               kwargs[c] :
-              CellField(arch, grid, TracerBoundaryConditions(grid), zeros(arch, grid))
+              CellField(arch, grid, TracerBoundaryConditions(grid))
               for c in tracer_names)
     return NamedTuple{tracer_names}(tracer_fields)
 end
@@ -65,8 +65,8 @@ be specified via a named tuple of `FieldBoundaryCondition`s.
 function TracerFields(arch, grid, tracer_names, bcs::NamedTuple)
     tracer_fields =
         Tuple(c ∈ keys(bcs) ?
-              CellField(arch, grid, bcs[c],                         zeros(arch, grid)) :
-              CellField(arch, grid, TracerBoundaryConditions(grid), zeros(arch, grid))
+              CellField(arch, grid, bcs[c]) :
+              CellField(arch, grid, TracerBoundaryConditions(grid))
               for c in tracer_names)
     return NamedTuple{tracer_names}(tracer_fields)
 end
@@ -87,10 +87,11 @@ Return a NamedTuple with pressure fields `pHY′` and `pNHS` initialized as
 `CellField`s on the architecture `arch` and `grid`. Fields may be passed via the
 optional keyword arguments `pHY′` and `pNHS`.
 """
-PressureFields(arch, grid;
-    pHY′ = CellField(arch, grid, PressureBoundaryConditions(grid), zeros(arch, grid)),
-    pNHS = CellField(arch, grid, PressureBoundaryConditions(grid), zeros(arch, grid))
[…]
