+++ src/Fields/field.jl
@@ -6,7 +6,7 @@ using OffsetArrays
 
 import Oceananigans.Architectures: architecture
 import Oceananigans.Utils: datatuple
-import Oceananigans.Grids: topology, x_topology, y_topology, z_topology
+import Oceananigans.Grids: halosize, topology, x_topology, y_topology, z_topology
 
 using Oceananigans.Architectures
 using Oceananigans.Grids
@@ -35,11 +35,11 @@ struct Field{X, Y, Z, A, G, B} <: AbstractField{X, Y, Z, A, G}
     boundary_conditions :: B
 
     function Field{X, Y, Z}(data, grid, bcs) where {X, Y, Z}
-        Tx, Ty, Tz = grid.Nx+2grid.Hx, grid.Ny+2grid.Hy, grid.Nz+2grid.Hz
+        Tx, Ty, Tz = total_size((X, Y, Z), grid)
 
         if size(data) != (Tx, Ty, Tz)
-            e = "Cannot construct field with size(data)=$(size(data)). " *
-                "Must have the same size as the grid with halos ($Tx, $Ty, $Tz)."
+            e = "Cannot construct field at ($X, $Y, $Z) with size(data)=$(size(data)). " *
+                "`data` must have size ($Tx, $Ty, $Tz)."
             throw(ArgumentError(e))
         end
 
@@ -54,7 +54,7 @@ Construct a `Field` on some architecture `arch` and a `grid` with some `data`.
 The field's location is defined by a tuple `L` of length 3 whose elements are
 `Cell` or `Face` and has field boundary conditions `bcs`.
 """
-Field(L::Tuple, arch, grid, bcs, data=zeros(arch, grid)) =
+Field(L::Tuple, arch, grid, bcs, data=zeros(eltype(grid), arch, grid, (L[1], L[2], L[3]))) =
     Field{L[1], L[2], L[3]}(data, grid, bcs)
 
 """
@@ -64,8 +64,8 @@ Construct a `Field` on some architecture `arch` and a `grid` with some `data`.
 The field's location is defined by `X`, `Y`, `Z` where each is either `Cell` or `Face`
 and has field boundary conditions `bcs`.
 """
-Field(X, Y, Z, arch, grid, bcs, data=zeros(eltype(grid), arch, grid, (X, Y, Z)) = 
-     Field((X, Y, Z), arch, grid, bcs, data)
+Field(X, Y, Z, arch, grid, bcs, data=zeros(eltype(grid), arch, grid, (X, Y, Z))) =
+    Field((X, Y, Z), arch, grid, bcs, data)
 
 """
     CellField([FT=eltype(grid)], arch::AbstractArchitecture, grid, bcs=TracerBoundaryConditions(grid), 
@@ -74,8 +74,12 @@ Field(X, Y, Z, arch, grid, bcs, data=zeros(eltype(grid), arch, grid, (X, Y, Z)) 
 Return a `Field{Cell, Cell, Cell}` on architecture `arch` and `grid` containing `data`
 with field boundary conditions `bcs`.
 """
-CellField(arch::AbstractArchitecture, grid, bcs=TracerBoundaryConditions(grid), 
-          data=zeros(FT, arch, grid, (Cell, Cell, Cell))) = Field(Cell, Cell, Cell, arch, grid, bcs, data)
+function CellField(arch::AbstractArchitecture, grid, 
+                   bcs=TracerBoundaryConditions(grid),
+                   data=zeros(eltype(grid), arch, grid, (Cell, Cell, Cell)))
+
+    return Field(Cell, Cell, Cell, arch, grid, bcs, data)
+end
 
 """
     XFaceField([FT=eltype(grid)], arch::AbstractArchitecture, grid, bcs=UVelocityBoundaryConditions(grid), 
[…]
