### 2020-02-25 c8ff702
Cleans up computations and fixes latent bug
