### 2020-02-25 d9b65ea
Extends total_size function to distinguish between the physical size of fields and the total size of underlying data
