+++ src/Grids/Grids.jl
@@ -54,7 +54,13 @@ z_topology(::AbstractGrid{FT, TX, TY, TZ}) where {FT, TX, TY, TZ} = TZ()
 
 size(grid::AbstractGrid) = (grid.Nx, grid.Ny, grid.Nz)
 length(grid::AbstractGrid) = (grid.Lx, grid.Ly, grid.Lz)
-halosize(grid) = (grid.Hx, grid.Hy, grid.Hz)
+halo_size(grid) = (grid.Hx, grid.Hy, grid.Hz)
+
+total_size(a) = size(a) # fallback
+
+total_size(grid::AbstractGrid) = (grid.Nx + grid.Hx, 
+                                  grid.Ny + grid.Hy, 
+                                  grid.Nz + grid.Hz)
 
 include("grid_utils.jl")
 include("regular_cartesian_grid.jl")
