+++ src/Diagnostics/horizontal_average.jl
@@ -1,6 +1,7 @@
 using Oceananigans.Architectures
 using Oceananigans.BoundaryConditions
 using Oceananigans.Utils
+using Oceananigans.Grids: total_size
 
 """
     HorizontalAverage{F, R, P, I, Ω} <: AbstractDiagnostic
@@ -37,7 +38,7 @@ model and you want to save the output to disk by passing it to an output writer.
 """
 function HorizontalAverage(field; frequency=nothing, interval=nothing, return_type=Array)
     arch = architecture(field)
-    result = zeros(arch, field.grid, 1, 1, size(parent(field))[3])
+    result = zeros(arch, field.grid, 1, 1, total_size(parent(field))[3])
     return HorizontalAverage(field, result, frequency, interval, 0.0, return_type, field.grid)
 end
 
