### 2020-02-25 b910b1d
Removes unnecessary import of renamed halo_size function
