### 2020-02-25 5bb9293
Revise face locations in NetCDF output
