### 2020-02-25 de20789
Nukes x_topology, etc in favor of more julian syntax; functionality for instantiated location types
