+++ src/Fields/field.jl
@@ -6,7 +6,7 @@ using OffsetArrays
 
 import Oceananigans.Architectures: architecture
 import Oceananigans.Utils: datatuple
-import Oceananigans.Grids: total_size, topology, x_topology, y_topology, z_topology
+import Oceananigans.Grids: total_size, topology
 
 using Oceananigans.Architectures
 using Oceananigans.Grids
@@ -54,8 +54,17 @@ Construct a `Field` on some architecture `arch` and a `grid` with some `data`.
 The field's location is defined by a tuple `L` of length 3 whose elements are
 `Cell` or `Face` and has field boundary conditions `bcs`.
 """
-Field(L::Tuple, arch, grid, bcs, data=zeros(eltype(grid), arch, grid, (L[1], L[2], L[3]))) =
-    Field{L[1], L[2], L[3]}(data, grid, bcs)
+function Field(L::Tuple, arch, grid, bcs, 
+               data=zeros(eltype(grid), arch, grid, (typeof(L[1]), typeof(L[2]), typeof(L[3]))))
+      
+    return Field{typeof(L[1]), typeof(L[2]), typeof(L[3])}(data, grid, bcs)
+end
+
+function Field(L::NTuple{3, DataType}, arch, grid, bcs, 
+               data=zeros(eltype(grid), arch, grid, (L[1], L[2], L[3])))
+
+    return Field{L[1], L[2], L[3]}(data, grid, bcs)
+end
 
 """
     Field(X, Y, Z, arch, grid, [data=zeros(arch, grid)], bcs)
@@ -157,10 +166,7 @@ architecture(o::OffsetArray) = architecture(o.parent)
 
 
 "Returns the topology of a fields' `grid`."
-@inline topology(f) = topology(f.grid)
-@inline x_topology(f) = x_topology(f.grid)
-@inline y_topology(f) = y_topology(f.grid)
-@inline z_topology(f) = z_topology(f.grid)
+@inline topology(f, args...) = topology(f.grid, args...)
 
 """
 Returns the length of a field located at `Cell` centers along a grid 
@@ -181,9 +187,9 @@ Returns the size of a field at `loc` on `grid`.
 This is a 3-tuple of integers corresponding to the number of interior nodes
 of `f` along `x, y, z`.
 """
-@inline size(loc, grid) = (dimension_length(loc[1], x_topology(grid), grid.Nx), 
-                           dimension_length(loc[2], y_topology(grid), grid.Ny), 
-                           dimension_length(loc[3], z_topology(grid), grid.Nz))
+@inline size(loc, grid) = (dimension_length(loc[1], topology(grid, 1), grid.Nx), 
+                           dimension_length(loc[2], topology(grid, 2), grid.Ny), 
+                           dimension_length(loc[3], topology(grid, 3), grid.Nz))
 
 """
     size(f::AbstractField{X, Y, Z}) where {X, Y, Z}
@@ -201,9 +207,9 @@ Returns the "total" size of a field at `loc` on `grid`.
 This is a 3-tuple of integers corresponding to the number of grid points
 contained by `f` along `x, y, z`.
 """
-@inline total_size(loc, grid) = (dimension_length(loc[1], x_topology(grid), grid.Nx, grid.Hx), 
-                                 dimension_length(loc[2], y_topology(grid), grid.Ny, grid.Hy), 
-                                 dimension_length(loc[3], z_topology(grid), grid.Nz, grid.Hz))
+@inline total_size(loc, grid) = (dimension_length(loc[1], topology(grid, 1), grid.Nx, grid.Hx), 
[…]
