### 2020-02-25 4e3d69b
Bugfix --- merge makes no sense for tuples
