### 2020-02-25 e4883e6
Moves AbstractCPUField, etc to set!.jl
