### 2020-02-25 2019cb7
NetCDF rescue with multiple dispatch
