### 2020-02-26 7809b1e
Fix NetCDFOutputWriter for GPUs
