### 2020-02-27 ad2e6ca
Improves getting time step from Simulation class
