### 2020-02-28 5eea0d4
Simulation handles closing of netcdfoutputwriter
