### 2020-02-28 3757242
Defines open and close for AbstractOutputwriter type
