### 2020-02-28 2373487
Remove second definition of `convert_to_arch`
