### 2020-02-29 e55490e
Removes tracer_fields kwarg from Model, adjusts checkpointer and TracerFields functions accordingly
