### 2020-02-29 a51f856
Nukes timestepper_method kwarg to model and makes appropriate changes to checkpointer
