### 2020-02-29 71899f2
Uncomment name checks for v and w in has_velocities
