### 2020-02-29 71b9c2f
Support `AbstractTime` clocks instead of just `DateTime` clocks
