+++ test/runtests.jl
@@ -36,6 +36,7 @@ using Oceananigans.Logger
 using Oceananigans.Utils
 
 using Dates: DateTime, Nanosecond
+using TimesDates: TimeDate
 using Statistics: mean
 using LinearAlgebra: norm
 using GPUifyLoops: @launch, @loop
