### 2020-02-29 c1c62fe
Add lazy velocity and tracers fields to CompressibleModel
