### 2020-02-29 9a62b5a
Lazy fields use the field, not `field.data`
