### 2020-02-29 5674553
Define `interior` for lazy fields
