### 2020-02-29 876eb3c
Simplify verification experiments using lazy fields
