### 2020-02-29 7620fed
Update verification experiment to account for new FaceField sizes
