### 2020-03-02 d5baa8f
Bump Oceananigans v0.24.2
