### 2020-03-03 7956c41
Moves some tests earlier in the script
