### 2020-03-04 2b12d21
Changes default value of time stepping parameter to 0
