### 2020-03-04 9ceb221
Adds more tests to confirm that the files are closed
