### 2020-03-04 18b28d2
Refactors regression tests slightly
