+++ test/regression_tests/rayleigh_benard_regression_test.jl
@@ -1,3 +1,4 @@
+
 function run_rayleigh_benard_regression_test(arch)
 
     #####
@@ -52,24 +53,16 @@ function run_rayleigh_benard_regression_test(arch)
     spinup_steps = 1000
       test_steps = 100
 
-    output_U(model) = datatuple(model.velocities)
-    output_Φ(model) = datatuple(model.tracers)
-    output_G(model) = datatuple(model.timestepper.Gⁿ)
-    outputfields = Dict(:U=>output_U, :Φ=>output_Φ, :G=>output_G)
+    prefix = "rayleigh_benard"
 
-    prefix = "data_rayleigh_benard_regression"
-    output_writer =
-      JLD2OutputWriter(model, outputfields, dir=joinpath(dirname(@__FILE__), "data"),
-                       prefix=prefix, frequency=test_steps, including=[])
+    checkpointer = Checkpointer(model, frequency=test_steps, prefix=prefix,
+                                dir=joinpath(dirname(@__FILE__), "data"))
 
     #####
     ##### Initial condition and spinup steps for creating regression test data
     #####
 
-    #=
-    @warn ("Generating new data for the Rayleigh-Benard regression test.
-           New regression test data generation will fail unless the JLD2
-           file $prefix.jld2 is manually deleted.")
+    @warn "Generating new data for the Rayleigh-Benard regression test."
 
     ξ(z) = a * rand() * z * (Lz + z) # noise, damped at the walls
     b₀(x, y, z) = (ξ(z) - z) / Lz
@@ -78,66 +71,72 @@ function run_rayleigh_benard_regression_test(arch)
     simulation.stop_iteration = spinup_steps-test_steps
     run!(simulation)
 
-    push!(simulation.output_writers, output_writer)
+    push!(simulation.output_writers, checkpointer)
     simulation.stop_iteration += 2test_steps
     run!(simulation)
-    =#
 
     #####
     ##### Regression test
     #####
 
     # Load initial state
-    u₀, v₀, w₀ = get_output_tuple(output_writer, spinup_steps, :U)
-    b₀, c₀ = get_output_tuple(output_writer, spinup_steps, :Φ)
-    Gu₀, Gv₀, Gw₀, Gb₀, Gc₀ = get_output_tuple(output_writer, spinup_steps, :G)
+    initial_filename = joinpath(dirname(@__FILE__), "data", prefix * "_iteration$spinup_steps.jld2")
 
-    model.velocities.u.data.parent .= ArrayType(u₀.parent)
-    model.velocities.v.data.parent .= ArrayType(v₀.parent)
-    model.velocities.w.data.parent[:, :, 1:Nz+2] .= ArrayType(w₀.parent)
-    model.tracers.b.data.parent    .= ArrayType(b₀.parent)
-    model.tracers.c.data.parent    .= ArrayType(c₀.parent)
+    solution₀, Gⁿ₀, G⁻₀ = get_fields_from_checkpoint(initial_filename)
 
-    model.timestepper.Gⁿ.u.data.parent .= ArrayType(Gu₀.parent)
-    model.timestepper.Gⁿ.v.data.parent .= ArrayType(Gv₀.parent)
-    model.timestepper.Gⁿ.w.data.parent[:, :, 1:Nz+2] .= ArrayType(Gw₀.parent)
-    model.timestepper.Gⁿ.b.data.parent .= ArrayType(Gb₀.parent)
-    model.timestepper.Gⁿ.c.data.parent .= ArrayType(Gc₀.parent)
+    model.velocities.u.data.parent .= ArrayType(solution₀.u)
+    model.velocities.v.data.parent .= ArrayType(solution₀.v)
[…]
