### 2020-03-04 d614833
Comments out regression test data generation sections
