### 2020-03-04 a193901
Exports data from Fields submodule
