+++ src/TimeSteppers/adams_bashforth.jl
@@ -47,6 +47,8 @@ function time_step!(model::IncompressibleModel{<:AdamsBashforthTimeStepper}, Δt
 
     calculate_explicit_substep!(Gⁿ, velocities, tracers, pressures, diffusivities, model)
 
+    ab2_time_step_tracers!(tracers, model.architecture, model.grid, Δt, χ, Gⁿ, G⁻)
+
     ab2_update_source_terms!(Gⁿ, model.architecture, model.grid, χ, G⁻)
 
     calculate_pressure_correction!(pressures.pNHS, Δt, Gⁿ, velocities, model)
