### 2020-03-04 f93af56
Refactors time-stepping; no longer calculating Gc^{n+1/2}
