+++ src/TimeSteppers/kernels.jl
@@ -109,87 +109,6 @@ function calculate_boundary_source_terms!(Gⁿ, arch, U, C, args...)
     return nothing
 end
 
-"""
-Update the horizontal velocities u and v via
-
-    `u^{n+1} = u^n + (Gu^{n+½} - δₓp_{NH} / Δx) Δt`
-
-Note that the vertical velocity is not explicitly time stepped.
-"""
-function update_velocities!(U, grid, Δt, G, pNHS)
-    @loop_xyz i j k grid begin
-        @inbounds U.u[i, j, k] += (G.u[i, j, k] - ∂xᶠᵃᵃ(i, j, k, grid, pNHS)) * Δt
-        @inbounds U.v[i, j, k] += (G.v[i, j, k] - ∂yᵃᶠᵃ(i, j, k, grid, pNHS)) * Δt
-    end
-    return nothing
-end
-
-"""
-Update the horizontal velocities u and v via
-
-    `u^{n+1} = u^n + Gu^{n+½} / Δt`
-
-Note that the vertical velocity is not explicitly time stepped.
-"""
-function update_velocities!(U, grid, Δt, G, ::Nothing)
-    @loop_xyz i j k grid begin
-        @inbounds U.u[i, j, k] += G.u[i, j, k] * Δt
-        @inbounds U.v[i, j, k] += G.v[i, j, k] * Δt
-    end
-    return nothing
-end
-
-"""
-Update tracers via
-
-    `c^{n+1} = c^n + Gc^{n+½} Δt`
-"""
-function update_tracer!(c, grid, Δt, Gc)
-    @loop_xyz i j k grid begin
-        @inbounds c[i, j, k] += Gc[i, j, k] * Δt
-    end
-    return nothing
-end
-
-"Update the solution variables (velocities and tracers)."
-function update_solution!(U, C, arch, grid, Δt, G, pNHS)
-    @launch device(arch) config=launch_config(grid, :xyz) update_velocities!(U, grid, Δt, G, pNHS)
-
-    #=
-    for i in 1:length(C)
-        @inbounds c = C[i]
-        @inbounds Gc = G[i+3]
-        @launch device(arch) config=launch_config(grid, :xyz) update_tracer!(c, grid, Δt, Gc)
-    end
-    =#
-
-    return nothing
-end
-
-function ab2_time_step_tracers!(C, arch, grid, Δt, χ, Gⁿ, G⁻)
-    for i in 1:length(C)
-        @inbounds c = C[i]
-        @inbounds Gcⁿ = Gⁿ[i+3]
-        @inbounds Gc⁻ = G⁻[i+3]
-        @launch device(arch) config=launch_config(grid, :xyz) ab2_time_step_tracer!(c, grid, Δt, χ, Gcⁿ, Gc⁻)
-    end
-
-    return nothing
-end
-
-"""
-Time step tracers via
-
-    `c^{n+1} = c^n + Δt ( (3/2 + χ) * Gc^{n} - (1/2 + χ) G^{n-1} )`
-
-"""
-function ab2_time_step_tracer!(c, grid::AbstractGrid{FT}, Δt, χ, Gcⁿ, Gc⁻) where FT
-    @loop_xyz i j k grid begin
-        @inbounds c[i, j, k] += Δt * ((FT(1.5) + χ) * Gcⁿ[i, j, k] - (FT(0.5) + χ) * Gc⁻[i, j, k])
-    end
-    return nothing
-end
-
 #####
 ##### Vertical integrals
 #####
