### 2020-03-04 4efa2a4
Use intermediate update with predictor velocities
