### 2020-03-04 9d1ae93
Nukes unused complete_pressure_correction function
