### 2020-03-05 919f5fa
Clean up and tidying up of time stepping functionality
