+++ src/TimeSteppers/adams_bashforth.jl
@@ -49,21 +49,21 @@ function time_step!(model::IncompressibleModel{<:AdamsBashforthTimeStepper}, Δt
 
     ab2_time_step_tracers!(tracers, model.architecture, model.grid, Δt, χ, Gⁿ, G⁻)
 
-    ab2_update_source_terms!(Gⁿ, model.architecture, model.grid, χ, G⁻)
-
+    # Fractional step. Note that predictor velocities share memory space with velocities.
     ab2_update_predictor_velocities!(predictor_velocities, model.architecture, model.grid, Δt, χ, Gⁿ, G⁻)
 
     calculate_pressure_correction!(pressures.pNHS, Δt, predictor_velocities, model)
 
-    ab2_time_step_velocities!(velocities, tracers, model.architecture,
-                              model.grid, Δt, Gⁿ, pressures.pNHS)
+    fractional_step_velocities!(velocities, tracers, model.architecture,
+                                model.grid, Δt, pressures.pNHS)
 
+    ab2_store_source_terms!(G⁻, model.architecture, model.grid, χ, Gⁿ)
+
+    # Compute w from recontinuity
     fill_halo_regions!(model.velocities, model.architecture,
                        boundary_condition_function_arguments(model)...)
 
     compute_w_from_continuity!(model)
-
-    ab2_store_source_terms!(G⁻, model.architecture, model.grid, χ, Gⁿ)
 
     tick!(model.clock, Δt)
 
@@ -71,7 +71,7 @@ function time_step!(model::IncompressibleModel{<:AdamsBashforthTimeStepper}, Δt
 end
 
 #####
-##### Adams-Bashforth-specific kernels
+##### Source term storage
 #####
 
 """ Store previous source terms before updating them. """
@@ -95,9 +95,9 @@ end
 """ Store source terms for `u`, `v`, and `w`. """
 function ab2_store_velocity_source_terms!(G⁻, grid::AbstractGrid{FT}, χ, Gⁿ) where FT
     @loop_xyz i j k grid begin
-        @inbounds G⁻.u[i, j, k] = ((FT(0.5) + χ) * G⁻.u[i, j, k] + Gⁿ.u[i, j, k]) / (FT(1.5) + χ)
-        @inbounds G⁻.v[i, j, k] = ((FT(0.5) + χ) * G⁻.v[i, j, k] + Gⁿ.v[i, j, k]) / (FT(1.5) + χ)
-        @inbounds G⁻.w[i, j, k] = ((FT(0.5) + χ) * G⁻.w[i, j, k] + Gⁿ.w[i, j, k]) / (FT(1.5) + χ)
+        @inbounds G⁻.u[i, j, k] = Gⁿ.u[i, j, k]
+        @inbounds G⁻.v[i, j, k] = Gⁿ.v[i, j, k]
+        @inbounds G⁻.w[i, j, k] = Gⁿ.w[i, j, k]
     end
     return nothing
 end
@@ -107,68 +107,6 @@ function ab2_store_tracer_source_term!(Gc⁻, grid::AbstractGrid{FT}, χ, Gcⁿ)
     @loop_xyz i j k grid begin
         @inbounds Gc⁻[i, j, k] = Gcⁿ[i, j, k]
     end
-    return nothing
-end
-
-"""
-Evaluate the right-hand-side terms for velocity fields and tracer fields
-at time step n+½ using a weighted 2nd-order Adams-Bashforth method.
-"""
[…]
