### 2020-03-05 186f1a8
Fixes bug in fill_north_halo
