### 2020-03-05 55e29c4
Adds 3 horizontal divergence operators
