### 2020-03-05 e359d56
Include boundary condition function arguments for filling tendency halos
