### 2020-03-05 2b9423f
Properly construct time stepper after restoring
