+++ src/OutputWriters/checkpointer.jl
@@ -153,7 +153,7 @@ function restore_from_checkpoint(filepath; kwargs=Dict{Symbol,Any}())
 
     # Restore time stepper
     kwargs[:timestepper] =
-        AdamsBashforthTimeStepper(eltype(grid), arch, grid, tracer_names;
+        AdamsBashforthTimeStepper(eltype(grid), arch, grid, kwargs[:velocities], tracer_names;
                                   G⁻ = TendencyFields(arch, grid, tracer_names; G⁻_tendency_field_kwargs...),
                                   Gⁿ = TendencyFields(arch, grid, tracer_names; Gⁿ_tendency_field_kwargs...))
 
