### 2020-03-05 72105b4
Change function names in RHS calculation to prepare for boundary-tendency calculations
