+++ src/TimeSteppers/generic_time_stepping.jl
@@ -44,16 +44,26 @@ contribution from non-hydrostatic pressure.
 """
 function calculate_tendencies!(tendencies, velocities, tracers, pressures, diffusivities, model)
 
-    calculate_interior_source_terms!(
-        tendencies, model.architecture, model.grid, model.coriolis, model.buoyancy,
-        model.surface_waves, model.closure, velocities, tracers, pressures.pHY′,
-        diffusivities, model.forcing, model.parameters, model.clock.time
-    )
+    # Note:
+    #
+    # "tendencies" is a NamedTuple of OffsetArrays corresponding to the tendency data for use
+    # in GPU computations.
+    #
+    # "model.timestepper.Gⁿ" is a NamedTuple of Fields, whose data also corresponds to 
+    # tendency data.
+    #
+    tendency_args = (tendencies, model.architecture, model.grid, model.coriolis, model.buoyancy,
+                     model.surface_waves, model.closure, velocities, tracers, pressures.pHY′,
+                     diffusivities, model.forcing, model.parameters, model.clock.time)
 
-    calculate_boundary_source_terms!(
+    calculate_interior_tendency_contributions!(tendency_args...)
+
+    calculate_boundary_tendency_contributions!(
         model.timestepper.Gⁿ, model.architecture, model.velocities,
-        model.tracers, boundary_condition_function_arguments(model)...
-    )
+        model.tracers, boundary_condition_function_arguments(model)...)
+
+    # Calculate
+    #calculate_tendencies_on_boundaries!(tendency_args...)
 
     return nothing
 end
