### 2020-03-05 ed37d46
More descriptive comments in calculate_interior_tendencies
