### 2020-03-05 2c1e89d
Adds new file defining tendency functions
