+++ src/TimeSteppers/TimeSteppers.jl
@@ -58,6 +58,7 @@ boundary_condition_function_arguments(model) =
      datatuple(model.tracers), model.parameters)
 
 include("generic_time_stepping.jl")
+include("momentum_and_tracer_tendencies.jl")
 include("time_stepping_kernels.jl")
 include("adams_bashforth.jl")
 
