### 2020-03-05 9e6f0aa
Minor docstring reformatting in velocity and tracer tendencies
