### 2020-03-06 10e1e55
Implements functions for calculating Gu, Gv, Gw on east, north, top boundaries
