+++ src/TimeSteppers/time_stepping_kernels.jl
@@ -4,7 +4,7 @@
 
 """ Store previous value of the source term and calculate current source term. """
 function calculate_interior_tendency_contributions!(G, arch, grid, coriolis, buoyancy, surface_waves, closure, 
-                                          U, C, pHY′, K, F, parameters, time)
+                                                    U, C, pHY′, K, F, parameters, time)
 
     # Manually choose thread-block layout here as it's ~20% faster.
     # See: https://github.com/climate-machine/Oceananigans.jl/pull/308
@@ -42,32 +42,154 @@ function calculate_interior_tendency_contributions!(G, arch, grid, coriolis, buo
     return nothing
 end
 
-""" Calculate the right-hand-side of the u-momentum equation. """
+"""
+    calculate_velocity_tendencies_on_boundaries!(tendency_calculation_args...)
+
+Calculate the velocity tendencies *on* east, north, and top boundaries, when the
+x-, y-, or z- directions have a `Bounded` topology.
+"""
+function calculate_velocity_tendencies_on_boundaries!(tendency_calculation_args...)
+
+    calculate_east_boundary_Gu!(tendency_calculation_args...)
+    calculate_north_boundary_Gv!(tendency_calculation_args...)
+    calculate_top_boundary_Gw!(tendency_calculation_args...)
+
+    return nothing
+end
+
+# Fallbacks for non-bounded topologies in x-, y-, and z-directions
+@inline calculate_east_boundary_Gu!(args...) = nothing
+@inline calculate_north_boundary_Gv!(args...) = nothing
+@inline calculate_top_boundary_Gw!(args...) = nothing
+
+"""
+    calculate_east_boundary_Gu!(G, arch, grid::AbstractGrid{FT, <:Bounded},
+                                coriolis, buoyancy, surface_waves, closure,
+                                U, C, pHY′, K, F, parameters, time) where FT
+
+Calculate `Gu` on east boundaries when the x-direction has `Bounded` topology.
+"""
+function calculate_east_boundary_Gu!(G, arch, grid::AbstractGrid{FT, <:Bounded},
+                                     coriolis, buoyancy, surface_waves, closure,
+                                     U, C, pHY′, K, F, parameters, time) where FT
+
+    @launch(device(arch), config=launch_config(grid, :yz), 
+            _calculate_east_boundary_Gu!(G.u, grid, coriolis, surface_waves, 
+                                         closure, U, C, K, F, pHY′, parameters, time))
+
+    return nothing
+end
+
+"""
+    calculate_north_boundary_Gu!(G, arch, grid::AbstractGrid{FT, <:Bounded},
+                                 coriolis, buoyancy, surface_waves, closure,
+                                 U, C, pHY′, K, F, parameters, time) where FT
+
+Calculate `Gv` on north boundaries when the y-direction has `Bounded` topology.
+"""
+function calculate_north_boundary_Gv!(G, arch, grid::AbstractGrid{FT, TX, <:Bounded},
+                                      coriolis, buoyancy, surface_waves, closure,
+                                      U, C, pHY′, K, F, parameters, time) where {FT, TX}
+
+    @launch(device(arch), config=launch_config(grid, :xz), 
+            _calculate_north_boundary_Gv!(G.v, grid, coriolis, surface_waves, 
+                                          closure, U, C, K, F, pHY′, parameters, time))
+
+    return nothing
+end
+
+"""
+    calculate_top_boundary_Gw!(G, arch, grid::AbstractGrid{FT, <:Bounded},
+                               coriolis, buoyancy, surface_waves, closure,
[…]
