### 2020-03-06 e40e36a
Moves boundary function functionality entirely to src/BoundaryConditions; requires defining placeholders in Oceananigans top-level
