### 2020-03-06 062c4f4
Update eady_turbulence.jl
