### 2020-03-06 2836e72
Update plotting in eady_turbulence.jl
