### 2020-03-06 15f9907
Test that Eady turbulence examples runs
