+++ test/test_examples.jl
@@ -114,9 +114,18 @@ end
         replace_strings = [
             ("Nh = 64", "Nh = 16"),
             ("Nz = 32", "Nz = 16"),
-            ("end_time = 3day", "end_time = 1")
+            ("end_time = 3day", "end_time = 1"),
+            # Get rid of anything PyPlot/PyCall related
+            ("using PyPlot, PyCall", ""),
+            ("GridSpec =", "#"),
+            ("fig =", "#"),
+            ("gs =", "#"),
+            ("fig.add_subplot", "#"),
+            ("gcf()", "#"),
+            ("makeplot!", "#makeplot!"),
+            ("function #makeplot!(axs, model)", "function makeplot!(axs, model)")
         ]
 
-        @test_skip run_example(replace_strings, "eady_turbulence")
+        @test run_example(replace_strings, "eady_turbulence")
     end
 end
