### 2020-03-07 7a5e110
CompatHelper: bump compat for "CUDAdrv" to "6.0"
