### 2020-03-07 eba9954
CompatHelper: bump compat for "CUDAapi" to "4.0"
