### 2020-03-07 aa93286
Masking and better plots
