### 2020-03-07 1c49d2a
Actually satisfy far field U boundary condition
