### 2020-03-07 0ee6035
Easy parallel plotting
