+++ verification/flow_around_cylinder/plot_velocity_fields.py
@@ -1,18 +1,21 @@
 from copy import copy
 
+import joblib
 import xarray as xr
+import xarray.ufuncs as xu
 import matplotlib.pyplot as plt
 import cmocean
 import ffmpeg
 
-ds = xr.open_dataset("flow_around_cylinder.nc")
 
-for n in range(ds.time.size):
+def plot_cylinder_frame(n):
+    ds = xr.open_dataset("flow_around_cylinder.nc")
+
     fig, axes = plt.subplots(nrows=2, figsize=(16, 9), dpi=300)
     
     fig.suptitle(f"Flow around a cylinder, Re = 100, t = {ds.time[n].values:.3f}", fontsize=16)
 
-    u = ds.u.isel(time=n).squeeze() - 1
+    u = ds.u.isel(time=n).squeeze()
     w = ds.w.isel(time=n).squeeze()
     
     # u = u.where(ds.xF**2 + ds.zC**2 > 1)
@@ -27,8 +30,8 @@ for n in range(ds.time.size):
     img_u = u.plot.pcolormesh(ax=ax_u, vmin=0,    vmax=1.5, cmap=palette_u, add_colorbar=False)
     img_w = w.plot.pcolormesh(ax=ax_w, vmin=-0.5, vmax=0.5, cmap=palette_w, add_colorbar=False)
 
-    ax_u.set_title("u - U")
-    ax_w.set_title("w")
+    ax_u.set_title("u-velocity")
+    ax_w.set_title("w-velocity")
 
     fig.colorbar(img_u, ax=axes[0], fraction=0.046, pad=0.04)
     fig.colorbar(img_w, ax=axes[1], fraction=0.046, pad=0.04)
@@ -37,10 +40,18 @@ for n in range(ds.time.size):
         ax.set_xlabel("x")
         ax.set_ylabel("z")
         ax.set_aspect("equal")
+        ax.label_outer()
 
     print(f"Saving frame {n}/{ds.time.size-1}...")
     plt.savefig(f"flow_around_cylinder_{n:05d}.png")
     plt.close("all")
+
+ds = xr.open_dataset("flow_around_cylinder.nc")
+
+joblib.Parallel(n_jobs=-1)(
+    joblib.delayed(plot_cylinder_frame)(n)
+    for n in range(ds.time.size)
+)
 
 (
     ffmpeg
