### 2020-03-07 0f16dfa
Forgot to zero out w
