### 2020-03-09 c8819fb
Update Lid-driven cavity script to Oceananigans v0.25
