+++ verification/lid_driven_cavity/lid_driven_cavity.jl
@@ -1,19 +1,13 @@
-using Plots, Printf
-
+using Printf
 using Oceananigans
+using Oceananigans: Face, Cell
 using Oceananigans.Diagnostics
+using Oceananigans.OutputWriters
 using Oceananigans.AbstractOperations
 
-using Oceananigans: Face, Cell, NoPenetrationBC
-
-# Workaround for plotting many frames.
-# See: https://github.com/JuliaPlots/Plots.jl/issues/1723
-# import GR
-# GR.inline("png")
-
-####
-#### Data from tables 1 and 2 of Ghia et al. (1982).
-####
+#####
+##### Data from tables 1 and 2 of Ghia et al. (1982).
+#####
 
 j̃ = [1,   8,      9,      10,     14,     23,     37,     59,     65,  80,     95,     110,    123,    124,    156,    126,    129]
 ỹ = [0.0, 0.0547, 0.0625, 0.0703, 0.1016, 0.1719, 0.2813, 0.4531, 0.5, 0.6172, 0.7344, 0.8516, 0.9531, 0.9609, 0.9688, 0.9766, 1.0]
@@ -23,68 +17,73 @@ ũ = Dict(
     400 => [0.0, -0.08186, -0.09266, -0.10338, -0.14612, -0.24299, -0.32726, -0.17119, -0.11477,  0.02135, 0.16256, 0.29093, 0.55892, 0.61756, 0.68439, 0.75837, 1.0]
 )
 
-####
-#### Model setup
-####
+#####
+##### Model setup
+#####
 
-Nx, Ny, Nz = 1, 128, 128
-Lx, Ly, Lz = 1, 1, 1
+Ny = Nz = 128
+Ly = Lz = 1.0
 
-vbcs = ChannelBCs(top    = BoundaryCondition(Value, 1.0),
-                  bottom = BoundaryCondition(Value, 0.0),
-                  north  = NoPenetrationBC(),
-                  south  = NoPenetrationBC())
+topology = (Flat, Bounded, Bounded)
+domain = (x=(0, 1), y=(0, Ly), z=(0, Lz))
+grid = RegularCartesianGrid(topology=topology, size=(1, Ny, Nz); domain...)
 
-wbcs = ChannelBCs(top    = NoPenetrationBC(),
-                  bottom = NoPenetrationBC(),
-                  north  = BoundaryCondition(Value, 0.0),
-                  south  = BoundaryCondition(Value, 0.0))
+v_bcs = VVelocityBoundaryConditions(grid,
+       top = ValueBoundaryCondition(1.0),
+    bottom = ValueBoundaryCondition(0.0)
+)
 
-bcs = ChannelSolutionBCs(v=vbcs, w=wbcs)
+w_bcs = WVelocityBoundaryConditions(grid,
+    north = ValueBoundaryCondition(0.0),
+    south = ValueBoundaryCondition(0.0)
+)
 
-grid = RegularCartesianGrid(size=(Nx, Ny, Nz), x=(0, Lx), y=(0, Ly), z=(0, Lz))
+Re = 100  # Reynolds number
 
-Re = 100
-model = NonDimensionalModel(grid=grid, Re=Re, Pr=Inf, Ro=Inf,
-                            coriolis=nothing, tracers=nothing, buoyancy=nothing,
-                            boundary_conditions=bcs, forcing=forcing)
-
-nan_checker = NaNChecker(model; frequency=10, fields=Dict(:w => model.velocities.w))
-push!(model.diagnostics, nan_checker)
+model = IncompressibleModel(
+    grid=grid,
+    buoyancy=nothing,
+    tracers=nothing,
+    coriolis=nothing,
[…]
