### 2020-03-09 aea2a23
Plotting lid-driven cavity vorticity
