### 2020-03-09 9f40e2a
Fix typo in application of y value BCs
