### 2020-03-09 da1e628
Fix `cell_diffusion_timescale` when running with no tracers
