### 2020-03-09 42eff53
Fix typo: `closure.κ` -> `closure.κh`
