+++ verification/lid_driven_cavity/lid_driven_cavity.jl
@@ -5,22 +5,6 @@ using Oceananigans.Diagnostics
 using Oceananigans.OutputWriters
 using Oceananigans.AbstractOperations
 
-#####
-##### Data from tables 1 and 2 of Ghia et al. (1982).
-#####
-
-j̃ = [1,   8,      9,      10,     14,     23,     37,     59,     65,  80,     95,     110,    123,    124,    156,    126,    129]
-ỹ = [0.0, 0.0547, 0.0625, 0.0703, 0.1016, 0.1719, 0.2813, 0.4531, 0.5, 0.6172, 0.7344, 0.8516, 0.9531, 0.9609, 0.9688, 0.9766, 1.0]
-
-ũ = Dict(
-    100 => [0.0, -0.03717, -0.04192, -0.04775, -0.06434, -0.10150, -0.15662, -0.21090, -0.20581, -0.13641, 0.00332, 0.23151, 0.68717, 0.73722, 0.78871, 0.84123, 1.0],
-    400 => [0.0, -0.08186, -0.09266, -0.10338, -0.14612, -0.24299, -0.32726, -0.17119, -0.11477,  0.02135, 0.16256, 0.29093, 0.55892, 0.61756, 0.68439, 0.75837, 1.0]
-)
-
-#####
-##### Model setup
-#####
-
 Ny = Nz = 128
 Ly = Lz = 1.0
 
@@ -74,7 +58,7 @@ function print_progress(simulation)
             progress, i, t, vmax, wmax, cfl(model), dcfl(model), simulation.Δt.Δt)
 end
 
-simulation = Simulation(model, Δt=wizard, stop_time=10, progress=print_progress, progress_frequency=20)
+simulation = Simulation(model, Δt=wizard, stop_time=100, progress=print_progress, progress_frequency=20)
 
 fields = Dict("v" => model.velocities.v, "w" => model.velocities.w, "ζ" => model -> ζ_comp(model))
 dims = Dict("ζ" => ("xC", "yF", "zF"))
