### 2020-03-10 c519a3f
Sweet velocity plots
