### 2020-03-10 6930d1b
Plotting vorticity
