### 2020-03-10 297f90a
Parallelize plotting and make movies of velocity and vorticity
