### 2020-03-10 #2018 @arfon
@danielskatz - I think this is because they have already been invited (I can see a pending invite on the repository). I don't think GitHub gives us any kind of semantic response when issuing these invites.
