### 2020-03-10 #2018 @arfon
OK, errors should be more semantic now as of https://github.com/openjournals/whedon-api/pull/94
