### 2020-03-11 fc94431
Simplify down to just `left_gradient` and `right_gradient`
