### 2020-03-11 0a66e79
dispatch on whether `closure.κ` is an empty named tuple or not
