+++ src/TurbulenceClosures/turbulence_closure_diagnostics.jl
@@ -5,33 +5,44 @@ min_Δz(grid) = grid.Δz
 
 cell_diffusion_timescale(model) = cell_diffusion_timescale(model.closure, model.diffusivities, model.grid)
 
-"Returns the time-scale for diffusion on a regular grid across a single grid cell."
+function cell_diffusion_timescale(closure::ConstantIsotropicDiffusivity{V, <:NamedTuple{()}},
+                                  diffusivities, grid) where V
+    return min_Δxyz(grid)^2 / closure.ν
+end
+
 function cell_diffusion_timescale(closure::ConstantIsotropicDiffusivity, diffusivities, grid)
     Δ = min_Δxyz(grid)
-    if length(closure.κ) == 0
-        return Δ^2 / closure.ν
-    else
-        max_κ = maximum(closure.κ)
-        return min(Δ^2 / closure.ν, Δ^2 / max_κ)
-    end
+    max_κ = maximum(closure.κ)
+    return min(Δ^2 / closure.ν, Δ^2 / max_κ)
+end
+
+function cell_diffusion_timescale(closure::ConstantAnisotropicDiffusivity{V, <:NamedTuple{()}, <:NamedTuple{()}},
+                                  diffusivities, grid) where V
+    Δh = min_Δxy(grid)
+    Δz = min_Δz(grid)
+    return min(Δh^2 / closure.νh, Δz^2 / closure.νv)
 end
 
 function cell_diffusion_timescale(closure::ConstantAnisotropicDiffusivity, diffusivities, grid)
     Δh = min_Δxy(grid)
     Δz = min_Δz(grid)
-    if length(closure.κ) == 0
-        return min(Δz^2 / closure.νv, Δh^2 / closure.νh)
-    else
-        max_κh = maximum(closure.κh)
-        max_κv = maximum(closure.κv)
-        return min(Δz^2 / closure.νv, Δh^2 / closure.νh,
-                   Δz^2 / max_κv, Δh^2 / max_κh)
-    end
+    max_κh = maximum(closure.κh)
+    max_κv = maximum(closure.κv)
+    return min(Δh^2 / closure.νh, Δz^2 / closure.νv,
+               Δh^2 / max_κh, Δz^2 / max_κv)
+end
+
+function cell_diffusion_timescale(closure::AnisotropicBiharmonicDiffusivity{V, <:NamedTuple{()}, <:NamedTuple{()}},
+                                  diffusivities, grid) where V
+    Δh = min_Δxy(grid)
+    Δz = min_Δz(grid)
+    return min(Δh^4 / closure.νh, Δz^4 / closure.νv)
 end
 
 function cell_diffusion_timescale(closure::AnisotropicBiharmonicDiffusivity, diffusivities, grid)
     Δh = min_Δxy(grid)
     Δz = min_Δz(grid)
+<<<<<<< HEAD
     if length(closure.κh) == 0
         return min(Δz^4 / closure.νv, Δh^4 / closure.νh)
     else
@@ -40,6 +51,12 @@ function cell_diffusion_timescale(closure::AnisotropicBiharmonicDiffusivity, dif
         return min(Δz^4 / closure.νv, Δh^4 / closure.νh,
                    Δz^4 / max_κv, Δh^4 / max_κh)
     end
+=======
+    max_κh = maximum(closure.κh)
+    max_κv = maximum(closure.κv)
+    return min(Δh^4 / closure.νh, Δz^4 / closure.νv,
+               Δh^4 / max_κh, Δz^4 / max_κv)
+>>>>>>> dispatch on whether `closure.κ` is an empty named tuple or not
 end
 
 function cell_diffusion_timescale(closure::AbstractSmagorinsky, diffusivities, grid)
