+++ src/TurbulenceClosures/turbulence_closure_diagnostics.jl
@@ -59,12 +59,40 @@ function cell_diffusion_timescale(closure::AnisotropicBiharmonicDiffusivity, dif
 >>>>>>> dispatch on whether `closure.κ` is an empty named tuple or not
 end
 
+function cell_diffusion_timescale(closure::SmagorinskyLilly{FT, P, <:NamedTuple{()}},
+                                  diffusivities, grid) where {FT, P}
+    Δ = min_Δxyz(grid)
+    max_ν = maximum(diffusivities.νₑ.data.parent)
+    return Δ^2 / max_ν
+end
+
+function cell_diffusion_timescale(closure::BlasiusSmagorinsky{ML, FT, P, <:NamedTuple{()}},
+                                  diffusivities, grid) where {ML, FT, P}
+    Δ = min_Δxyz(grid)
+    max_ν = maximum(diffusivities.νₑ.data.parent)
+    return Δ^2 / max_ν
+end
+
 function cell_diffusion_timescale(closure::AbstractSmagorinsky, diffusivities, grid)
     Δ = min_Δxyz(grid)
     min_Pr = minimum(closure.Pr)
     max_κ = maximum(closure.κ)
     max_νκ = maximum(diffusivities.νₑ.data.parent) * max(1, 1/min_Pr)
     return min(Δ^2 / max_νκ, Δ^2 / max_κ)
+end
+
+function cell_diffusion_timescale(closure::RozemaAnisotropicMinimumDissipation{FT, <:NamedTuple{()}},
+                                  diffusivities, grid) where FT
+    Δ = min_Δxyz(grid)
+    max_ν = maximum(diffusivities.νₑ.data.parent)
+    return Δ^2 / max_ν
+end
+
+function cell_diffusion_timescale(closure::VerstappenAnisotropicMinimumDissipation{FT, PK, PN, <:NamedTuple{()}},
+                                  diffusivities, grid) where {FT, PK, PN}
+    Δ = min_Δxyz(grid)
+    max_ν = maximum(diffusivities.νₑ.data.parent)
+    return Δ^2 / max_ν
 end
 
 function cell_diffusion_timescale(closure::AbstractAnisotropicMinimumDissipation, diffusivities, grid)
