+++ verification/lid_driven_cavity/lid_driven_cavity.jl
@@ -5,69 +5,83 @@ using Oceananigans.Diagnostics
 using Oceananigans.OutputWriters
 using Oceananigans.AbstractOperations
 
-Ny = Nz = 128
-Ly = Lz = 1.0
+function simulate_lid_driven_cavity(; Re, end_time=10.0)
+    Ny = Nz = 128
+    Ly = Lz = 1.0
 
-topology = (Flat, Bounded, Bounded)
-domain = (x=(0, 1), y=(0, Ly), z=(0, Lz))
-grid = RegularCartesianGrid(topology=topology, size=(1, Ny, Nz); domain...)
+    topology = (Flat, Bounded, Bounded)
+    domain = (x=(0, 1), y=(0, Ly), z=(0, Lz))
+    grid = RegularCartesianGrid(topology=topology, size=(1, Ny, Nz); domain...)
 
-v_bcs = VVelocityBoundaryConditions(grid,
-       top = ValueBoundaryCondition(1.0),
-    bottom = ValueBoundaryCondition(0.0)
-)
+    v_bcs = VVelocityBoundaryConditions(grid,
+           top = ValueBoundaryCondition(1.0),
+        bottom = ValueBoundaryCondition(0.0)
+    )
 
-w_bcs = WVelocityBoundaryConditions(grid,
-    north = ValueBoundaryCondition(0.0),
-    south = ValueBoundaryCondition(0.0)
-)
+    w_bcs = WVelocityBoundaryConditions(grid,
+        north = ValueBoundaryCondition(0.0),
+        south = ValueBoundaryCondition(0.0)
+    )
 
-Re = 100  # Reynolds number
+    model = IncompressibleModel(
+                       grid = grid,
+                   buoyancy = nothing,
+                    tracers = nothing,
+                   coriolis = nothing,
+        boundary_conditions = (v=v_bcs, w=w_bcs),
+                    closure = ConstantIsotropicDiffusivity(ν=1/Re)
+    )
 
-model = IncompressibleModel(
-    grid=grid,
-    buoyancy=nothing,
-    tracers=nothing,
-    coriolis=nothing,
-    boundary_conditions = (v=v_bcs, w=w_bcs),
-    closure = ConstantIsotropicDiffusivity(ν=1/Re)
-)
+    u, v, w = model.velocities
+    ζ_op = ∂y(w) - ∂z(v)
+    ζ = Field(Cell, Face, Face, model.architecture, model.grid, TracerBoundaryConditions(grid))
+    ζ_computation = Computation(ζ_op, ζ)
 
-u, v, w = model.velocities
-ζ_op = ∂y(w) - ∂z(v)
-ζ = Field(Cell, Face, Face, model.architecture, model.grid, TracerBoundaryConditions(grid))
-ζ_comp = Computation(ζ_op, ζ)
+    fields = Dict(
+        "v" => model.velocities.v,
+        "w" => model.velocities.w,
+        "ζ" => model -> ζ_computation(model)
+    )
 
-max_Δt = 0.25 * model.grid.Δy^2 * Re / 2
-wizard = TimeStepWizard(cfl=0.1, Δt=1e-6, max_change=1.1, max_Δt=max_Δt)
-cfl = AdvectiveCFL(wizard)
-dcfl = DiffusiveCFL(wizard)
+    dims = Dict("ζ" => ("xC", "yF", "zF"))
+    global_attributes = Dict("Re" => Re)
+    output_attributes = Dict("ζ" => Dict("longname" => "vorticity", "units" => "1/s"))
+
+    field_output_writer =
+        NetCDFOutputWriter(model, fields, filename="lid_driven_cavity_Re$Re.nc", interval=0.01,
+                           global_attributes=global_attributes, output_attributes=output_attributes,
+                           dimensions=dims)
+
[…]
