### 2020-03-11 c2a6fef
Combine velocity and vorticity plots
