+++ verification/lid_driven_cavity/plot_lid_driven_cavity.py
@@ -33,17 +33,18 @@ y_ζ_Ghia = [0.0000, 0.0625, 0.1250, 0.1875, 0.2500, 0.3125, 0.3750, 0.4375, 0.5
     100: [nan, 40.0110, 22.5378, 16.2862, 12.7844, 10.4199, 8.69628, 7.43218, 6.57451, 6.13973, 6.18946, 6.82674, 8.22110, 10.7414, 15.6591, 30.7923, nan]
 }
 
-def plot_velocity_frame(Re, n):
+def plot_lid_driven_cavity_frame(Re, n):
     ds = xr.open_dataset(f"lid_driven_cavity_Re{Re}.nc")
     Ny = ds.yC.size
     Nz = ds.zC.size
 
-    fig, axes = plt.subplots(nrows=2, ncols=2, figsize=(16, 9), dpi=200)
+    fig, axes = plt.subplots(nrows=3, ncols=2, figsize=(16, 16), dpi=200)
     plt.subplots_adjust(hspace=0.25)
     fig.suptitle(f"Lid-driven cavity, Re = {Re}, t = {ds.time[n].values:.2f}", fontsize=16)
 
     ax_v_line, ax_v_mesh = axes[0, 0], axes[0, 1]
     ax_w_line, ax_w_mesh = axes[1, 0], axes[1, 1]
+    ax_ζ_line, ax_ζ_mesh = axes[2, 0], axes[2, 1]
 
     v_line = ds.v.isel(time=n, yF=Ny//2)
     ax_v_line.plot(y_Ghia, v_Ghia[Re], label="Ghia et al. (1982)", color="tab:blue", linestyle="", marker="o", fillstyle="none")
@@ -79,68 +80,41 @@ def plot_velocity_frame(Re, n):
     ax_w_mesh.set_ylabel("z")
     ax_w_mesh.set_aspect("equal")
 
-    print(f"Saving velocity frame {n}/{ds.time.size-1}...")
-    plt.savefig(f"lid_driven_cavity_velocity_Re{Re}_{n:05d}.png")
-    plt.close("all")
-
-def plot_vorticity_frame(Re, n):
-    ds = xr.open_dataset(f"lid_driven_cavity_Re{Re}.nc")
-    Ny = ds.yC.size
-    Nz = ds.zC.size
-
-    fig, axes = plt.subplots(ncols=2, figsize=(16, 9), dpi=200)
-    fig.suptitle(f"Lid-driven cavity, Re = {Re}, t = {ds.time[n].values:.2f}", fontsize=16)
-
-    ax_line, ax_mesh = axes[0], axes[1]
-
     ζ_line = -ds.ζ.isel(time=n, zF=Nz-1)
-    ax_line.plot(y_ζ_Ghia, ζ_Ghia[Re], label="Ghia et al. (1982)", color="tab:red", linestyle="", marker="o", fillstyle="none")
-    ax_line.plot(ds.yF, ζ_line.values.flatten(), label="Oceananigans.jl", color="tab:red")
-    ax_line.legend(loc="lower left", bbox_to_anchor=(0, 1.01, 1, 0.2), ncol=2, frameon=False)
-    ax_line.set_xlabel("y")
-    ax_line.set_ylabel("vorticity $\zeta$")
-    ax_line.set_xlim([0, 1])
-    ax_line.set_ylim(bottom=0)
+    ax_ζ_line.plot(y_ζ_Ghia, ζ_Ghia[Re], label="Ghia et al. (1982)", color="tab:red", linestyle="", marker="o", fillstyle="none")
[…]
