### 2020-03-13 b47aa26
Adds function to extract state from Model, and greatly simplifies forcing function and boundary condition function signatures
