### 2020-03-14 65830a8
Fixes Rayleigh-Benard test for new forcing function syntax
