+++ src/TimeSteppers/velocity_and_tracer_tendencies.jl
@@ -1,6 +1,6 @@
 """
     u_velocity_tendency(i, j, k, grid, coriolis, surface_waves, 
-                        closure, U, C, K, F, pHY′, parameters, clock)
+                        closure, U, C, K, F, pHY′, clock)
 
 Return the tendency for the horizontal velocity in the x-direction, or the east-west 
 direction, ``u``, at grid point `i, j, k`.
@@ -22,7 +22,7 @@ forcing functions, `pHY′` is the hydrostatic pressure anomaly.
 and `clock` is the physical clock of the model.
 """
 @inline function u_velocity_tendency(i, j, k, grid, coriolis, surface_waves, 
-                                     closure, U, C, K, F, pHY′, parameters, clock)
+                                     closure, U, C, K, F, pHY′, clock)
 
     return ( - div_ũu(i, j, k, grid, U)
              - x_f_cross_U(i, j, k, grid, coriolis, U)
@@ -35,7 +35,7 @@ end
 
 """
     v_velocity_tendency(i, j, k, grid, coriolis, surface_waves, 
-                        closure, U, C, K, F, pHY′, parameters, clock)
+                        closure, U, C, K, F, pHY′, clock)
 
 Return the tendency for the horizontal velocity in the y-direction, or the north-south 
 direction, ``v``, at grid point `i, j, k`.
@@ -57,7 +57,7 @@ forcing functions, `pHY′` is the hydrostatic pressure anomaly.
 and `clock` is the physical clock of the model.
 """
 @inline function v_velocity_tendency(i, j, k, grid, coriolis, surface_waves, 
-                                     closure, U, C, K, F, pHY′, parameters, clock)
+                                     closure, U, C, K, F, pHY′, clock)
 
     return ( - div_ũv(i, j, k, grid, U)
              - y_f_cross_U(i, j, k, grid, coriolis, U)
@@ -70,7 +70,7 @@ end
 
 """
     w_velocity_tendency(i, j, k, grid, coriolis, surface_waves, 
-                        closure, U, C, K, F, parameters, clock)
+                        closure, U, C, K, F, clock)
                         
 Return the tendency for the vertical velocity ``w`` at grid point `i, j, k`.
 The tendency for ``w`` is called ``G_w`` and defined via
@@ -90,7 +90,7 @@ forcing functions, `pHY′` is the hydrostatic pressure anomaly.
 and `clock` is the physical clock of the model.
 """
 @inline function w_velocity_tendency(i, j, k, grid, coriolis, surface_waves, 
-                                     closure, U, C, K, F, parameters, clock)
+                                     closure, U, C, K, F, clock)
 
     return ( - div_ũw(i, j, k, grid, U)
              - z_f_cross_U(i, j, k, grid, coriolis, U)
@@ -101,8 +101,8 @@ and `clock` is the physical clock of the model.
 end
 
 """
-    tracer_tendency(i, j, k, grid, c, tracer_index, closure, buoyancy, U, C, K, Fc,
-                    parameters, clock)
+    tracer_tendency(i, j, k, grid, c, tracer_index, 
+                    closure, buoyancy, U, C, K, Fc, clock)
 
 Return the tendency for a tracer field `c` with index `tracer_index` 
 at grid point `i, j, k`.
@@ -118,11 +118,10 @@ The arguments `U`, `C`, and `K` are `NamedTuple`s with the three velocity compon
 tracer fields, and  precalculated diffusivities where applicable. 
 `Fc` is the user-defined forcing function for tracer `c`.
 
-`parameters` is a `NamedTuple` of scalar parameters for user-defined forcing functions 
-and `clock` is the physical clock of the model.
[…]
