+++ src/TimeSteppers/time_stepping_kernels.jl
@@ -4,7 +4,7 @@
 
 """ Store previous value of the source term and calculate current source term. """
 function calculate_interior_tendency_contributions!(G, arch, grid, coriolis, buoyancy, surface_waves, closure, 
-                                                    U, C, pHY′, K, F, parameters, clock)
+                                                    U, C, pHY′, K, F, clock)
 
     # Manually choose thread-block layout here as it's ~20% faster.
     # See: https://github.com/climate-machine/Oceananigans.jl/pull/308
@@ -22,13 +22,13 @@ function calculate_interior_tendency_contributions!(G, arch, grid, coriolis, buo
     end
 
     @launch(device(arch), threads=(Tx, Ty), blocks=(Bx, By, Bz),
-            calculate_Gu!(G.u, grid, coriolis, surface_waves, closure, U, C, K, F, pHY′, parameters, clock))
+            calculate_Gu!(G.u, grid, coriolis, surface_waves, closure, U, C, K, F, pHY′, clock))
 
     @launch(device(arch), threads=(Tx, Ty), blocks=(Bx, By, Bz),
-            calculate_Gv!(G.v, grid, coriolis, surface_waves, closure, U, C, K, F, pHY′, parameters, clock))
+            calculate_Gv!(G.v, grid, coriolis, surface_waves, closure, U, C, K, F, pHY′, clock))
 
     @launch(device(arch), threads=(Tx, Ty), blocks=(Bx, By, Bz),
-            calculate_Gw!(G.w, grid, coriolis, surface_waves, closure, U, C, K, F, parameters, clock))
+            calculate_Gw!(G.w, grid, coriolis, surface_waves, closure, U, C, K, F, clock))
 
     for tracer_index in 1:length(C)
         @inbounds Gc = G[tracer_index+3]
@@ -36,7 +36,7 @@ function calculate_interior_tendency_contributions!(G, arch, grid, coriolis, buo
         @inbounds  c = C[tracer_index]
 
         @launch(device(arch), threads=(Tx, Ty), blocks=(Bx, By, Bz),
-                calculate_Gc!(Gc, grid, c, Val(tracer_index), closure, buoyancy, U, C, K, Fc, parameters, clock))
+                calculate_Gc!(Gc, grid, c, Val(tracer_index), closure, buoyancy, U, C, K, Fc, clock))
     end
 
     return nothing
@@ -65,17 +65,17 @@ end
 """
     calculate_east_boundary_Gu!(G, arch, grid::AbstractGrid{FT, <:Bounded},
                                 coriolis, buoyancy, surface_waves, closure,
-                                U, C, pHY′, K, F, parameters, clock) where FT
+                                U, C, pHY′, K, F, clock) where FT
 
 Calculate `Gu` on east boundaries when the x-direction has `Bounded` topology.
 """
 function calculate_east_boundary_Gu!(G, arch, grid::AbstractGrid{FT, <:Bounded},
                                      coriolis, buoyancy, surface_waves, closure,
-                                     U, C, pHY′, K, F, parameters, clock) where FT
+                                     U, C, pHY′, K, F, clock) where FT
 
     @launch(device(arch), config=launch_config(grid, :yz), 
             _calculate_east_boundary_Gu!(G.u, grid, coriolis, surface_waves, 
-                                         closure, U, C, K, F, pHY′, parameters, clock))
+                                         closure, U, C, K, F, pHY′, clock))
 
     return nothing
 end
@@ -83,17 +83,17 @@ end
 """
     calculate_north_boundary_Gu!(G, arch, grid::AbstractGrid{FT, <:Bounded},
                                  coriolis, buoyancy, surface_waves, closure,
-                                 U, C, pHY′, K, F, parameters, clock) where FT
+                                 U, C, pHY′, K, F, clock) where FT
 
[…]
