+++ src/Forcing/Forcing.jl
@@ -8,5 +8,6 @@ zeroforcing(args...) = 0
 
 include("simple_forcing.jl")
 include("model_forcing.jl")
+include("parameterized_forcing.jl")
 
 end
