### 2020-03-14 fa7948d
Updates eady turbulence example for new forcing function signature
