### 2020-03-14 b5e90fa
Adds parameterized boundary conditions
