### 2020-03-14 a1a1493
Removes unnecessary explicit grid reference in fill halo regions
