### 2020-03-14 40d1658
Updates stratified couette flow verification for nuking of model.parameters
