### 2020-03-14 6415de6
Updates wind mixing and convection example for new boundary condition function signature
