### 2020-03-14 ffb7205
Fixes bug in buoyancy forcing function in eady turbulence example
