### 2020-03-14 db330c8
Bug fixes in tests for new function and boundary condition signature
