+++ test/test_forcings.jl
@@ -1,5 +1,6 @@
 add_one(args...) = 1.0
 
+""" Instantiate ModelForcing. """
 function test_forcing(fld)
     kwarg = Dict(fld => add_one)
     forcing = ModelForcing(; kwarg...)
@@ -7,44 +8,55 @@ function test_forcing(fld)
     return f() == 1.0
 end
 
+""" Take one time step with three forcing functions on u, v, w. """
 function time_step_with_forcing_functions(arch)
-    @inline Fu(i, j, k, grid, time, U, Φ, params) = @inbounds ifelse(k == grid.Nz, -U.u[i, j, k] / 60, 0)
-    @inline Fv(i, j, k, grid, time, U, Φ, params) = @inbounds ifelse(k == grid.Nz, -U.v[i, j, k] / 60, 0)
-    @inline Fw(i, j, k, grid, time, U, Φ, params) = @inbounds ifelse(k == grid.Nz, -U.w[i, j, k] / 60, 0)
+    @inline Fu(i, j, k, grid, clock, state) = @inbounds ifelse(k == grid.Nz, -state.velocities.u[i, j, k] / 60, 0)
+    @inline Fv(i, j, k, grid, clock, state) = @inbounds ifelse(k == grid.Nz, -state.velocities.v[i, j, k] / 60, 0)
+    @inline Fw(i, j, k, grid, clock, state) = @inbounds ifelse(k == grid.Nz, -state.velocities.w[i, j, k] / 60, 0)
 
     forcing = ModelForcing(u=Fu, v=Fv, w=Fw)
 
     grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
     model = IncompressibleModel(grid=grid, architecture=arch, forcing=forcing)
     time_step!(model, 1, euler=true)
+
     return true
 end
 
-function time_step_with_forcing_functions_params(arch)
-    @inline Fu(i, j, k, grid, time, U, Φ, params) = @inbounds ifelse(k == grid.Nz, -U.u[i, j, k] / params.τ, 0)
-    @inline Fv(i, j, k, grid, time, U, Φ, params) = @inbounds ifelse(k == grid.Nz, -U.v[i, j, k] / params.τ, 0)
-    @inline Fw(i, j, k, grid, time, U, Φ, params) = @inbounds ifelse(k == grid.Nz, -U.w[i, j, k] / params.τ, 0)
+""" Take one time step with ParameterizedForcing forcing functions. """
+function time_step_with_parameterized_forcing(arch)
+    @inline Fu_func(i, j, k, grid, clock, state, params) = @inbounds ifelse(k == grid.Nz, -state.velocities.u[i, j, k] / params.τ, 0)
+    @inline Fv_func(i, j, k, grid, clock, state, params) = @inbounds ifelse(k == grid.Nz, -state.velocities.v[i, j, k] / params.τ, 0)
+    @inline Fw_func(i, j, k, grid, clock, state, params) = @inbounds ifelse(k == grid.Nz, -state.velocities.w[i, j, k] / params.τ, 0)
+
+    Fu = ParameterizedForcing(Fu_func, (τ=60,))
+    Fv = ParameterizedForcing(Fv_func, (τ=60,))
+    Fw = ParameterizedForcing(Fw_func, (τ=60,))
 
     forcing = ModelForcing(u=Fu, v=Fv, w=Fw)
 
     grid = RegularCartesianGrid(size=(16, 16, 16), length=(1, 1, 1))
-    model = IncompressibleModel(grid=grid, architecture=arch, forcing=forcing, parameters=(τ=60,))
+    model = IncompressibleModel(grid=grid, architecture=arch, forcing=forcing)
     time_step!(model, 1, euler=true)
+
[…]
