### 2020-03-15 f36a6aa
Import Adapt into Models module
