### 2020-03-15 15beefe
Deletes obsolete comments about model.parameters from ocean wind mixing and convection example
