### 2020-03-15 bebbf27
Refactors and simplifies field constructors
