### 2020-03-15 c5d841e
Adds forced flow verification setup
