### 2020-03-15 5f6115a
Adds convenience constructors for boundary functions applied to tracer or velocity fields. Allows boundary function to have embedded parameters
