+++ src/BoundaryConditions/boundary_function.jl
@@ -1,5 +1,5 @@
 """
-        BoundaryFunction{B, X1, X2}(func)
+        BoundaryFunction{B, X1, X2}(func, parameters=nothing)
 
     A wrapper for user-defined boundary condition functions on the
     boundary specified by symbol `B` and at location `(X1, X2)`.
@@ -12,22 +12,144 @@
     (::BoundaryFunction{:z,Cell,Cell,getfield(Main, Symbol("##7#8"))}) (generic function with 1 method)
 
     julia> top_tracer_bc = BoundaryCondition(Flux, top_tracer_flux);
-"""
-struct BoundaryFunction{B, X1, X2, F} <: Function
-    func :: F
 
-    function BoundaryFunction{B, X1, X2}(func) where {B, X1, X2}
+    julia> momentum_flux_func(x, y, t, p) = cos(p.k * x) * cos(p.ω * t);
+
+    julia> parameterized_u_velocity_flux = BoundaryFunction{:z, Face, Cell}(flux_func, (k=4π, ω=3.0))
+
+    julia> top_u_bc = BoundaryCondition(Flux, parameterized_u_velocity_flux)
+
+"""
+struct BoundaryFunction{B, X1, X2, P, F} <: Function
+    func :: F
+    parameters :: P
+
+    function BoundaryFunction{B, X1, X2}(func, parameters=nothing) where {B, X1, X2}
         B ∈ (:x, :y, :z) || throw(ArgumentError("The boundary B at which the BoundaryFunction is
                                                 to be applied must be either :x, :y, or :z."))
-        return new{B, X1, X2, typeof(func)}(func)
+        return new{B, X1, X2, typeof(parameters), typeof(func)}(func, parameters)
     end
 end
 
-@inline (bc::BoundaryFunction{:x, Y, Z})(j, k, grid, clock, state) where {Y, Z} =
+@inline (bc::BoundaryFunction{:x, Y, Z, Nothing})(j, k, grid, clock, state) where {Y, Z} =
     bc.func(ynode(Y, j, grid), znode(Z, k, grid), clock.time)
 
-@inline (bc::BoundaryFunction{:y, X, Z})(i, k, grid, clock, state) where {X, Z} =
+@inline (bc::BoundaryFunction{:y, X, Z, Nothing})(i, k, grid, clock, state) where {X, Z} =
     bc.func(xnode(X, i, grid), znode(Z, k, grid), clock.time)
 
-@inline (bc::BoundaryFunction{:z, X, Y})(i, j, grid, clock, state) where {X, Y} =
+@inline (bc::BoundaryFunction{:z, X, Y, Nothing})(i, j, grid, clock, state) where {X, Y} =
     bc.func(xnode(X, i, grid), ynode(Y, j, grid), clock.time)
+
+@inline (bc::BoundaryFunction{:x, Y, Z})(j, k, grid, clock, state) where {Y, Z} =
+    bc.func(ynode(Y, j, grid), znode(Z, k, grid), clock.time, bc.parameters)
+
+@inline (bc::BoundaryFunction{:y, X, Z})(i, k, grid, clock, state) where {X, Z} =
+    bc.func(xnode(X, i, grid), znode(Z, k, grid), clock.time, bc.parameters)
+
+@inline (bc::BoundaryFunction{:z, X, Y})(i, j, grid, clock, state) where {X, Y} =
+    bc.func(xnode(X, i, grid), ynode(Y, j, grid), clock.time, bc.parameters)
+
+#####
+##### Convenience constructors
+#####
+
+UVelocityBoundaryLocation(::Val{:x}) = (Cell, Cell) # yz boundary coordinates
+UVelocityBoundaryLocation(::Val{:y}) = (Face, Cell) # xz boundary coordinates
[…]
