### 2020-03-15 52401c4
Exports new convenience constructors from BoundaryConditions module
