### 2020-03-15 1a7e516
Adds test for instantiation via new boundary function constructors
