### 2020-03-15 d4333cf
Removes forced flow verification
