### 2020-03-15 703be48
Less code, less boiler plate
