### 2020-03-15 c04e29f
Improves docstring
