### 2020-03-16 f135fd5
Shortcircuits has_velocities for empty tuples, adds a test
