### 2020-03-16 39c708a
Adds back forced flow verification functionality
