+++ verification/forced_flow/run_free_slip_simulations.jl
@@ -1,0 +1,6 @@
+# Run 4 simulations:
+
+setup_and_run_free_slip_simulation(Nx=16, Nz=16, CFL=1e-3)
+setup_and_run_free_slip_simulation(Nx=32, Nz=32, CFL=1e-3)
+setup_and_run_free_slip_simulation(Nx=64, Nz=64, CFL=1e-3)
+setup_and_run_free_slip_simulation(Nx=128, Nz=128, CFL=1e-3)
