### 2020-05-20 #2018 @kthyng
Hi everyone! We are officially starting JOSS back up today. We know that everyone has different circumstances with work and home life right now, so we want to figure out a way forward that will work for everyone.

@ysimillides — are you still able to review this submission? If so, what timeline would work for you?
