### 2020-05-21 #2018 @ysimillides
Hello, I can't access the checklist above, but by going through the items my main concern is that there is no comparison / mention of other similar software which can be found ( state of the field on the checklist). If a few lines in the paper could be added to reflect this, that would be ideal ( and likewise if their are no similar packages, this should be explicitly stated).

Otherwise everything seems fine and the package follows best practices for Julia packages ( integrated with the package manager, automated tests +documentation )
