### 2020-06-02 #2018 @kthyng
Hi @ysimillides! Were you able to accept the invitation above to gain permissions for your checklist? Have you been able to install the software and try it out?
