### 2020-07-06 #2018 @kthyng
@mancellin — would you be interested in reviewing this submission to JOSS? We already have one review completed but the other reviewer went MIA. You look like a good reviewer knowing Julia and related CFD.
