### 2020-07-10 #2018 @ali-ramadhan
@kthyng No worries about the delay, thank you for keeping this review moving forward!

I've fallen off the grid in the past few months myself but I believe we've addressed most of the points from @funsim's review so I'll reply more formally to the review soon.

@mancellin Thanks so much for agreeing to review our submission! Please let me know if you hit any issues or have any questions.
