### 2020-07-23 CliMA/Oceananigans.jl#611 @glwagner
This was resolved by #767
