### 2020-07-23 CliMA/Oceananigans.jl#608 @glwagner
I think this is resolved?
