### 2020-07-23 CliMA/Oceananigans.jl#608 @ali-ramadhan
Haha so the link is dead again since we don't have a `benchmarks.jl` file. It should link to the documentation page on performance benchmarks.

I will open a PR.
