### 2020-08-04 #2018 @mancellin
Overall, this is a very nice package. The documentation is extensive, the package is easy to install and works as expected.

I have some issues running simulations on GPU, but I believe it to be due to my own setup and not to the package itself. I'll give it another try in the next days and open an issue if I need help.

One minor issue: the review check-list requires the authors to compare their code to other commonly-used packages. Could you add a few lines in the paper about the state of the art, citing other simulation codes targetting the same problem (if they exist)?

Besides, have the minor issues raised by the other reviewer in January been addressed?
