### 2020-08-04 #2018 @ali-ramadhan
Thank you for the review and the kind words @mancellin!

> I have some issues running simulations on GPU, but I believe it to be due to my own setup and not to the package itself. I'll give it another try in the next days and open an issue if I need help.

Apologies about this. We discovered a major bug concerning running GPU simulations that we fixed in Oceananigans v0.32.0 last week, so maybe try a more recent version if you were trying an older one. Otherwise we'd love to hear more about the issue so we can fix it.

> One minor issue (already mentioned by @ysimillides): the review check-list requires the authors to compare their code to other commonly-used packages. Could you add a few lines in the paper about the state of the art, citing other simulation codes targetting the same problem (if they exist)?

Thank you for bringing this up. I will update the software paper with a discussion of other similar packages.

> Besides, have the minor issues raised by @funsim in January been addressed?

Yes I believe so (the issues have been closed) but I think I still need to post a formal reply. Sorry about this delay on my part. I will do it today.
