### 2020-08-11 CliMA/Oceananigans.jl#609 @ali-ramadhan
This was resolved in PR #794
