### 2020-08-21 CliMA/Oceananigans.jl#864 @ali-ramadhan
`scalar getindex is disallowed` with `topology = (Periodic, Bounded, Bounded)` using Oceananigans v0.34.1
Reported by @sandreza

Must have missed this (or tests didn't cover this case) in PR https://github.com/CliMA/Oceananigans.jl/pull/851
