### 2020-08-27 CliMA/Oceananigans.jl#873 @ali-ramadhan
Run convergence tests as part of the verification testset
Otherwise they might go out of date. Also good to keep checking convergence I suppose.

X-Ref: https://github.com/openjournals/joss-reviews/issues/2018
