### 2020-08-27 CliMA/Oceananigans.jl#864 @ali-ramadhan
Not sure why tests did not catch this in #851, but running tests on Tartarus is catching it now 🤷

Fixing and will open a PR.
