### 2020-08-27 #2018 @mancellin
@ali-ramadhan Thank you for your reply.

> Let me know if you've had any luck with more recent version of Oceananigans.jl, otherwise I will look into the issue

Yes, I managed to run a simulation on GPU after updating Oceananigans!
However, some tests from the Oceananigans test suite seem to be still failing on GPU. I'll open an issue on that.

>      https://clima.github.io/OceananigansDocumentation/stable/verification/convergence_tests/

FYI, the plots seem to be missing from the documentation page.
In the "Verification experiments" section of the documentation, you could maybe add links to the relevant source code on Github.
