### 2020-08-27 #2018 @mancellin
@kthyng Everything looks good to me. I can recommend the paper for publication.
