### 2020-08-27 CliMA/Oceananigans.jl#873 @glwagner
Agree this is a necessity. Some tests are cheap enough to verify convergence continuously. Some of them are expensive so I think we should just ensure that they continue to execute (like we do for the other verification tests).
