### 2020-08-27 CliMA/Oceananigans.jl#879 @ali-ramadhan
Include all convergence test plots in documentation
Thanks @mancellin for pointing this out.

X-Ref: https://github.com/openjournals/joss-reviews/issues/2018
