### 2020-08-27 #2018 @ali-ramadhan
Thanks @mancellin!

Not sure if these are similar to the GPU test failures you experienced but I fixed some GPU tests yesterday (see issue https://github.com/CliMA/Oceananigans.jl/issues/864 and PR https://github.com/CliMA/Oceananigans.jl/pull/878). But if you can open an issue we'd appreciate it very much!

Also thanks for pointing out the convergence tests. I'll make sure all the convergence tests run and I'll include plots for each one in the documentation.

I opened two issues to work on:
https://github.com/CliMA/Oceananigans.jl/issues/873
https://github.com/CliMA/Oceananigans.jl/issues/879
