### 2020-08-27 #2018 @ali-ramadhan
Thanks @kthyng!

Just curious whether the package version in the original post (v0.18.0) matters? v0.18.0 is an ancient version at this point and we're about to release v0.35.0 but it doesn't seem to show up in the JOSS paper so maybe it doesn't matter.
