### 2020-08-27 CliMA/Oceananigans.jl#879 @ali-ramadhan
Ah look like only the forced flow figures show up at https://clima.github.io/OceananigansDocumentation/stable/verification/convergence_tests/

The others seem to have dead image links I think

There are only two figures here: https://github.com/CliMA/Oceananigans.jl/tree/master/docs/src/verification/convergence_plots
