### 2020-08-27 CliMA/Oceananigans.jl#879 @ali-ramadhan
Resolved by #880
