### 2020-08-28 CliMA/Oceananigans.jl#882 @ali-ramadhan
Thanks so much for taking the time to report this @mancellin!

Wasn't able to immediately reproduce it but I'll try on a couple of other different machines.

Actually, from doing a little bit of searching around I found two CUDA.jl issues with similar errors (https://github.com/JuliaGPU/CUDA.jl/issues/190 and https://github.com/JuliaGPU/CUDA.jl/issues/286) and it looks like they added a note about the error to the CUDA.jl docs: https://juliagpu.gitlab.io/CUDA.jl/installation/troubleshooting/#LLVM-error:-Cannot-cast-between-two-non-generic-address-spaces

> LLVM error: Cannot cast between two non-generic address spaces
>
> You are using an unpatched copy of LLVM, likely caused by using Julia as packaged by your Linux distribution. These often decide to use a global copy of LLVM instead of using the one built and patched by Julia during the compilation process. This is not supported: LLVM cannot easily be used like a regular shared library, as Julia (and other users of LLVM alike) has an extensive list of patches to be applied to the specific versions of LLVM that are supported.
>
> It is thus recommended to use the official binaries, or use a version of Julia built without setting `USE_SYSTEM_LLVM=1` (which you can suggest to maintainers of your Linux distribution).

So maybe it's just a case of needing to use Julia binaries from https://julialang.org/downloads/ ?
