### 2020-08-28 CliMA/Oceananigans.jl#882 @ali-ramadhan
Don't feel obligated to back and run the tests with a Julia binary from https://julialang.org/downloads/ but let me know if you feel like it was likely enough to be the problem to close this issue.
