### 2020-08-28 #2018 @ali-ramadhan
@mancellin @kthyng The docs on latest version of Oceananigans.jl should have all the convergence plots now: https://clima.github.io/OceananigansDocumentation/latest/verification/convergence_tests/
