### 2020-09-03 #2018 @ali-ramadhan
We released Oceananigans v0.35.0 yesterday which should contain all the changes suggested by @funsim and @mancellin.
