### 2020-09-04 #2018 @kthyng
Oops, too soon. Anyway, I read through your paper and it looks great. After you report back here with the Zenodo archive, we'll be good to finish up!
