### 2020-09-04 #2018 @ali-ramadhan
Ah thanks for the reminder! We'll merge the `joss-paper` branch into the `master` branch (forgot to do this) and tag v0.36.0 which includes some new bug fixes and should show up on Zenodo.
