### 2020-09-08 #2018 @ali-ramadhan
Oceananigans v0.36.0 has been archived on Zenodo with the following DOI: [10.5281/zenodo.4019272](https://doi.org/10.5281/zenodo.4019272)

I've edited the authors of the Zenodo entry to reflect the author list on the software paper but I don't think the change has been reflected in the Zenodo page yet?
