### 2020-09-08 #2018 @ali-ramadhan
Ah sorry was just hoping to bump the version to v0.36.0 to reflect the Zenodo entry.
