### 2020-09-10 #2018 @danielskatz
👋 @funsim - I see three boxes you have not checked off - are these done now from your point of view?  If so, can you check them off?
