### 2020-09-10 #2018 @ali-ramadhan
Not sure if @funsim is still active on GitHub (their review was posted a while back) but @kthyng suggested she might be able to check those boxes off in an earlier comment: https://github.com/openjournals/joss-reviews/issues/2018#issuecomment-681924429
