### 2020-09-15 #2018 @ali-ramadhan
@kthyng Are you able to check off the missing boxes or should we wait for @funsim?
