### 2020-09-18 #2018 @kthyng
@openjournals/dev Could you see why I can generate the paper but not accept it? Thank you.
