### 2020-09-22 #2018 @arfon
I think Whedon is struggling with an unusual DOI. This PR should fix the issue: https://github.com/CliMA/Oceananigans.jl/pull/966
