### 2020-09-22 #2018 @arfon
@kthyng - I'm not sure if you had other checks to make here but this paper is now compiling properly.
