### 2020-09-22 #2018 @kthyng
@ali-ramadhan Ok great! You are developing quickly. Do you want to use your now newest version in this publication or stop at v0.36? Also can you change the metadata in the JOSS zenodo archive so that the title and author list exactly match your JOSS paper?
