### 2020-09-22 #2018 @ali-ramadhan
@kthyng Ah I think we'll stick with v0.36.0 here as we specifically tagged it for JOSS (if that's okay with JOSS).

I thought I did change the JOSS Zenodo archive metadata for v0.36.0 so the authors match the JOSS paper authors exactly (https://zenodo.org/record/4019272 from https://github.com/openjournals/joss-reviews/issues/2018#issuecomment-690279423).

There is a newer different Zenodo entry for v0.37.0 that pulled the authors from GitHub contributors. I can change the authors on this Zenodo archive as well if needed?
