### 2020-09-22 #2018 @kthyng
> @kthyng Ah I think we'll stick with v0.36.0 here as we specifically tagged it for JOSS (if that's okay with JOSS).
>

No problem!

> I thought I did change the JOSS Zenodo archive metadata for v0.36.0 so the authors match the JOSS paper authors exactly (https://zenodo.org/record/4019272 from [#2018 (comment)](https://github.com/openjournals/joss-reviews/issues/2018#issuecomment-690279423)).

The title at that link is "CliMA/Oceananigans.jl: v0.36.0" but your submission title is "Oceananigans.jl: Fast and friendly geophysical fluid dynamics on GPUs". Author list looks fine on a quick glance.
